% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bmr}
\alias{bmr}
\title{Boix-Miller-Rosato dichotomous coding of democracy, 1800-2020, version 4.0}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19775 rows and 16 columns.
}
\source{
C. Boix, M. Miller, and S. Rosato. "A Complete Data Set of Political Regimes, 1800–2007". In: \emph{Comparative Political Studies} 46.12 (2012), pp. 1523-1554. DOI: \href{https://dx.doi.org/10.1177/0010414012463905}{10.1177/0010414012463905}. Available at
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/FJLMKT}
(versions 1-3) and \url{https://sites.google.com/site/mkmtwo/data} (version 4.0).
}
\usage{
bmr
}
\description{
A dataset containing the Boix-Miller-Rosato (BMR) dichotomous coding of
democracy, first used in the paper C. Boix, M. Miller, and S. Rosato. "A Complete Data Set of Political Regimes, 1800–2007". In: \emph{Comparative Political Studies} 46.12 (2012), pp. 1523-1554. DOI: \href{https://dx.doi.org/10.1177/0010414012463905}{10.1177/0010414012463905}. The complete dataset can be
downloaded at \url{https://sites.google.com/site/mkmtwo/data}. The
complete codebook is available there as well.
}
\details{
\describe{ \item{bmr_country}{The country name in BMR.}

\item{bmr_ccode}{Numeric polity code code in BMR. }

\item{abbreviation}{World Bank 3-letter code.}

\item{abbreviation_undp}{UNDP 3-letter code.}

\item{year}{Year}

\item{democracy}{Dichotomous democracy measure. 1 = Democracy.}

\item{sovereign}{Dichotomous indicator of sovereignty/independence (if 0,
democracy is NA).}

\item{democracy_trans}{-1 if democratic breakdown, 0 if no change, 1 if
democratic transition.}

\item{democracy_breakdowns}{Previous number of democratic breakdowns.}

\item{democracy_duration}{Consecutive years of current regime type.}

\item{democracy_omitteddata}{This is the same measure as democracy, except it
records an NA for countries occupied during an international war (e.g., the
Netherlands 1940-44) or experiencing state collapse during a civil war (e.g.,
Lebanon 1976-89). The democracy variable instead fills in these years as
continuations of the same regime type.}

\item{democracy_femalesuffrage}{This adjusts democracy by also requiring that
at least half of adult women have the right to vote. 30 countries change
values.} }
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other events data: 
\code{\link{bnr}},
\code{\link{prc_gasiorowski}},
\code{\link{ulfelder}}

Other dichotomous democracy indexes: 
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{kailitz}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{wahman_teorell_hadenius}}
}
\concept{democracy}
\concept{dichotomous democracy indexes}
\concept{events data}
\keyword{datasets}
