% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bollen_pmm}
\alias{bollen_pmm}
\alias{bollen}
\title{The Bollen measure of democracy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 510 rows and 7 columns.
}
\source{
K. Bollen and P. Paxton. "Subjective Measures of Liberal Democracy". In: \emph{Comparative Political Studies} 33.1 (2000), pp. 58-86. DOI: \href{https://dx.doi.org/10.1177/0010414000033001003}{10.1177/0010414000033001003}.

K. A. Bollen. \emph{Cross-National Indicators of Liberal Democracy, 1950-1990}. Inter-university Consortium for Political and Social Research (ICPSR) [distributor], 2001. DOI: \href{https://dx.doi.org/10.3886/ICPSR02532.v2}{10.3886/ICPSR02532.v2}. Original data available
at \url{https://www.icpsr.umich.edu/web/ICPSR/studies/2532/publications}. Taken from
D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM.
}
\usage{
bollen_pmm
}
\description{
0-100 index of democracy from K. A. Bollen. \emph{Cross-National Indicators of Liberal Democracy, 1950-1990}. Inter-university Consortium for Political and Social Research (ICPSR) [distributor], 2001. DOI: \href{https://dx.doi.org/10.3886/ICPSR02532.v2}{10.3886/ICPSR02532.v2}. Original data
available at \url{https://www.icpsr.umich.edu/web/ICPSR/studies/2532/publications}.
Taken from D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM.
}
\section{Variables}{


\describe{

\item{original_country}{The country name, as in the PMM dataset. Use
\code{extended_country_name} instead if you want a consistent name.}

\item{year}{The calendar year.}

\item{bollen_pmm}{Democracy index. Ranges from 0 to 100.} }
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{fh_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{polity_pmm}},
\code{\link{prc_gasiorowski}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wgi_legacy}}
}
\concept{PMM replication data}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
