% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cite_dataset}
\alias{cite_dataset}
\title{Retrieve the citation for a specific dataset included in this package}
\usage{
cite_dataset(dataset_name, to_bibtex = FALSE)
}
\arguments{
\item{dataset_name}{The name of the dataset to cite.}

\item{to_bibtex}{Whether to provide bibtex/biblatex markup.}
}
\value{
Prints the citation information. Returns the \code{\link[RefManageR:BibEntry]{RefManageR::BibEntry()}} object.
}
\description{
This is a convenient wrapper for \code{print(bibliography)}, searching by
\code{key = dataset_name}, with \code{BibOptions(use.regex = TRUE)}. You can
also search and manipulate the \code{\link{bibliography}} object directly
using the \link[RefManageR:RefManageR-package]{RefManageR::RefManageR} package functions.
}
\examples{

cite_dataset("gwf")

cite_dataset("pmm")

cite_dataset("uds") # Same output as the previous one
}
