% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_year_coder.R
\name{country_year_coder}
\alias{country_year_coder}
\title{Country-Year State System Coding}
\usage{
country_year_coder(
  tbl,
  country_col,
  date_col,
  code_col,
  code_type = "cown",
  to_system = c("GW", "polity", "cow"),
  match_type = c("country and code", "code only"),
  include_in_output = c("extended_country_name", "GWn", "cown", "in_GW_system"),
  verbose = TRUE,
  debug = FALSE,
  match_final_year = FALSE
)
}
\arguments{
\item{tbl}{A data frame with (at least) country and year columns.}

\item{country_col}{The (unquoted) name of the country column. Defaults to
\code{country}. If missing, but \code{code_col} is given, the function defaults to
matching only on code and does not match on country names. This saves time,
but is not always possible and can introduce errors if the code column is
incorrect.}

\item{date_col}{The (unquoted) name of the year column. Defaults to \code{year}.
Must exist in \code{tbl} - otherwise just use \link[countrycode:countrycode]{countrycode::countrycode} if you don't care
about country-year matching.}

\item{code_col}{The (unquoted) name of the code column. Defaults to \code{NULL}.}

\item{code_type}{Type of code to match on (only needed if \code{code_col} is not
\code{NULL}). Defaults to "cown". Can be any unambiguous substring of the
following:

\itemize{

\item "GWn": Numeric code for the country in the Gleditsch-Ward system of
states (\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}).
This is very
similar, but not identical, to the Correlates of War and Polity codes. See
the vignette in this package for the differences.

\item "GWc": Three letter code for the country in the Gleditsch-Ward system of
states (\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}).
This is very
similar, but not identical, to the Correlates of War and Polity codes. See
the vignette in this package for an overview of the differences between
these systems.

\item "extended_GWn": Identical to \code{GWn}, but includes a number of not commonly
used codes for units of doubtful sovereignty (751 for Hyderabad before its
incorporation into India, 666.001 and 666.002 for Israel pre 1967 borders
and occupied territories, and 605 for Western Sahara). These codes are used
in some UCDP/PRIO (\url{https://www.prio.org/Data/Armed-Conflict/UCDP-PRIO/})
datasets.

\item "extended_GWc": Identical to \code{GWc}, but includes a number of not commonly
used codes for units of doubtful sovereignty (HYD for Hyderabad before its
incorporation into India, ISR.1 and ISR.2 for Israel pre 1967 borders amd
occupied territories, and SAH for Western Sahara). These codes are used in
some UCDP/PRIO (\url{https://www.prio.org/Data/Armed-Conflict/UCDP-PRIO/})
datasets.

\item "cown": Numeric code for the country in the Correlates of War system of
states (\url{https://correlatesofwar.org/data-sets/cow-country-codes/}). This
is very similar, but not identical, to the Gleditsch and Ward and Polity
codes. See the vignette in this package for an overview of the differences
between these systems.

\item "cowc": Three letter code for the country in the Correlates of War
system of
states (\url{https://correlatesofwar.org/data-sets/cow-country-codes/}). This
is very similar, but not identical, to the Gleditsch and Ward and Polity
codes. See the vignette in this package for an overview of the differences
between these systems.

\item "polity_ccode": Numeric code for the country in the Polity IV
dataset (\url{https://www.systemicpeace.org/polity/polity4.htm}). This is very
similar, but not identical, to the Gleditsch and Ward and COW codes. See
the vignette in this package for the differences.

\item "polity_scode": Three letter code for the country in the Polity IV
dataset (\url{https://www.systemicpeace.org/polity/polity4.htm}). This is very
similar, but not identical, to the Gleditsch and Ward and COW codes. See
the vignette in this package for an overview of the differences between
these systems.

\item "eurostat": Eurostat 2-letter country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "eu28": Whether the state is a member of the European Union (as of
December 2015), without special territories. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "fao": Food and Agriculture Organization of the United Nations numerical
country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "fips105": FIPS 10-5 (Federal Information Processing Standard) 2-letter
country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "icao": International Civil Aviation Organization 2-letter country code.
Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "icao_region": International Civil Aviation Organization region code.
Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "imf": International Monetary Fund numeric country code. Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "ioc": International Olympic Committee 3-letter country code. Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso2c": ISO-2 character. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso3c": ISO-3 character. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso2n": ISO-2 numeric. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso3n": ISO-3 numeric. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "un": United Nations numerical code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb": World Bank (very similar but not identical to iso3c). Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb_api2c": World Bank API 2 character code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb_api3c": World Bank API 3 character code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wvs": World Values Survey numeric code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.
}}

\item{to_system}{State system to match to. Default is Gledistsch and Ward.}

\item{match_type}{How to match countries when provided with a code column;
the possible options are "country and code" (looks at the country column as
well as the code column, if the latter is supplied) or "code only" (matches
only on the code column). The latter option should only be used if you are
confident about the code column being accurate and not missing any codes,
or if not country name column is present. Default is "country and code".
Can be abbreviated (e.g., "country" is the same as "country and code")}

\item{include_in_output}{Columns to include in output. Defaults to
c("extended_country_name", "GWn", "cown", "in_GW_system"), but can be any
character vector containing any combination of the column names in the
internal database of country codes.

The possibilities are:

\itemize{

\item "GW_country_name": The name of the country in the
Gleditsch-Ward system of states
(\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html})
(Gleditsch and War 1999).

\item "extended_country_name": The name of the country in the
Gleditsch-Ward system of states
(\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html})
(Gleditsch and War 1999), of the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad).

\item "cow_country_name": The name of the state in the Correlates of War state
system (2016 version). See
\url{https://correlatesofwar.org/data-sets/cow-country-codes/} for details. Is
\code{NA} if the country is not in the COW system of states.

\item "polity_country_name": The name of the state in the official Polity
dataset. Is \code{NA} if the country is not in Polity.

\item "country_name_en": The official name of the country (official short
English country name), as defined by the ISO organization. Taken from \link[countrycode:countrycode]{countrycode::countrycode}. See also
\url{https://en.wikipedia.org/wiki/ISO_3166-1}.
Some names are missing because the state no longer exists or there is
controversy about its sovereign status.

\item "GWn": Numeric code for the country in the Gleditsch-Ward system of
states
(\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html})
(Gleditsch and War 1999). This is very
similar, but not identical, to the Correlates of War and Polity codes. See
the vignette in this package for the differences.

\item "extended_GWn": Identical to \code{GWn}, but includes a number of not commonly
used codes for units of doubtful sovereignty (751 for Hyderabad before its
incorporation into India, 666.001 and 666.002 for Israel pre 1967 borders
amd occupied territories, and 605 for Western Sahara). These codes are used
in some UCDP/PRIO (\url{https://www.prio.org/Data/Armed-Conflict/UCDP-PRIO/})
datasets.

\item "GWc": Three letter code for the country in the Gleditsch-Ward system
of states
(\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html})
(Gleditsch and War 1999). This is very
similar, but not identical, to the Correlates of War and Polity codes. See
the vignette in this package for an overview of the differences between
these systems.

\item "extended_GWc": Identical to \code{GWc}, but includes a number of not commonly
used codes for units of doubtful sovereignty (HYD for Hyderabad before its
incorporation into India, 666.001 and 666.002 for Israel pre 1967 borders
amd occupied territories, and 605 for Western Sahara). These codes are used
in some UCDP/PRIO (\url{https://www.prio.org/Data/Armed-Conflict/UCDP-PRIO/})
datasets.

\item "cown": Numeric code for the country in the Correlates of War system of
states (\url{https://correlatesofwar.org/data-sets/cow-country-codes/}), 2016
release. This is very similar, but not identical, to the Gleditsch and Ward
and Polity codes. See the vignette in this package for an overview of the
differences between these systems.

\item "cowc": Three letter code for the country in the Correlates of War
system of
states (\url{https://correlatesofwar.org/data-sets/cow-country-codes/}), 2016
release. This is very similar, but not identical, to the Gleditsch and Ward
and Polity codes. See the vignette in this package for an overview of the
differences between these systems.

\item "polity_ccode": Numeric code for the country in the Polity IV
dataset (\url{https://www.systemicpeace.org/polity/polity4.htm}). This is very
similar, but not identical, to the Gleditsch and Ward and COW codes. See
the vignette in this package for the differences.

\item "polity_scode": Three letter code for the country in the Polity IV
dataset (\url{https://www.systemicpeace.org/polity/polity4.htm}). This is very
similar, but not identical, to the Gleditsch and Ward and COW codes. See
the vignette in this package for an overview of the differences between
these systems.

\item "regex": The regular expression used to match the country name. Useful
for debugging.

\item "GW_startdate": The entry date of the state into the international system
(usually the date of independence, though there can be more than one if the
country entered or exited the state system at different times), according
to Gleditsch and Ward, or 1 January 1816, whichever is later. Is \code{NA} if
the country was never in the Gleditsch and Ward system of states.

\item "GW_enddate": The exit date of the state from the international system
(usually the date the state lost its independence, though there can be more
than one), according to Gleditsch and Ward, or \code{NA} if the country is still
in the international system as of the time the dataset was compiled. Is
\code{NA} if the country was never in the Gleditsch and Ward system of states.

\item "cow_startdate": The entry date of the state into the international
system (usually the date of independence, though there can be more than one
if the country entered or exited the state system at different times),
according to Correlates of War, or 1 January 1800, whichever is later. Is
always \code{NA} if the country was never in the Correlates of War system of
states.

\item "cow_enddate": The exit date of the state from the international system
(usually the date the state lost its independence, though there can be more
than one), according to the Correlates of War system, or \code{NA} if the
country is still in the international system as of the time the dataset was
compiled. Is always \code{NA} if the country was never in the Correlates of War
system of states.

\item "official_region": Regions as defined in the World Bank Development
Indicators. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "extended_region": The same as "official_region", except it adds region
information for historical states. It assigns "Austria-Hungary" the
(invented) region "Central Europe", since Hungary is assigned to Eastern
Europe and Austria to Western Europe in the World Bank scheme.

\item "official_continent": Continents as defined in the World Bank Development
Indicators. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "extended_continent": The same as "official_continent", except it adds
continent information for historical states. Is \code{NA} only for Antarctica
and a number of isolated islands.

\item "lon": The rough longitude of the state, obtained by geocoding the
official country name via \link[ggmap:geocode]{ggmap::geocode} (with some manual adjustments
for historical states).

\item "lat": The rough latitude of the state, obtained by geocoding the
official country name via \link[ggmap:geocode]{ggmap::geocode} (with some manual adjustments
for historical states).

\item "microstate": Whether the state is a microstate, according to Gleditsch.
His tentative list of microstates is available via the Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.

\item "in_GW_system": Whether the state is "in system" (that is, is independent
and sovereign), according to Gleditsch and Ward, for a particular date.

\item "in_cow_system": Whether the state is "in system" (that is, is
independent and sovereign), according to the Correlates of War system, for
a particular date.

\item "in_polity_system": Whether the state is "in system" (that is, is
independent and sovereign), according to the Polity IV system, for a
particular date.

\item ar5: IPCC's regional mapping used both in the Fifth Assessment Report
(AR5) and for the Reference Concentration Pathways (RCP). Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item continent: Continent as defined in the World Bank Development Indicators.
Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "eurocontrol_pru": European Organisation for the Safety of Air Navigation
region. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "eurocontrol_statfor": European Organisation for the Safety of Air
Navigation region. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "eurostat": Eurostat 2-letter country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "eu28": Whether the state is a member of the European Union (as of
December 2015), without special territories. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "fao": Food and Agriculture Organization of the United Nations numerical
country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "fips105": FIPS 10-5 (Federal Information Processing Standard) 2-letter
country code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "icao": International Civil Aviation Organization 2-letter country code.
Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "icao_region": International Civil Aviation Organization region code. Taken
from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "imf": International Monetary Fund numeric country code. Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "ioc": International Olympic Committee 3-letter country code. Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso2c": ISO-2 character. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso3c": ISO-3 character. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso2n": ISO-2 numeric. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "iso3n": ISO-3 numeric. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "un": United Nations numerical code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb": World Bank (very similar but not identical to iso3c). Taken from
\link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb_api2c": World Bank API 2 character code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wb_api3c": World Bank API 3 character code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "wvs": World Values Survey numeric code. Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_ar": country name (Arabic). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_de": country name (German). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_es": country name (Spanish). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_fr": country name (French). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_ru": country name (Russian). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.

\item "country_name_zh": country name (Chinese). Taken from \link[countrycode:countrycode]{countrycode::countrycode}.
}}

\item{verbose}{Prints information about countries matched multiple times,
name changes, and other problems encountered while processing the data.
Default is \code{TRUE}.}

\item{debug}{Returns additional columns useful for debugging. Default is
\code{FALSE}.}

\item{match_final_year}{Whether to classify the final year of a country as
"in system". For example, if \code{TRUE}, the country-year pair "Republic of
Vietnam 1975"  will appear as "in system" in 1975, though the country ended
in 1975; if \code{FALSE} it will appear as not in system, as the Republic of
Vietnam (South Vietnam) ceased to exist by the end of 1975. Use with care;
this may result in multiple matches. Default is \code{FALSE}.}
}
\value{
a tidy data frame of country-years merged with information about the
system of states (GWn, COW, or Polity IV), country codes, and standardized
country names.
}
\description{
Given a set of country-years, find appropriate country codes, standardized
country names, and determine state system membership
}
\examples{
\donttest{
library(dplyr)
bmr <- redownload_bmr(return_raw = TRUE)
country_year_coder(bmr, country_col = country, date_col = year,
                   code_col = ccode, code_type = "polity_ccode")}
}
