% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_panel}
\alias{create_panel}
\title{Create a panel of countries belonging in the state system.}
\usage{
create_panel(system = c("cow", "GW", "polity"), max_year = NULL)
}
\arguments{
\item{system}{The state system to target. The possibilities are "cow"
(\url{https://correlatesofwar.org/data-sets/state-system-membership/}, the
default), "GW" (Gleditsch and Ward, from Gleditsch, Kristian S. & Michael
D. Ward. 1999. "Interstate System Membership: A Revised List of the
Independent States since 1816." International Interactions 25: 393-413,
available via the Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}),
or "polity" (for the \url{https://www.systemicpeace.org/polityproject.html}
system of states).}

\item{max_year}{The maximum year in the panel. If NULL, uses the current year.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the country code in the relevant system of states,
the official country name and years when the country belongs to the system,
and the relevant start and end dates for entry and exit into the system.
}
\description{
Create a panel of countries belonging in the state system.
}
\examples{
create_panel("cow") # The correlates of War system is the default
create_panel("GW") # Use this for the Gleditsch and Ward system
create_panel("polity", max_year = 1997) # You can specify a maximum year
}
