% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\name{democracy_scores}
\alias{democracy_scores}
\title{Extract UD scores from a UD model}
\usage{
democracy_scores(model, adjust_to_dichotomous = TRUE, as_prob = TRUE)
}
\arguments{
\item{model}{a \link[mirt:SingleGroupClass-class]{mirt::SingleGroupClass} model.}

\item{adjust_to_dichotomous}{Whether to calculate an adjusted score where the
midpoint represents the average cutpoint for dichotomous scores. See the
vignette for more details. Default is \code{TRUE}.}

\item{as_prob}{Whether to output scores as 0-1 probability scales. See the
vignette for more details. Default is \code{TRUE}.}
}
\value{
A data frame with latent variable democracy scores (the equivalent of
the UDS posterior means) for all country-years in the data, with standard
errors and 95\% confidence intervals.
}
\description{
This function is a simple wrapper for \code{fscores(model, full.scores =
TRUE, full.scores.SE = TRUE, ...)} that returns scores in a tidy data frame
instead of a matrix. More fine-grained control can be achieved by using
\link[mirt:fscores]{mirt::fscores} directly.
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\examples{
\donttest{
# Replicate the official UDS scores (2011 release)
library(dplyr)
library(mirt)
democracy_data <- generate_democracy_scores_dataset(
  selection = "_pmm", output_format = "wide", exclude_pmm_duplicates = FALSE
  )

democracy_data <- prepare_democracy_data(democracy_data)

replication_2011_model <- mirt(democracy_data |>
  select(matches("pmm")), model = 1, itemtype = "graded", SE = TRUE, verbose = FALSE)

democracy_scores(model = replication_2011_model)

}
}
