% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/democracyData-package.R
\docType{package}
\name{democracyData-package}
\alias{democracyData}
\alias{democracyData-package}
\title{democracyData}
\description{
This package archives a large number of democracy datasets and their
documentation, many of historical interest. It also contains a number of
functions to download and process the original source data and to determine
membership of a country in standard membership lists of the state system
(according to Gleditsch and Ward, Correlates of War, or Polity).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xmarquez.github.io/democracyData/}
  \item \url{https://github.com/xmarquez/democracyData}
  \item Report bugs at \url{https://github.com/xmarquez/democracyData/issues}
}

}
\author{
\strong{Maintainer}: Xavier Marquez \email{marquez.x@gmail.com} (\href{https://orcid.org/0000-0001-7653-9334}{ORCID})

}
