% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_wgi_voice_and_accountability}
\alias{download_wgi_voice_and_accountability}
\alias{wgi}
\alias{wgi_democracy}
\title{Democracy, Voice, and Accountability Index from the World Governance
Indicators}
\source{
Worldwide Governance Indicators. Data and documentation available at \url{https://www.worldbank.org/en/publication/worldwide-governance-indicators}.
}
\usage{
download_wgi_voice_and_accountability(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)
}
\arguments{
\item{url}{The URL of the dataset. Defaults to \url{https://www.worldbank.org/content/dam/sites/govindicators/doc/wgidataset_with_sourcedata-2025.xlsx}.}

\item{verbose}{Whether to print a running commentary of what the function is
doing while processing the data.}

\item{return_raw}{Whether to return the raw data, with minimal processing.
Default is \code{FALSE}.}

\item{...}{Other parameters passed to \link{country_year_coder}.}
}
\value{
A dataset containing the country-year version of the latest update
(WGI 2.0, covering 1996-2024) of the World Governance Indicators
"Democracy, Voice and Accountability" index, compiled by Daniel Kaufmann
and Aart Kraay. The original data are available at
\url{https://www.worldbank.org/en/publication/worldwide-governance-indicators}, along with a full explanation of
the methodology used to compile the index.

According to the compilers, "Voice and accountability captures perceptions
of the extent to which a country's citizens are able to participate in
selecting their government, as well as freedom of expression, freedom of
association, and a free media."

\describe{

\item{wb_country}{Country name (World Bank standard).}

\item{wb_code}{World bank 3 character country code.}

\item{year}{Year.}

\item{Estimate}{Estimate of governance (ranges from approximately -2.5
(weak) to 2.5 (strong) governance performance), with a mean of 0.}

\item{StdErr}{Standard error reflects variability around the point estimate
of governance.}

\item{NumSrc}{Number of data sources on which estimate is based. The
methodology document lists, among these, the EIU Democracy Index, Freedom
House, the World Economic Forum Global Competitiveness Report, the Gallup
World Poll, the Institutional Profiles Database, the Political Risk
Services International Country Risk Guide, and Reporters Without Borders
Press Freedom Index. See \url{https://www.worldbank.org/content/dam/sites/govindicators/doc/va.pdf} for more
details.}

\item{Rank}{Percentile rank among all countries (ranges from 0 (lowest) to
100 (highest) rank).}

\item{Lower}{Lower bound of 90\% confidence interval for governance, in
percentile rank terms.}

\item{Upper}{Upper bound of 90\% confidence interval for governance, in
percentile rank terms} }
}
\description{
Download the latest version (WGI 2.0, 2025 revision, with data up to 2024) of
the Democracy, Voice, and Accountability index from the World Governance
Indicators and process it via \link{country_year_coder}.
}
\details{
\strong{Methodology break (WGI 2.0, 2025 revision).} In 2025 the World Bank
released a full methodological revision of the WGI. The source pool was
expanded (now 35 sources including the Arab, Asian, and Euro barometers), the
aggregation model was changed, and all estimates were recomputed back to
1996. Scores obtained through this function are therefore \strong{not directly
comparable} to scores produced by earlier versions of this package (any
\code{democracyData} release before 0.7.0). Use \link{wgi_legacy} for the archived
pre-0.7.0 series (1996-2023) with the same column schema. See the WGI
methodology PDF linked from the documentation URL for details.
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note}{


The datasets downloaded by the \verb{download_*} family of functions are not
directly available in this package. You will always need to directly
download them in order to use them.
}

\examples{
\dontrun{
download_wgi_voice_and_accountability()}
}
\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{eiu}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wgi_legacy}}
}
\concept{continuous democracy indexes}
\concept{democracy}
