% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_uds_documentation.R
\docType{data}
\name{extended_uds}
\alias{extended_uds}
\title{Extended UDS}
\format{
A \code{\link[tibble:tibble]{tibble::tibble()}} with
41986 rows and
20 columns.

A dataset extending the Unified Democracy Scores of Pemstein, Meserve, and
Melton (2010) to the 19th century (and sometimes before), updating it with
the latest democracy datasets included in \code{democracyData} through
2025, and
calculating scores for countries not in the official UD release. Cite both
Pemstein, Meserve, and Melton 2010 as well as Marquez 2016.
}
\usage{
extended_uds
}
\description{
Extended UDS
}
\section{Overview}{


This dataset contains an extension of the Unified Democracy Scores of
Pemstein, Meserve, and Melton (2010). PMM 2010 use a latent variable
approach to combine diverse measurements of democracy for a broad panel of
countries from 1946 to 2012; this dataset extends the scores by using more
democracy measures with broader temporal and spatial coverage, generating
democracy scores for
41986
country-years
(299
distinct countries and non-sovereign territories). It is generated by the
function \link{generate_extended_uds}.

The measurement of democracy is complicated and controversial. The extended
UDS makes few judgments about what measures of democracy should be used to
generate latent democracy scores, as long as they have been used in
scholarly work. It uses dichotomous, trichotomous, ordinal, and continuous
indices; indices that focus primarily on the "competition" and indices that
focus on the "participation" dimension of democracy; "thick" indices that
attempt to measure a wide variety of characteristics plausibly attributed
to democracy; and "minimalist" indexes that restrict themselves to the bare
minimum of competition. Nevertheless, most of these indexes agree that
democracy has something to do with competition and participation, even if
they weight these dimensions somewhat differently, and even if they include
other things, such as civil rights.

The latent variable score is calculated using the democracy measures from
the following datasets: \link{anckar}, \link{anrr}, \link{arat}, \link{blm}, \link{bmr}
(\code{democracy_femalesuffrage} indicator only), \link{bnr_extended}, \link{bti},
\link{bollen_pmm}, \link{svmdi} (continuous indicator), \link{doorenspleet}, \link{eiu}, \link{fh}
(\code{fh_total_reversed} indicator), \link{fh_electoral}, \link{gwf_all_extended},
\link{hadenius_pmm}, \link{LIED}, \link{magaloni_extended}, \link{mainwaring}, \link{munck_pmm},
\link{pacl}, \link{pacl_update}, \link{peps} (PEPS1v indicator), \link{pitf} (non-binary
indicator), \link{polity} (polity 2 indicator), \link{polyarchy} (both contestation
and participation dimensions), \link{polyarchy_dimensions} (both contestation
and participation dimensions), \link{prc}, \link{REIGN}, \link{svolik},
\link{ulfelder_extended}, \link{utip} (trichotomous indicator, with one-party
democracy as the middle category), the \code{v2x_polyarchy} indicator from
\link{vdem_simple}, \link{vanhanen} (the \code{vanhanen_democratization} indicator only),
\link{vaporeg} (the \code{vaporeg_trichotomous} indicator), \link{wgi}, and
\link{wahman_teorell_hadenius} (the \code{wth_democrobust} variable). Please refer
to the documentation of \link{generate_democracy_scores_dataset} for more
details on these measures of democracy, and to Marquez 2016 for more
details on the construction of the extended latent variable model used to
generate the scores.
}

\section{Variables}{


\describe{

\item{z1, z1_adj, z1_as_prob, z1_adj_as_prob}{The latent democracy score.
\code{z1} is the "raw" (unadjusted) score, with mean approximately 0 and sd 1;
\code{z1_adj} is the version adjusted so that 0 represents the average cutpoint
for the dichotomous indicators composing the index; the \verb{_as_prob} versions
adjust these to the 0-1 range by applying \link[stats:Normal]{stats::pnorm} (treating them as
probability measures).}

\item{se_z1}{The standard error of the score.}

\item{_pct025, _pct975}{The latent democracy score 95\% confidence
intervals.}

\item{num_measures}{The number of measures that contributed to the index
for the particular country-year. Varies from
1
to
31;
one can filter the dataset to include only scores where more than, say, 5
measures have contributed to the index.}

\item{measures}{The names of the specific measures that contributed to the
index in that particular country-year. This is a list-column; use
\link[tidyr:unnest]{tidyr::unnest} to unnest. See the documentation for
\link{generate_democracy_scores_dataset} for specific info about any particular
measure.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\references{
X. Márquez. "A Quick Method for Extending the Unified Democracy Scores". In: \emph{Available at SSRN 2753830} (2016). DOI: \href{https://dx.doi.org/10.2139/ssrn.2753830}{10.2139/ssrn.2753830}.

D. Pemstein, S. Meserve, and J. Melton. "Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type". In: \emph{Political Analysis} 18.4 (2010), pp. 426-449. DOI: \href{https://dx.doi.org/10.1093/pan/mpq020}{10.1093/pan/mpq020}.
}
\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other latent variable indexes of democracy: 
\code{\link{polyarchy_dimensions}},
\code{\link{uds_2014}}
}
\author{
Xavier Marquez, Political Science and International Relations
Programme, Victoria University of Wellington, marquez.x@gmail.com
}
\concept{democracy}
\concept{latent variable indexes of democracy}
\keyword{datasets}
