% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\name{generate_extended_uds}
\alias{generate_extended_uds}
\title{Generates the extended UDS scores from the latest democracy data in this
package}
\usage{
generate_extended_uds(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to print diagnostic information about what the
function is doing, including information from \link[mirt:mirt]{mirt::mirt} about the model
being fit. Default is \code{FALSE}.}
}
\value{
A \link{tibble} described in \link{extended_uds}. Right now there's no
flexibility at all to choose the variables to generate this data, though
this might change in the future. See
\code{vignette("Replicating_and_extending_the_UD_scores")} for more info on how
to generate these scores for arbitrary measures.
}
\description{
Generates the extended UDS scores from the latest democracy data in this
package
}
\examples{
\dontrun{
extended_uds <- generate_extended_uds()
}
}
