% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gwf_all}
\alias{gwf_all}
\alias{gwf_all_extended}
\alias{gwf}
\alias{gwf_autocratic}
\alias{gwf_autocratic_extended}
\title{The Geddes Wright and Frantz Autocratic Regimes dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7954 rows and 20 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9411 rows and 20 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4591 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5301 rows and 21 columns.
}
\source{
B. Geddes, J. Wright, and E. Frantz. "Autocratic Breakdown and Regime Transitions: A New Data Set". In: \emph{Perspectives on Politics} 12.1 (2014), pp. 313-331. DOI: \href{https://dx.doi.org/10.1017/S1537592714000851}{10.1017/S1537592714000851}. The full data and codebook can be
downloaded here: \url{https://sites.psu.edu/dictators/}.
}
\usage{
gwf_all

gwf_all_extended

gwf_autocratic

gwf_autocratic_extended
}
\description{
The dataset described in B. Geddes, J. Wright, and E. Frantz. "Autocratic Breakdown and Regime Transitions: A New Data Set". In: \emph{Perspectives on Politics} 12.1 (2014), pp. 313-331. DOI: \href{https://dx.doi.org/10.1017/S1537592714000851}{10.1017/S1537592714000851}. The full data and codebook can be
downloaded at \url{https://sites.psu.edu/dictators/}. The documentation
below is derived from the codebook. Some variables apply to the country-year
version, some to the case version and some to the "extended" version
(extended back before 1945 with info from the case descriptions).
}
\details{
\describe{

\item{gwf_cowcode}{Correlates of War (CoW) country code}

\item{year}{Calendar year. In the "extended_yearly" version of
the dataset, these have been calculated back before 1945 from the information
available in the casenames.}

\item{gwf_country}{Country name}

\item{gwf_casename}{Autocratic regime case name (country name and years);
these are the units of observation for duration analysis}

\item{gwf_regimetype}{Regime type: not independent, foreign-occupied,
warlord, democracy, provisional, monarchy, personalist, military, party,
party-personalist, party-military, military-personalist,
party-military-personalist. Only includes non-democratic regime types.}

\item{gwf_duration}{Duration of regime up to
time t. In the "extended" datasets this always starts at 1.}

\item{gwf_fail}{Binary indicator for regime failure}

\item{gwf_disagree}{Binary indicator for country-year observations where GWF
coding of democracy differs from \href{pacl.html}{CGV} (here \link{pacl}) coding of democracy}

\item{gwf_startdate}{Day-Month-Year for the calendar date of the autocratic
regime start event. Only in \code{gwf_autocratic} datasets.}

\item{gwf_enddate}{Day-Month-Year for the calendar date
of the autocratic regime failure event (31-12-2010 for right-censored
autocratic regimes). Only in \code{gwf_autocratic} datasets.}

\item{gwf_spell}{Time-invariant duration of autocratic
regime}

\item{gwf_fail_subs}{Categorical variable marking the subsequent
regime type. 1: subsequent regime is democracy 2: subsequent regime is
autocratic 3: subsequent regime is warlord, foreign-occupied or ceases to
exist 0: no regime failure at duration time t ; and regime still in power
December 31, 2010}

\item{gwf_fail_type}{Categorical variable marking how the
autocratic regime ends. 1: regime insiders change rules of regime 2:
incumbent loses elections 3: no incumbent runs in competitive election won by
opponent 4: popular uprising 5: military coup 6: insurgents, revolutionaries,
or combatants fighting a civil war 7: foreign imposition or invasion 8: new
autocratic leader selected, changes rules, and remains in power 9: state
ceases to exist ends or government fails to control most of the country's
territory 0: regime still in power on December 31, 2010}

\item{gwf_fail_violent}{Categorical variable marking the level of violence
during the autocratic regime failure event. 1: no deaths 2: 1-25 deaths 3:
26-1000 deaths 4: >1000 0: regime still in power on December 31, 2010}

\item{gwf_party}{Binary indicator of party regime type (groups party-based,
party-personal, party-military, party-personal-military, oligarchy, and Iran
1979-2010)}

\item{gwf_personal}{Binary indicator of personalist regime type}

\item{gwf_military}{Binary indicator of military regime type (groups
military, military-personal, indirect military)}

\item{gwf_monarchy}{Binary
indicator of monarchy regime type} }
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note}{
 The "extended" datasets \verb{*_extended} are extended back in time
using the \code{gwf_duration} variable, so they include years before 1946.The
original country-year datasets did not include the period before 1946
explicitly.
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other authoritarianism: 
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}
}
\concept{authoritarianism}
\concept{democracy}
\keyword{datasets}
