% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hadenius_pmm}
\alias{hadenius_pmm}
\alias{hadenius}
\title{Axel Hadenius' Index of Democracy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 129 rows and 7 columns.
}
\source{
D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM.
}
\usage{
hadenius_pmm
}
\description{
0-10 index of democracy from A. Hadenius. \emph{Democracy and Development}. New York: Cambridge University Press, 1992.
Taken from D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM. Higher values are more democratic.
}
\section{Variables}{


\describe{

\item{pmm_country}{The original country name in the PMM replication data.}

\item{year}{The calendar year. This is approximate. The surveys cover
specific periods in the original data that do not always overlap with a
single year. In particular, the year 1981 is "skipped" -- a single survey
covers Jan.1981 - Aug. 1982 and its value is assigned to 1982 here.}

\item{hadenius_pmm}{0-10 index of democracy from A. Hadenius. \emph{Democracy and Development}. New York: Cambridge University Press, 1992.
Only available for 129 countries in 1988.} }
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bollen_pmm}},
\code{\link{fh_pmm}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{polity_pmm}},
\code{\link{prc_gasiorowski}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wgi_legacy}}
}
\concept{PMM replication data}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
