% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mainwaring}
\alias{mainwaring}
\alias{mainwaring_pmm}
\title{Mainwaring, Brinks, and Perez Liñán's democracy measure for Latin America}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2156 rows and 14 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 835 rows and 7 columns.
}
\source{
S. Mainwaring, A. Pérez-Liñán, and D. Brinks. "Political Regimes in Latin America, 1900-2007 (with Daniel Brinks)". In: \emph{Democracies and Dictatorships in Latin America: Emergence, Survival, and Fall}. New York: Cambridge University Press, 2014. \url{https://web.archive.org/web/20120119050029/http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf}.

D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM.
}
\usage{
mainwaring

mainwaring_pmm
}
\description{
Trichotomous democracy indicator from S. Mainwaring, D. Brinks, and A. Pérez-Liñán. "Classifying Political Regimes in Latin America". In: \emph{Studies in Comparative International Development} 36.1 (2001), pp. 37-65. DOI: \href{https://dx.doi.org/10.1007/bf02687584}{10.1007/bf02687584}.
Extended in S. Mainwaring, A. Pérez-Liñán, and D. Brinks. "Political Regimes in Latin America, 1900-2007 (with Daniel Brinks)". In: \emph{Democracies and Dictatorships in Latin America: Emergence, Survival, and Fall}. New York: Cambridge University Press, 2014. \url{https://web.archive.org/web/20120119050029/http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf}. Original data description
archived at \url{https://web.archive.org/web/20120119050029/http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf}.
}
\section{Variables}{


\describe{

\item{mainwaring_country}{The original country name.}

\item{year}{The calendar year. }

\item{mainwaring}{Trichotomous numerical democracy indicator from S. Mainwaring, D. Brinks, and A. Pérez-Liñán. "Classifying Political Regimes in Latin America". In: \emph{Studies in Comparative International Development} 36.1 (2001), pp. 37-65. DOI: 10.1007/bf02687584.  S. Mainwaring, A. Pérez-Liñán, and D. Brinks. "Political Regimes in Latin America, 1900-2007 (with Daniel Brinks)". In: \emph{Democracies and Dictatorships in Latin America: Emergence, Survival, and Fall}. New York: Cambridge University Press, 2014. \url{https://web.archive.org/web/20120119050029/http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf}.
0 = non-democracy, 0.5 = hybrid, 1 = democracy.}

\item{Regime}{Trichotomous democracy indicator. From the published
article:

When governments commit no violation of any of the four criteria
(elections, franchise, civil liberties, and civilian power), they are coded
as democratic. They rank as authoritarian if they present one or more major
violations or as semi-democratic if they display only partial failures in
one or more categories.}

\item{Elections}{Measure of whether the principle of free and fair
elections is violated. MV = major violation, NV = No violation, PV =
partial violation. From the published article:

A major violation of this democratic principle occurs if:
\itemize{
\item the head of government or the legislature is not elected.
\item the government uses its resources (patronage, repression, or a
combination of both) to ensure electoral victory--i.e., there are
systematic complaints about fraud or repression, and there is virtual
certainty about the outcome of presidential elections (e.g., Mexico
1945-88, Argentina 1952-55, E1 Salvador 1952-63, Paraguay 1960-89).
\item through fraud, manipulation, or outright repression, the government makes
it impossible for a wide gamut of parties to compete (or if they do
compete, to take office).
}

A partial violation occurs if:
\itemize{
\item there are systematic complaints of rigged elections and/or harassment of
the opposition but there is still uncertainty about electoral outcomes and
the government fails to capture large majorities in the legislature; or the
military vetoed a few "unacceptable" but important presidential candidates
(e.g., Argentina 1958-66); fraud affected but did not thoroughly skew
electoral results; or the elections were conducted under substantially
unequal playing rules (e.g., Nicaragua in 1984 because the Sandinistas
dominated the media and pressured opposition groups, E1 Salvador in the
1980s because the left faced massive repression). }
}

\item{Franchise}{Measure of whether the principle of inclusive franchise is
violated. MV = major violation, NV = No violation, PV = partial violation.
NA if the electoral principle is coded MV. From the published article:

A major violation of this democratic principle occurs if a large part of
the adult population is disenfranchised on ethnic, class, gender, or
educational grounds in ways that:
\itemize{
\item likely prevent very different electoral outcomes (or so is widely
believed); or
\item are unusually exclusionary for that historical period; or
\item trigger mass social protests.
}

A partial violation occurs if disenfranchisement of some social groups
occurs in ways that are not likely to significantly shape electoral
outcomes. }

\item{Civil.Liberties}{Measure of whether the civil liberties are violated.
MV = major violation, NV = No violation, PV = partial violation. NA if the
electoral principle is coded MV. From the published article;

A major violation of democratic principles occurs if:
\itemize{
\item gross human rights violations or censorship against opposition media
occur systematically; or
\item political parties are not free to organize--i.e., most major parties are
banned, just a single party is allowed to exist, or a few parties are
tightly controlled by the government (e.g., Panama 1968-1980, Paraguay
1947-59, Brazil 1965-79).
}

A partial violation occurs if:
\itemize{
\item violations of human rights are less widespread but still affect the
opposition's capacity to organize in some geographic areas or some social
sectors; or
\item there is intermittent censorship of the media or regular prohibition of
one major party or candidate. }
}

\item{Civilian.Power}{Measure of whether the elected rulers enjoy real
governing capacity (e.g., are not overruled by the military). MV = major
violation, NV = No violation, PV = partial violation. NA if the electoral
principle is coded MV. From the published article:

A major violation of this democratic principle occurs if:
\itemize{
\item military leaders or the military as an institution openly dominate major
policy areas not strictly related to the armed forces; or
\item if the elected head of government is a puppet, such that the electoral
process does not really determine who governs.
}

A partial violation occurs if military leaders or the military as an
institution are able to veto important policies in a few areas not related
to the armed forces (e.g., Ecuador 1961-62).
}

\item{From}{First year of the regime.}

\item{To}{Last year of the regime.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note}{


The \link{mainwaring_pmm} version of this data is taken from D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: 10.7910/DVN/WWYOHU. HDL: 1902.1/PMM.
It only contains a numeric version of the
Mainwaring Regime indicator, ranging from -1 (Authoritarian) to 0
(Semidemocracy) to 1 (Democracy). The original dataset is found in
\link{mainwaring}. PMM's replication data is actually missing a fair number
of the data points in the original dataset, but the original dataset is not
missing any of their data points, and there are no differences between the
data points wherever both the original and the replication data have values.
It is included here for completeness.
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bollen_pmm}},
\code{\link{fh_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{polity_pmm}},
\code{\link{prc_gasiorowski}}

Other regional democracy indexes: 
\code{\link{blm}}

Other trichotomous democracy indexes: 
\code{\link{blm}},
\code{\link{kailitz}},
\code{\link{prc_gasiorowski}},
\code{\link{utip}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}}
}
\concept{PMM replication data}
\concept{democracy}
\concept{regional democracy indexes}
\concept{trichotomous democracy indexes}
\keyword{datasets}
