% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redownload_functions.R
\name{redownload_lied}
\alias{redownload_lied}
\title{Downloads the 2022 update (v. 6.4) of the Lexical Index of Electoral Democracy (annual time series, data to
2021) and processes it using \link{country_year_coder}.}
\source{
S. Skaaning, J. Gerring, and H. Bartusevičius. "A Lexical Index of Electoral Democracy". In: \emph{Comparative Political Studies} 48.12 (2015), pp. 1491-1525. DOI: 10.1177/0010414015581050.
Original data and variable descriptions available at
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/WPKNIT}
}
\usage{
redownload_lied(url, verbose = TRUE, return_raw = FALSE, ...)
}
\arguments{
\item{url}{The URL of the dataset. Defaults to
\url{https://dataverse.harvard.edu/api/access/datafile/6290760}}

\item{verbose}{Whether to print a running commentary of what the function is
doing while processing the data.}

\item{return_raw}{Whether to return the raw data, without any processing.
Default is \code{FALSE}.}

\item{...}{Other parameters passed to \link{country_year_coder}.}
}
\description{
Downloads the 2022 update (v. 6.4) of the Lexical Index of Electoral Democracy (annual time series, data to
2021) and processes it using \link{country_year_coder}.
}
\examples{
\dontrun{
redownload_lied()}
}
