% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Removes empty rows (rows with all NAs) from a data frame}
\usage{
remove_empty_rows(df, vars)
}
\arguments{
\item{df}{A data frame}

\item{vars}{The variables to select}
}
\value{
The original data frame, with the empty rows for the variables
selected removed
}
\description{
Removes empty rows (rows with all NAs) from a data frame
}
\examples{
df <- data.frame(a = c(1,2,3,NA, 5), b = c(NA,1,2,NA, 3), c = 1:5)
df \%>\% remove_empty_rows(a:b)

# Different from
na.omit(df)
}
