% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{utip}
\alias{utip}
\title{The University of Texas Inequality Project Categorical Dataset of Political
Regimes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6215 rows and 18 columns.
}
\source{
S. Hsu. "The Effect of Political Regimes on Inequality, 1963-2002". In: \emph{UTIP Working Paper} (2008).
}
\usage{
utip
}
\description{
The University of Texas Inequality Project Categorical Dataset of Political
Regimes. Described in S. Hsu. "The Effect of Political Regimes on Inequality, 1963-2002". In: \emph{UTIP Working Paper} (2008).
\url{http://utip.gov.utexas.edu/papers/utip_53.pdf}. Data available for
download at \url{https://utip.gov.utexas.edu/datasets.html}.
}
\section{Description}{


This is a categorical dataset of regime types, used in papers on inequality
by the UTIP group. The coding rules are not well documented. The category
"one party democracy" is particularly difficult to understand. The
categories are nevertheless mutually exclusive.

\describe{

\item{utip_country}{The original country name in the dataset.}

\item{year}{Calendar year}.

\item{other_dictatorship}{Whether the regime is a non-military
dictatorship. 0 = no, 1 = yes.}

\item{military_dictatorship}{Whether the regime is a military dictatorship.
0 = no, 1 = yes.}

\item{civil_war}{Whether the regime is in a civil war. 0 = no, 1 = yes.}

\item{one_party_democracy}{Whether the regime is a "one party democracy",
where "fair" elections are held but only one party competes. 0 = no, 1 =
yes.}

\item{communist}{Whether the regime is communist. 0 = no, 1 = yes.}

\item{islamic_republic}{Whether the regime is an Islamic republic. 0 = no,
1 = yes. In practice, only Libya and Iran qualify.}

\item{social_democracy}{Whether the regime is social democracy, using the
Esping-Andersen (1990) classification of welfare states. 0 = no, 1 = yes.}

\item{conservative_democracy}{Whether the regime is a democracy with a
conservative or liberal welfare state, using the Esping-Andersen (1990)
classification of welfare states. 0 = no, 1 = yes.}

\item{utip_dichotomous}{Calculated, not in original dataset. Dichotomous
index of whether the regime is a social democracy, conservative democracy,
or one party democracy. 0 = no, 1 = yes.}

\item{utip_dichotomous_strict}{Calculated, not in original dataset.
Dichotomous index of whether the regime is a social democracy or a
conservative democracy (excludes one party democracy). 0 = no, 1 = yes.}

\item{utip_trichotomous}{Calculated, not in original dataset. Trichotomous
index of democracy: 2= regime is a social democracy or a conservative
democracy (excludes one party democracy), 1 = regime is a one party
democracy, 0 = other.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other authoritarianism: 
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{gwf_all}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{wahman_teorell_hadenius}}

Other dichotomous democracy indexes: 
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{kailitz}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{wahman_teorell_hadenius}}

Other trichotomous democracy indexes: 
\code{\link{blm}},
\code{\link{kailitz}},
\code{\link{mainwaring}},
\code{\link{prc_gasiorowski}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}}
}
\concept{authoritarianism}
\concept{democracy}
\concept{dichotomous democracy indexes}
\concept{trichotomous democracy indexes}
\keyword{datasets}
