% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vaporeg}
\alias{vaporeg}
\title{The Varieties of Political Regimes (VaPoReg) dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30674 rows and 40 columns.
}
\source{
S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Dataset. Version 3.2}. Hannah Arendt Institute for Totalitarianism Studies. Dresden, 2026. Dataset download at
\url{https://www.va-poreg.de/download/Kailitz,\%20Varieties\%20of\%20Political\%20Regimes\%20March\%202026.rds}.

S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Codebook. Version 3.2}. Tech. rep. Dresden: Hannah Arendt Institute for Totalitarianism Studies, 2026. Codebook PDF at
\url{https://www.va-poreg.de/download/Kailitz\%2C\%20Varieties\%20of\%20Political\%20Regimes\%2C\%20Codebook\%203.2.pdf}.
}
\usage{
vaporeg
}
\description{
\code{vaporeg} contains the March 2026 VaPoReg release in its current
country-year schema. The package does not rename these variables into the
legacy \code{democracyData} interface; use \link{vaporeg_2024} for the archived legacy
schema.
}
\section{Introduction}{


VaPoReg classifies political regimes from 1900 to 2025 using multiple
classification schemes, including detailed, compact, quadruple, triple,
binary-democracy, and binary-autocracy versions. It also codes dominant
political authorities in dependent entities, monarchy status, regime spell
timing, and population. The documentation below is based on the VaPoReg
Codebook, Version 3.2 (March 2026), especially the variable list on
pages 163-167.
}

\section{Variables}{


\describe{

\item{country_name}{Political unit name.}

\item{year}{Calendar year. Annual codings use July 1 as the reference date.}

\item{vaporeg_regtype_detailed}{Detailed regime classification. Codes:
10 Liberal Democracy; 20 Electoral Democracy; 30 Electoral Oligarchy;
40 Democratizing Regime; 50 Electoral Hybrid Regime; 60 Electoral
Autocracy; 70 Electoral Oligarchical Autocracy; 80 Non-Electoral
Transitional Regime; 90 Constitutional Monarchy; 100 Autocratic Monarchy;
110 One-Party Autocracy; 120 Right-Wing Autocracy; 130 Communist
Ideocracy; 140 Islamist Ideocracy; 150 Military Autocracy; 160 Personalist
Autocracy; 170 Direct Rule Colonial Regime; 180 Indirect Rule Colonial
Regime; 190 Direct Rule Occupation Regime; 200 Indirect Rule Occupation
Regime; 210 No Central Authority; 220 Part of Other Country.}

\item{vaporeg_regtype_reports}{Country reports regime classification. Uses
the same 22-category coding as \code{vaporeg_regtype_detailed}.}

\item{vaporeg_regtype_compact}{Compact regime classification. Codes:
10 Democracy; 20 Electoral and Liberal Hybrid Regime; 30 Provisional
Regime; 40 Electoral Autocracy; 50 Ruling Monarchy; 60 One-Party
Autocracy; 70 Right-Wing Autocracy; 80 Communist Ideocracy; 90 Islamist
Ideocracy; 100 Military Autocracy; 110 Personalist Autocracy;
120 Colonial Regime; 130 Occupation Regime; 140 No Central Authority.}

\item{vaporeg_regtype_quadruple}{Four-category regime classification:
1 Democracy; 2 Hybrid Regime; 3 Electoral Autocracy; 4 Closed Autocracy.}

\item{vaporeg_regtype_triple}{Three-category regime classification:
1 Democracy; 2 Hybrid Regime; 3 Autocracy.}

\item{vaporeg_regtype_bindem}{Binary democracy classification:
0 Non-Democracy; 1 Democracy.}

\item{vaporeg_regtype_binaut}{Binary autocracy classification:
0 Non-Autocracy; 1 Autocracy.}

\item{vaporeg_identifier}{VaPoReg unit identifier. The codebook describes
this as an identifier based on the Correlates of War dataset.}

\item{cowcode}{Correlates of War country code, where available.}

\item{vaporeg_dpa_country}{Dominant political authority in cases of dependence.}

\item{vaporeg_dpa_regdetailed}{Detailed regime classification of the
dominant political authority. It uses the same coding structure as the
detailed regime scale, but the current codebook labels category 20 as
Defective Democracy and category 80 as Non-Electoral Autocratic
Transitional Regime.}

\item{vaporeg_dpa_regcompact}{Compact regime classification of the dominant
political authority.}

\item{vaporeg_monarchy}{Monarchy status. Codes: 10 Autocratic Monarchy;
20 Governed by Autocratic Monarchy; 30 Constitutional Monarchy;
40 Governed by Constitutional Monarchy; 50 Ceremonial Monarchy;
60 Governed by Ceremonial Monarchy; 70 Republic; 80 Governed by Republic.}

\item{vaporeg_regtype_rwasubtype}{Right-wing autocracy subtype.
Codes: 0 Not applicable; 1 Fascist; 2 Corporatist.}

\item{vaporeg_continent}{Continent code. Codes: 0 missing or not assigned;
1 Africa; 2 Americas; 3 Asia; 4 Europe; 5 Oceania.}

\item{vaporeg_region}{UN-style region code. Codes: 0 missing or not
assigned; 1 Australia and New Zealand; 2 Caribbean; 3 Central America;
4 Central Asia; 5 Eastern Africa; 6 Eastern Asia; 7 Eastern Europe;
8 Melanesia, Micronesia, Polynesia; 9 Middle Africa; 10 Northern Africa;
11 Northern America; 12 Northern Europe; 13 South America;
14 South-Eastern Asia; 15 Southern Africa; 16 Southern Asia;
17 Southern Europe; 18 Western Africa; 19 Western Asia; 20 Western Europe.}

\item{vaporeg_regstart_reports, vaporeg_regend_reports, vaporeg_regduration_reports,
vaporeg_regchange_reports}{Spell start date, end date, duration in years,
and annual change indicator for the country reports classification. Change
equals 1 if the July 1 regime differs from the previous July 1.}

\item{vaporeg_regstart_detailed, vaporeg_regend_detailed, vaporeg_regduration_detailed,
vaporeg_regchange_detailed}{Spell start date, end date, duration in years,
and annual change indicator for the detailed classification.}

\item{vaporeg_regstart_compact, vaporeg_regend_compact, vaporeg_regduration_compact,
vaporeg_regchange_compact}{Spell start date, end date, duration in years,
and annual change indicator for the compact classification.}

\item{vaporeg_regduration_quadruple, vaporeg_regchange_quadruple}{Spell
duration in years and annual change indicator for the quadruple
classification.}

\item{vaporeg_regduration_triple, vaporeg_regchange_triple}{Spell duration
in years and annual change indicator for the triple classification.}

\item{vaporeg_regduration_binaut, vaporeg_regchange_binaut}{Spell duration
in years and annual change indicator for the binary autocracy
classification.}

\item{vaporeg_regduration_bindem, vaporeg_regchange_bindem}{Spell duration
in years and annual change indicator for the binary democracy
classification.}

\item{vaporeg_population}{Population series based on Gapminder and UN
data. The codebook notes that units coded as "Part of Other Country" are
excluded from this series.}

\item{vaporeg_population_sources}{Source flag for \code{vaporeg_population}.
Codes: 1 Modified based on Gapminder; 2 Modified based on UN;
3 UN Data; 4 Gapminder Data.}

}
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other institutions: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{vaporeg_2024}}

Other dichotomous democracy indexes: 
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{kailitz}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vaporeg_2024}},
\code{\link{wahman_teorell_hadenius}}

Other trichotomous democracy indexes: 
\code{\link{blm}},
\code{\link{kailitz}},
\code{\link{mainwaring}},
\code{\link{prc_gasiorowski}},
\code{\link{utip}},
\code{\link{vaporeg_2024}}
}
\concept{democracy}
\concept{dichotomous democracy indexes}
\concept{institutions}
\concept{trichotomous democracy indexes}
\keyword{datasets}
