% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vaporeg_2024}
\alias{vaporeg_2024}
\title{The Varieties of Political Regimes (VaPoReg) dataset, legacy 2024 schema}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 29840 rows and 30 columns.
}
\source{
S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Codebook. Version 1.7}. Tech. rep. Dresden: Hannah Arendt Institute for Totalitarianism Studies, 2024.  S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Dataset}. Dresden, 2024.
}
\usage{
vaporeg_2024
}
\description{
\code{vaporeg_2024} preserves the pre-0.7.0 \code{democracyData} interface for VaPoReg.
It is the last packaged version of the old schema, covering 1900 to 2024 and
including the package's standardized country identifiers and derived
democracy measures. Use \link{vaporeg} for the current March 2026 edition.
}
\section{Variables}{


\describe{

\item{vaporeg_country}{The name of the country or political unit for a given year.}

\item{year}{The calendar year.}

\item{vaporeg_code}{The VaPoReg unit identifier.}

\item{GeoNamesID}{Legacy geographic identifier for linking with spatial datasets.}

\item{vaporeg_cowcode}{The Correlates of War country code, where available.}

\item{vaporeg_cr}{Legacy compatibility alias of \code{vaporeg_s}.}

\item{vaporeg_s}{The standard classification of political regimes.}

\item{vaporeg_a}{An alternative classification that merges democracies and
semi-democracies and groups monarchies together as ruling monarchies.}

\item{vaporeg_governing_country}{For colonial and occupation regimes, the ruling country.}

\item{vaporeg_s_gc, vaporeg_a_gc}{Standard and alternative regime
classifications for the governing country.}

\item{vaporeg_s_start, vaporeg_s_end, vaporeg_s_duration, vaporeg_s_change}{Start date,
end date, duration, and annual change indicator for the standard
classification.}

\item{vaporeg_a_start, vaporeg_a_end, vaporeg_a_duration, vaporeg_a_change}{Start date,
end date, duration, and annual change indicator for the alternative
classification.}

\item{Gapminder_pop, gm_world_pop}{Population variables included in the legacy
interface.}

\item{un_region, un_continent}{UN region and continent codes.}

\item{extended_country_name, GWn, cown, in_GW_system}{Standardized country
identifiers added by \code{democracyData}.}

\item{vaporeg_binary_strict}{A dichotomous democracy indicator coded 1 only
for the most demanding VaPoReg democracy category.}

\item{vaporeg_binary_non_strict}{A dichotomous democracy indicator coded 1
for both full and defective democracies.}

\item{vaporeg_trichotomous}{A trichotomous democracy indicator coded 2 for
full democracy, 1 for defective democracy, and 0 otherwise.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other institutions: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{vaporeg}}

Other dichotomous democracy indexes: 
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{kailitz}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vaporeg}},
\code{\link{wahman_teorell_hadenius}}

Other trichotomous democracy indexes: 
\code{\link{blm}},
\code{\link{kailitz}},
\code{\link{mainwaring}},
\code{\link{prc_gasiorowski}},
\code{\link{utip}},
\code{\link{vaporeg}}
}
\concept{democracy}
\concept{dichotomous democracy indexes}
\concept{institutions}
\concept{trichotomous democracy indexes}
\keyword{datasets}
