% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wgi_legacy}
\alias{wgi_legacy}
\title{Democracy, Voice, and Accountability Index from the World Governance
Indicators, legacy series}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5350 rows and 13 columns.
}
\source{
D. Kaufmann and A. Kraay. \emph{Worldwide Governance Indicators}. The World Bank Research Group, 2020. \url{http://www.govindicators.org}.
}
\usage{
wgi_legacy
}
\description{
\code{wgi_legacy} preserves the pre-0.7.0 \code{democracyData} version of the World
Governance Indicators "Voice and Accountability" index. It uses the
pre-2025 WGI methodology, covers 1996 to 2023, and is not directly
comparable to the revised \link{wgi} series introduced in 0.7.0.
}
\details{
The variable schema is identical to \link{wgi}; the descriptions below are
unchanged from the earlier packaged release.
}
\section{Variables}{


\describe{

\item{wb_country}{Country name (World Bank standard).}

\item{wb_code}{World bank 3 character country code.}

\item{year}{Year.}

\item{Estimate}{Estimate of governance (ranges from approximately -2.5
(weak) to 2.5 (strong) governance performance), with a mean of 0.}

\item{StdErr}{Standard error reflects variability around the point estimate
of governance.}

\item{NumSrc}{Number of data sources on which estimate is based. The
methodology document lists, among these, the EIU Democracy Index, Freedom
House, the World Economic Forum Global Competitiveness Report, the Gallup
World Poll, the Institutional Profiles Database, the Political Risk
Services International Country Risk Guide, and Reporters Without Borders
Press Freedom Index. See \url{https://www.worldbank.org/content/dam/sites/govindicators/doc/va.pdf} for more
details.}

\item{Rank}{Percentile rank among all countries (ranges from 0 (lowest) to
100 (highest) rank).}

\item{Lower}{Lower bound of 90\% confidence interval for governance, in
percentile rank terms.}

\item{Upper}{Upper bound of 90\% confidence interval for governance, in
percentile rank terms.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}}
}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
