% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{anckar}
\alias{anckar}
\title{The Anckar-Fredriksson dataset of political regimes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18777 rows and 13 columns.
}
\source{
C. Anckar and C. Fredriksson. "Classifying Political Regimes 1800–2016: A Typology and a New Dataset". In: \emph{European Political Science} 18.1 (2018), pp. 84-96. DOI: \href{https://dx.doi.org/10.1057/s41304-018-0149-8}{10.1057/s41304-018-0149-8}. Version 2 available at \url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/7SSSAH/DIZXSI&version=2.0}.
}
\usage{
anckar
}
\description{
Dataset described in C. Anckar and C. Fredriksson. "Classifying Political Regimes 1800–2016: A Typology and a New Dataset". In: \emph{European Political Science} 18.1 (2018), pp. 84-96. DOI: \href{https://dx.doi.org/10.1057/s41304-018-0149-8}{10.1057/s41304-018-0149-8}. Data and codebook at the
link. This is version 2 of the data, available at \url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/7SSSAH/DIZXSI&version=2.0}.
}
\section{Variables}{


\describe{

\item{anckar_country}{The country name, as in the original dataset, with
minimal modification. Use \code{extended_country_name} instead if you want
a consistent name.}

\item{anckar_ccode}{The codebook says this is the COW code; it looks like a
modified version of the Polity code (with the USSR = 364, for example). Use
\code{cown} or \code{GWn} instead.}

\item{abbreviation}{World Bank country abbreviation.}

\item{year}{The calendar year.}

\item{democracy}{This should be identical to version 2 of Boix, Miller and
Rosato's Dichotomous democracy measure (\code{democracy_omitteddata}; see
\link{bmr}). Supplemented for all countries for the period
2011-2016, as well as for Liechtenstein 1866-1990, Monaco 1862-2016, and
San Marino 1800-1992, by Anckar and Fredriksson. 1 = democracy, 0 =
non-democracy.}

\item{monarchy}{Distinguishes between monarchies and republics in
democratic countries. It is \code{NA} for non-democracies. Vales: 0
Republic, 1 Monarchy, 7 Hybrid.}

\item{regimebroadcat}{Regime type. Classification with broad categories:
Parliamentarism, Semi-presidentialism, Presidentialism, Semi-monarchy,
Party-based rule, Personalist rule, Military rule, Absolute monarchy,
Oligarchy}

\item{regimenarrowcat}{Regime type. Classification with narrow categories:
Parliamentarism, Semi-presidentialism, Presidentialism, Semi-monarchy,
Single-party rule, Multi-party authoritarian rule, Personalist rule,
Military rule, Absolute monarchy, Monarchic oligarchy, Other oligarchy.}

\item{popelection}{Indicates if the head of state is popularly elected or
not in democratic republics. Values 0 Head of state not popularly elected;
1 Head of state popularly elected. In a few cases this variable is not
\code{NA} for non-democracies; it is unclear why.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found via the
Wayback Machine at
\url{https://web.archive.org/web/20130627160240/http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{https://correlatesofwar.org/data-sets/state-system-membership/} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}},
\code{\link{wgi_legacy}}

Other authoritarianism: 
\code{\link{REIGN}},
\code{\link{gwf_all}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}

Other dichotomous democracy indexes: 
\code{\link{anrr}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{kailitz}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vaporeg}},
\code{\link{vaporeg_2024}},
\code{\link{wahman_teorell_hadenius}}
}
\concept{authoritarianism}
\concept{democracy}
\concept{dichotomous democracy indexes}
\keyword{datasets}
