% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{UTF-8}
\name{bibliography}
\alias{bibliography}
\title{References to all the datasets included in this package.}
\format{
An object of class \code{BibEntry} (inherits from \code{bibentry}) of length 58.
}
\usage{
bibliography
}
\description{
This is a \code{\link[RefManageR:BibEntry]{RefManageR::BibEntry()}} object that can be manipulated using the
\link[RefManageR:RefManageR-package]{RefManageR::RefManageR} package functions or the standard citation functions
in R.
}
\keyword{datasets}
