% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bmr}
\alias{bmr}
\title{Boix-Miller-Rosato dichotomous coding of democracy, 1800-2010, version 2.0}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17632 rows and 14 columns.}
\source{
Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A
Complete Data Set of Political Regimes, 1800-2007. Comparative Political
Studies 46 (12): 1523-1554. Available at
\url{https://sites.google.com/site/mkmtwo/data}
}
\usage{
bmr
}
\description{
A dataset containing the Boix-Miller-Rosato (BMR) dichotomous coding of democracy,
used in the paper Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A
Complete Data Set of Political Regimes, 1800-2007. Comparative Political
Studies 46 (12): 1523-1554. The complete dataset can be downloaded here:
\url{https://sites.google.com/site/mkmtwo/BMR-PolRegimes.pdf?attredirects=0}.
The complete codebook is available here:
\url{https://sites.google.com/site/mkmtwo/data}.
}
\details{
\describe{
\item{bmr_country}{The country name in BMR.}
\item{bmr_ccode}{Numeric polity code code in BMR. }
\item{abbreviation}{World Bank 3-letter code.}
\item{year}{Year}
\item{democracy}{Dichotomous democracy measure. 1 = Democracy.}
\item{sovereign}{Dichotomous indicator of sovereignty/independence (if 0, democracy is NA).}
\item{democracy_trans}{-1 if democratic breakdown, 0 if no change, 1 if democratic transition.}
\item{democracy_breakdowns}{Previous number of democratic breakdowns.}
\item{democracy_duration}{Consecutive years of current regime type.}
\item{democracy_omitteddata}{This is the same measure as democracy, except
it records an NA for countries occupied during an international war (e.g.,
the Netherlands 1940-44) or experiencing state collapse during a civil war
(e.g., Lebanon 1976-89). The democracy variable instead fills in these
years as continuations of the same regime type.}
}
}
\section{Standard descriptive variables (generated by this package)}{

 \describe{

 \item{extended_country_name}{The name of the country in the Gleditsch-Ward
 system of states, or the official name of the
 entity (for non-sovereign entities and states not in the Gleditsch and Ward
 system of states) or else a common name for disputed cases that do not have
 an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
 scheme sometimes indicates the common name of the country and (in
 parentheses) the name of an earlier incarnation of the state: thus, they
 have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
 For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
 System Membership: A Revised List of the Independent States since 1816."
 International Interactions 25: 393-413. The list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
 Ward list of independent states. For details, see Gleditsch, Kristian S. &
 Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the
 Independent States since 1816." International Interactions 25: 393-413. The
 list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{cown}{The Correlates of War numeric country code, 2016 version. This
 differs from Gleditsch and Ward's numeric country code in a few cases. See
 \url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
 the full list.}

 \item{in_GW_system}{Whether the state is "in system" (that is, is
 independent and sovereign), according to Gleditsch and Ward, for this
 particular date. Matches at the end of the year; so, for example South
 Vietnam 1975 is `FALSE` because, according to Gleditsch and Ward, the
 country ended on April 1975 (being absorbed by North Vietnam). It is also
 `TRUE` for dates beyond 2012 for countries that did not end by then, depsite
 the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{arat_pmm}}, \code{\link{blm}},
  \code{\link{bnr}}, \code{\link{bollen_pmm}},
  \code{\link{doorenspleet}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{fh_pmm}}, \code{\link{gwf_all}},
  \code{\link{hadenius_pmm}}, \code{\link{kailitz}},
  \code{\link{magaloni}}, \code{\link{mainwaring}},
  \code{\link{munck_pmm}}, \code{\link{pacl}},
  \code{\link{peps}}, \code{\link{pitf}},
  \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{prc_gasiorowski}},
  \code{\link{svolik_regime}}, \code{\link{uds_2014}},
  \code{\link{ulfelder}}, \code{\link{utip}},
  \code{\link{vanhanen}},
  \code{\link{wahman_teorell_hadenius}}

Other events data: \code{\link{bnr}},
  \code{\link{prc_gasiorowski}}, \code{\link{ulfelder}}

Other dichotomous democracy indexes: \code{\link{bnr}},
  \code{\link{doorenspleet}}, \code{\link{kailitz}},
  \code{\link{pacl}}, \code{\link{svolik_regime}},
  \code{\link{ulfelder}}, \code{\link{utip}},
  \code{\link{wahman_teorell_hadenius}}
}
\keyword{datasets}
