% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bnr}
\alias{bnr}
\alias{bnr_extended}
\title{The Bernhard, Nordstrom & Reenock Event History Coding of Democratic
Breakdowns}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3651 rows and 8 columns.}
\source{
Michael Bernhard, Timothy Nordstrom, and Christopher
Reenock, "Economic Performance, Institutional Intermediation and Democratic
Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding
description available at
\url{http://users.clas.ufl.edu/bernhard/content/data/data.htm}
}
\usage{
bnr

bnr_extended
}
\description{
The Bernhard, Nordstrom & Reenock Event History Coding of Democratic
Breakdowns. From Michael Bernhard, Timothy Nordstrom, and Christopher
Reenock, "Economic Performance, Institutional Intermediation and Democratic
Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding
description available at
\url{http://users.clas.ufl.edu/bernhard/content/data/data.htm}
}
\section{Description}{


Taken from the BNR dataset description on the website:

(Time Series:  1913-2005, n: 3654, N: 126, Mean number of countries per
year: 39.28 , Mean number of years per country: 29.46).

The \code{event} variable is a binary coding of all democracies from 1913 until
2005 prepared for use in event history analysis.  Countries that meet the
minimum conditions for democracy (see below) enter the dataset and are
coded "0."  When countries cease to meet those minimum criteria they are
coded "1" and exit from the dataset.  The time frame onset in 1913 is a
function of when the first country (Norway) meets the minimum conditions.
All series terminate in either in a breakdown in various years or right
censorship in 2005.

The minimal conditions are based on Dahl's notion of polyarchy
(competitiveness, inclusiveness) combined with Linz and Stepan's stateness
criteria.

Competitiveness: We include countries that have competitive elections
without gross distortions of fairness.  We excluded cases from our set of
democracies in which we detected outcome changing vote fraud, in which
there was either extensive or extreme violence that inhibited voters from
expressing themselves, or in which political parties representing a
substantial portion of the population were banned.

Inclusivness:  We only include competitive polities in which at least fifty
percent of all adult citizens are enfranchised to vote in our set of
democracies.

Stateness:  We also paid careful attention to questions of sovereignty, not
including in our set of democracies many cases, where despite competitive
elections, full sovereignty was not formally achieved.  We also excluded in
which internal wars in which twenty percent of the population or territory
was out of control of the state were in progress.

\describe{

\item{bnr_country}{The original country name in the dataset.}

\item{bnr_cowcode}{The original COW country code in the dataset.}

\item{year}{Calendar year}.

\item{event}{The event coding. 0 = democracy, 1 = termination.}

\item{bnr}{Only in the \code{bnr_extended} dataset. 0 = non-democracy, 1 =
democracy. Years with a termination event are coded 0. Note that this is
the reverse of the \code{event} variable, where 0 indicates democracy; this is
because the \code{bnr} dataset was originally created for survival analysis.}

}
}

\section{Standard descriptive variables (generated by this package)}{

 \describe{

 \item{extended_country_name}{The name of the country in the Gleditsch-Ward
 system of states, or the official name of the
 entity (for non-sovereign entities and states not in the Gleditsch and Ward
 system of states) or else a common name for disputed cases that do not have
 an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
 scheme sometimes indicates the common name of the country and (in
 parentheses) the name of an earlier incarnation of the state: thus, they
 have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
 For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
 System Membership: A Revised List of the Independent States since 1816."
 International Interactions 25: 393-413. The list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
 Ward list of independent states.}

 \item{cown}{The Correlates of War numeric country code, 2016 version. This
 differs from Gleditsch and Ward's numeric country code in a few cases. See
 \url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
 the full list.}

 \item{in_GW_system}{Whether the state is "in system" (that is, is
 independent and sovereign), according to Gleditsch and Ward, for this
 particular date. Matches at the end of the year; so, for example South
 Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
 country ended on April 1975 (being absorbed by North Vietnam). It is also
 \code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
 the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note about the "extended" version of the dataset}{


The \code{bnr_extended} version of this data is the same as the \code{bnr} data, but
it uses the COW system of states to impute a value of "non-democracy" to
years where the state was in the state system but does not appear in the
original \code{bnr} dataset. For example, Albania first appears in the \code{bnr}
dataset in 1992 as a democracy; but it was in the state system since much
earlier. Since the \code{bnr} dataset claims to record all democratization
events in the COW system of states since 1913, we assume Albania was
non-democratic between 1913 and 1992. This procedure generates a large
number of additional country-years with information about democracy.
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{arat_pmm}}, \code{\link{blm}},
  \code{\link{bmr}}, \code{\link{bollen_pmm}},
  \code{\link{doorenspleet}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{eiu}}, \code{\link{fh_pmm}},
  \code{\link{gwf_all}}, \code{\link{hadenius_pmm}},
  \code{\link{kailitz}}, \code{\link{magaloni}},
  \code{\link{mainwaring}}, \code{\link{munck_pmm}},
  \code{\link{pacl}}, \code{\link{peps}},
  \code{\link{pitf}}, \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{prc_gasiorowski}},
  \code{\link{svolik_regime}}, \code{\link{uds_2014}},
  \code{\link{ulfelder}}, \code{\link{utip}},
  \code{\link{vanhanen}},
  \code{\link{wahman_teorell_hadenius}}

Other events data: \code{\link{bmr}},
  \code{\link{prc_gasiorowski}}, \code{\link{ulfelder}}

Other dichotomous democracy indexes: \code{\link{bmr}},
  \code{\link{doorenspleet}}, \code{\link{kailitz}},
  \code{\link{pacl}}, \code{\link{svolik_regime}},
  \code{\link{ulfelder}}, \code{\link{utip}},
  \code{\link{wahman_teorell_hadenius}}
}
\keyword{datasets}
