% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\name{cutpoints}
\alias{cutpoints}
\title{Extract cutpoints from a UD model in a tidy format.}
\usage{
cutpoints(model, type = "score")
}
\arguments{
\item{model}{A \link[mirt:SingleGroupClass-class]{mirt::SingleGroupClass} model
of the democracy scores.}

\item{type}{A string specifying the cutpoint type. Can be (an abbreviation
of) "score" (for score cutpoints) or "discrimination" (for discrimination
parameters). Default is "score."}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with either score cutpoints for each variable used to
construct the latent scores in terms of the latent variable (the default), or
discrimination parameters for each variable used to construct the index.
For the score cutpoints (\code{type = 'score'}), the columns
\code{estimate}, \code{pct975}, and \code{pct025} report the IRT
parametrization of the model estimates, a normalized measure in the same
scale as the latent variable.
}
\description{
This function takes a model of the democracy scores and extracts the
discrimination parameters, score cutpoints, and standard errors for all the
variables involved, putting these into a tidy data frame.
}
\examples{
\donttest{
# Replicate the official UDS 2011 release and calculate its cutpoints
library(dplyr)
library(mirt)
democracy_data <- generate_democracy_scores_dataset(
  selection = "_pmm", output_format = "wide", exclude_pmm_duplicates = FALSE
  )

democracy_data <- prepare_democracy_data(democracy_data)

replication_2011_model <- mirt(democracy_data |> 
  select(matches("pmm")), model = 1, itemtype = "graded", SE = TRUE, verbose = FALSE)

cutpoints(replication_2011_model)
}
}
