% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{doorenspleet}
\alias{doorenspleet}
\title{Renske Doorenspleet's Democracy Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13877 rows and 10 columns.}
\source{
Doorenspleet, Renske. 2000. Reassessing the Three Waves of
Democratization. World Politics 52 (03): 384-406. DOI:
10.1017/S0043887100016580.
\url{http://dx.doi.org/10.1017/S0043887100016580}.
}
\usage{
doorenspleet
}
\description{
A dataset containing the country-year version of the democracy classification
described in Doorenspleet, Renske. 2000. Reassessing the Three Waves of
Democratization. World Politics 52 (03): 384-406. DOI:
10.1017/S0043887100016580. \url{http://dx.doi.org/10.1017/S0043887100016580}.
}
\details{
\describe{
\item{doorenspleet_country}{Country name as in Doorenspleet.}
\item{doorenspleet_regime}{Regime type: A = Authoritarian, D = Democracy, I = interruption.}
\item{doorenspleet}{Regime type: 1 = Authoritarian, 2 = Democracy, interruption codes as NA.}
\item{year}{Year. Regimes are coded as of the 31 Dec of the year.}
\item{start_year}{First year of the regime.}
\item{end_year}{Last year of the regime.}
}
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{anckar}}, \code{\link{arat_pmm}},
  \code{\link{blm}}, \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{bollen_pmm}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{eiu}}, \code{\link{fh_pmm}},
  \code{\link{gwf_all}}, \code{\link{hadenius_pmm}},
  \code{\link{kailitz}}, \code{\link{magaloni}},
  \code{\link{mainwaring}}, \code{\link{munck_pmm}},
  \code{\link{pacl}}, \code{\link{peps}},
  \code{\link{pitf}}, \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{prc_gasiorowski}},
  \code{\link{svmdi}}, \code{\link{svolik_regime}},
  \code{\link{uds_2014}}, \code{\link{ulfelder}},
  \code{\link{utip}}, \code{\link{vanhanen}},
  \code{\link{wahman_teorell_hadenius}}

Other dichotomous democracy indexes: \code{\link{anckar}},
  \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{kailitz}}, \code{\link{pacl}},
  \code{\link{svolik_regime}}, \code{\link{ulfelder}},
  \code{\link{utip}}, \code{\link{wahman_teorell_hadenius}}
}
\concept{democracy}
\concept{dichotomous democracy indexes}
\keyword{datasets}
