% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{eiu}
\alias{eiu}
\title{The Economist Intelligence Unit's Democracy Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2839 rows and 7 columns.
}
\source{
The Economist Intelligence Unit. \emph{Democracy Index 2024: What's Wrong with Representative Democracy?} Tech. rep. London, UK: The Economist Intelligence Unit Limited, 2025..

T. E. I. Unit. \emph{Democracy Index 2023: Age of Conflict}. Tech. rep. London, United Kingdom: Economist Intelligence Unit, 2024..
}
\usage{
eiu
}
\description{
A dataset containing the country-year version of the democracy index compiled
by the Economist Intelligence Unit.
}
\details{
\describe{

\item{eiu_country}{Country name as in the original tables.}

\item{eiu}{Democracy index, ranging from 0 (least democratic) to to 10 (most
democratic). The report says that the index "is based on five categories:
electoral process and pluralism; civil liberties; the functioning of
government; political participation; and political culture" which form "one
interrelated whole".}

}
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}}
}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
