% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fh_pmm}
\alias{fh_pmm}
\alias{fh}
\title{Freedom House "Freedom in the World" data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6438 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9965 rows and 11 columns.
}
\source{
D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: \href{https://dx.doi.org/10.7910/DVN/WWYOHU}{10.7910/DVN/WWYOHU}.

F. House. \emph{Freedom in the World 2025: The Uphill Battle to Safeguard Rights}. Tech. rep. Freedom House, Feb. 2025.
}
\usage{
fh_pmm

fh
}
\description{
\code{fh_pmm} archives the Freedom House data included in D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: \href{https://dx.doi.org/10.7910/DVN/WWYOHU}{10.7910/DVN/WWYOHU}.,
while \code{fh} archives the 2025 edition of the Freedom
in the World data (updated to 2024). For the latest version of the Freedom House data, always
use \link{download_fh}.
}
\details{
The Freedom House dataset released in 2025 is different from the archived
PMM data for 49 country years:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{fh |>
  dplyr::inner_join(fh_pmm) |>
  dplyr::select(extended_country_name, year, fh_total_reversed, pmm_freedomhouse) |>
  dplyr::mutate(fh_total_reversed = (fh_total_reversed/2)+ 1) |>
  dplyr::filter(fh_total_reversed != pmm_freedomhouse) |>
  print(n = 50)
#> Joining with `by = join_by(year, extended_country_name, GWn, cown,
#> in_GW_system)`
#> # A tibble: 49 x 4
#>    extended_country_name                 year fh_total_reversed pmm_freedomhouse
#>    <chr>                                <dbl>             <dbl>            <dbl>
#>  1 Antigua & Barbuda                     1982               6                5.5
#>  2 Argentina                             1982               2.5              5  
#>  3 Bahamas                               1982               6.5              6  
#>  4 Bangladesh                            1982               3                2.5
#>  5 Bolivia                               1982               2.5              5.5
#>  6 Brazil                                1982               4.5              5  
#>  7 Burundi                               1982               1.5              2  
#>  8 Comoros                               1982               3.5              4  
#>  9 Congo                                 1993               5                4  
#> 10 Congo, Democratic Republic of (Zair~  1977               2                1.5
#> 11 Congo, Democratic Republic of (Zair~  1978               2                1.5
#> 12 Congo, Democratic Republic of (Zair~  1982               2                1.5
#> 13 Cote D'Ivoire                         1980               3                2.5
#> 14 Cote D'Ivoire                         1999               2.5              3  
#> 15 Swaziland                             1978               3                2.5
#> 16 Grenada                               1982               2.5              2  
#> 17 Guinea-Bissau                         1982               2                1.5
#> 18 Guyana                                1982               3.5              3  
#> 19 Honduras                              1982               5.5              5  
#> 20 India                                 1976               4                4.5
#> 21 Kiribati                              1982               6                6.5
#> 22 Liberia                               1982               2                3  
#> 23 Liberia                               2001               2                2.5
#> 24 Malta                                 1982               5.5              5  
#> 25 Mauritius                             1982               5.5              6  
#> 26 Mozambique                            1982               1                1.5
#> 27 Myanmar (Burma)                       1982               1.5              1  
#> 28 Panama                                1982               3                3.5
#> 29 Peru                                  2000               4.5              5  
#> 30 Philippines                           1982               3.5              3  
#> 31 Poland                                1982               2                2.5
#> 32 Portugal                              1982               6                6.5
#> 33 Sao Tome and Principe                 1982               2                1  
#> 34 Saudi Arabia                          1982               2                1.5
#> 35 Korea, Republic of                    1982               3                2.5
#> 36 Spain                                 1982               5.5              6.5
#> 37 Sri Lanka (Ceylon)                    1982               5.5              4.5
#> 38 Surinam                               1982               2                1.5
#> 39 Surinam                               1990               3                4.5
#> 40 Syria                                 1982               2                1.5
#> 41 Taiwan                                1978               3                3.5
#> 42 Taiwan                                1980               3                2.5
#> 43 Taiwan                                1991               5                3  
#> 44 Tajikistan                            1991               3                5  
#> 45 Turkey (Ottoman Empire)               1982               3                3.5
#> 46 Uganda                                1982               3                3.5
#> 47 Uruguay                               1982               3                3.5
#> 48 Vanuatu                               1982               6                5  
#> 49 Zimbabwe (Rhodesia)                   1982               4                3.5
}\if{html}{\out{</div>}}

Many, but not all of these cases seem to be concentrated around the problematic year of 1982.
}
\section{Variables}{


\describe{

\item{pmm_country}{The original country name in the PMM replication data.}

\item{year}{The calendar year. This is approximate. The surveys cover
specific periods in the original data that do not always overlap with a
single year. In particular, the year 1981 is "skipped" -- a single survey
covers Jan.1981 - Aug. 1982 and its value is assigned to 1982 here.}

\item{pmm_freedomhouse }{A reversed measure of political rights and civil liberties,
from 1 to 7, with higher values being more democratic (the opposite of the
actual FH data).}

\item{fh_total_reversed }{A reversed measure of political rights and civil liberties,
from 0 to 12, with higher values being more democratic (the opposite of the
actual FH data). To convert to the same scale as \code{pmm_freedomhouse}, divide by two and add 1,
or average \code{pr} and \code{cl} and then reverse.}

\item{pr}{The political rights scores (from 1 to 7) in the archived Freedom House data (\code{fh}).}

\item{cl}{The civil rights scores (from 1 to 7) in the archived Freedom House data (\code{fh}).}

\item{status}{The status of the country in the archived Freedom House data (Free, Partly Free, Not Free).}

\item{fh_total}{The sum of the political rights and civil liberties scores. Ranges from 2 to 14.}
}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
\link{download_fh} and \link{download_fh_electoral}

Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other Freedom House: 
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bollen_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{polity_pmm}},
\code{\link{prc_gasiorowski}}

Other ordinal democracy indexes: 
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{polityIV}},
\code{\link{polity_pmm}}
}
\concept{Freedom House}
\concept{PMM replication data}
\concept{democracy}
\concept{ordinal democracy indexes}
\keyword{datasets}
