% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fh_pmm}
\alias{fh_pmm}
\alias{fh}
\title{Freedom House "Freedom in the World" data from Pemstein, Meserve, and
Melton's replication dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6438 rows and 7 columns.
}
\source{
Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013.
"Replication data for: Democratic Compromise: A Latent Variable Analysis of
Ten Measures of Regime Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}.
}
\usage{
fh_pmm
}
\description{
This is the Freedom House data included in Pemstein, Daniel, Stephen A.
Meserve, and James Melton. 2013. "Replication data for: Democratic
Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In:
Harvard Dataverse. \url{http://hdl.handle.net/1902.1/PMM}. It differs for
some country-years from the current version of the Freedom House data, which
can be accessed using \link{download_fh}; hence its inclusion here.
}
\section{Variables}{


\describe{

\item{pmm_country}{The original country name in the PMM replication data.}

\item{year}{The calendar year. This is approximate. The surveys cover
specific periods in the original data that do not always overlap with a
single year. In particular, the year 1981 is "skipped" -- a single survey
covers Jan.1981 - Aug. 1982 and its value is assigned to 1982 here.}

\item{fh_pmm}{A reversed measure of political rights and civil liberties,
from 1 to 7, with higher values being more democratic (the oppositie of the
actual FH data).} }
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
\link{download_fh} and \link{download_fh_electoral}

Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_reign}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{wahman_teorell_hadenius}}

Other Freedom House: 
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bollen_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{polity_pmm}},
\code{\link{prc_gasiorowski}}

Other ordinal democracy indexes: 
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{polityIV}},
\code{\link{polity_pmm}}
}
\concept{Freedom House}
\concept{PMM replication data}
\concept{democracy}
\concept{ordinal democracy indexes}
\keyword{datasets}
