% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_democracy_scores_dataset.R
\name{generate_democracy_scores_dataset}
\alias{generate_democracy_scores_dataset}
\title{Generate a data frame containing all the democracy measures accessible from
this package}
\usage{
generate_democracy_scores_dataset(
  datasets,
  selection,
  output_format = "long",
  use_extended = TRUE,
  verbose = TRUE,
  force_redownload = FALSE,
  scale_scores = FALSE,
  uds_release_year = c(2014, 2011, 2010),
  svmdi_release_year = c(2020, 2016),
  exclude_pmm_duplicates = FALSE,
  keep_dissolution_year = FALSE,
  keep_cow_extras = FALSE,
  keep_PIPE_extras = FALSE,
  keep_polity_extras = FALSE,
  keep_only_last_in_year = TRUE
)
}
\arguments{
\item{datasets}{Character vector indicating which datasets to use in
producing the combined data frame. Can be any or all of (an unambiguous
abbreviation of) "\link{anckar}", "\link{anrr}", "\link{LIED}", "\link{PIPE}", "\link{arat_pmm}",
"\link{blm}", "\link{blm_pmm}", "\link{bmr}", "\link{bnr}", "\link{bnr_extended}", "\link{bti}",
"\link{bollen_pmm}", "\link{doorenspleet}", "\link{eiu}", "\link{fh_pmm}", "\link{gwf_all}",
"\link{gwf_all_extended}", "\link{hadenius_pmm}", "\link{kailitz}", "\link{magaloni}",
"\link{magaloni_extended}", "\link{mainwaring}", "\link{mainwaring_pmm}", "\link{munck_pmm}",
"\link{pacl}", "\link{pacl_pmm}", "\link{pacl_update}", "\link{peps}", "\link{pitf}",
"\link{polity_pmm}", "\link{polyarchy}", "\link{polyarchy_dimensions}", "\link{polyarchy_pmm}",
"\link{prc_gasiorowski}", "\link{prc_pmm}", "\link{svmdi}", "\link{svolik_regime}", "\link{uds}",
"\link{ulfelder}", "\link{ulfelder_extended}", "\link{utip}", "\link{vanhanen}", "\link{vdem}",
"\link{wahman_teorell_hadenius}", "\link{reign}" or "\link[=reign]{REIGN}", "\link{polityIV}",
"\link{polity5}", "\link{fh}", "\link{fh_electoral}", "\link{wgi}". It can also be empty; the
default is all of them.}

\item{selection}{A regular expression for selecting among the datasets.
Optional.}

\item{output_format}{Character. Whether to output a "wide" (each measure of
democracy in a separate column) or a "long" (a column with measure names, a
column with values) version of the data frame. Default is "long".}

\item{use_extended}{Whether to use "extended" (that is, including values
before 1945 for some regimes) versions of some datasets (\link{gwf}, \link{ulfelder},
\link{bnr}, and \link{magaloni}). Default is \code{TRUE}.}

\item{verbose}{Provides a running commentary on what the function is is
doing. Default is \code{TRUE}.}

\item{force_redownload}{Whether to re-download all datasets that can be
re-downloaded, including those archived with this package. Used only for
debugging; default is \code{FALSE}.}

\item{scale_scores}{Whether to scale each measure (substracting their mean
and dividing by their standard deviation). Default is \code{FALSE}.}

\item{uds_release_year}{Which release of the original \link{uds} dataset to
include. Defaults to the latest, 2014. Can be 2014, 2011, or 2010, or all
of them.}

\item{svmdi_release_year}{Which release of the \link{svmdi} dataset to include.
Defaults to the latest, 2020. Can be 2020 or 2016, or both of them.}

\item{keep_only_last_in_year}{Whether to keep only the last regime
measurement in a given country-year. Some datasets (e.g., \link{prc}, \link{reign})
contain more than one regime measurement per country-year in some cases (if
the regime changed multiple times during the year); setting this to
\code{TRUE} discards all except the regime measurement as of 31 December of
the year, the standard practice in most datasets. Default is \code{TRUE}.
This setting is only of interest if you set \code{output_format = "long"},
since it is ignored when \code{output_format = "wide"}, which automatically
discards all regime measurements except the last in the year.}

\item{target_system}{Character vector describing which state system to use
for the combined file, and which country codes to use. Can be one or both
of "GWn" (Gleditsch and Ward, numeric codes) and "cown" (numeric codes for
the Correlates of War system). The default is \emph{both} Gleditsch-ward and
Correlates of War (\code{c("GWn", "cown")}).}

\item{exclude_extra_pmm}{Whether to exclude versions of some measures found
in Pemstein, Meserve, and Melton's replication dataset for their 2010 piece
introducing the Unified Democracy Scores (Pemstein, Meserve, and Melton
2010, 2013). See \link{blm_pmm}, \link{prc_pmm}, \link{fh_pmm}, \link{pacl_pmm},
\link{vanhanen_pmm}, and \link{polity_pmm} for details. This is included mostly to
extend or replicate the \link{uds} scores. Default is \code{TRUE}.}
}
\value{
A \link{tibble} with the selected democracy measures and state system
information, in two versions: "long" and "wide". These contain the
following variables:
}
\description{
This function automates the process of joining together all the archived and
downloadable datasets in this package, producing a comprehensive data frame
with every democracy measure. It automatically converts each democracy
measure to a numeric value, ensures that larger values are associated with
more democracy, and that all country-years are appropriately matched.
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Long version}{


In the "long" version of the dataset (\code{format = "long"}), the output
data frame also contains the following variables:

\describe{

\item{year}{The calendar year. Most measures of democracy reflect the
country's situation as of 31 December of the year. If
\code{keep_last_in_year = FALSE} (and \code{format = "long"}), a single
country year may nevertheless contain more than one measurement for some
measures (e.g., \link{prc}).}

\item{measure}{The name of the measure. (e.g., "blm", "fh_total_reversed").
}

\item{variable}{The numerical value of the measure, in the original scale
(if \code{scale_scores = FALSE}) or as a z-score (if \code{scale_scores =
  TRUE}).}

\item{index_type}{The index type (dichotomous, trichotomous,
ordinal/graded, continuous).}

\item{dataset}{The name of the dataset.} }
}

\section{Wide version}{


In the "wide" version of the dataset (\code{format = "wide"}), the output
data frame can also contain any of the following variables (in the scales
described below, unless \code{scale_scores = TRUE}, in which case the
measures are converted to z-scores):

\describe{

\item{year}{The calendar year. In the "wide" version, all  measures of
democracy reflect the country's situation as of 31 December of the year as
much as possible.}

\item{anckar_democracy}{The \link{anckar} measure of democracy, as a numeric
value. Up to 2010 this should be identical to \code{bmr_democracy_omitteddata}.
0 = non-democracy, 1 = democracy.}

\item{arat}{The original \link{arat} measure of democracy, converted to numeric
form. Ranges from 29 (least democratic) to 109 (most democratic).
See \link{arat} for details.}

\item{anrr_democracy}{The \link{anrr} measure of democracy, as a numeric value. 0 =
non-democracy, 1 = democracy.}

\item{blm}{The \link{blm} measure of democracy, as a numeric value. Can be 0
(authoritarian), 0.5 (hybrid), or 1 (democracy). }

\item{bmr_democracy}{The \link{bmr} measure of democracy, as a numeric value.
Can be 0 (non-democracy) or 1 (democracy). }

\item{bmr_democracy_omitteddata}{According to the \link{bmr} codebook, "this is
the same measure as bmr, except it records an NA for countries occupied
during an international war (e.g., the Netherlands 1940-44) or experiencing
state collapse during a civil war (e.g., Lebanon 1976-89). The democracy
variable instead fills in these years as continuations of the same regime
type." There are some -1 values that I've converted to \code{NA} (the measure is
supposed to be between 0 and 1).}

\item{bmr_democracy_femalesuffrage}{According to the \link{bmr} codebook, this
is the same measure as \code{bmr}, except that it also requires that at least
half of adult women have the right to vote. 30 countries change values.}

\item{bnr}{The \link{bnr} event measure of democracy, reversed, so that 0
indicates non-democracy and 1 indicates democracy. Since the \code{event}
variable of the \link{bnr} dataset only codes terminations of democracy
(ignoring years when the country is non-democratic), this variable is
mostly equal to 1.}

\item{bnr_extended}{The \link{bnr_extended} variable, with 0 indicating
non-democracy and 1 indicating democracy. See the documentation for
\link{bnr_extended} for details of how this variable extends \link{bnr}, filling in
years of non-democracy back to 1913.}

\item{bti}{The \link{bti} - Berteslmann Transformation Democracy Status index.
Ranges from 1 to 10. }

\item{doorenspleet}{The \link{doorenspleet} measure of democracy, with 1
indicating non-democracy and 2 indicating democracy.}

\item{eiu}{The \link{eiu} measure of democracy, ranging from 0 (least
democratic) to 10 (most democratic). The report says that the index "is
based on five categories: electoral process and pluralism; civil liberties;
the functioning of government; political participation; and political
culture" which form "one interrelated whole"}

\item{fh_electoral}{The \link{fh_electoral} measure of electoral democracy, with
0 indicating a lack of electoral democracy, 1 indicating electoral
democracy status.}

\item{fh_total_reversed}{The \href{fh}{Freedom House} combined measure of
political and civil rights, reversed, so that 0 is least free and 12 is
most free.}

\item{gwf_democracy}{A measure of democracy from \link{gwf_all}, obtained by
coding all democracies as 1 and all non-democracies, including all
non-democratic nonautocracies, as 0.}

\item{gwf_democracy_extended}{A measure of democracy from
\link{gwf_all_extended}, obtained by coding all democracies as 1 and all
non-democracies, including all non-democratic nonautocracies, as 0. See the
documentation for \link{gwf_all_extended} for details on how the dataset was
extended to the period before 1945 for some countries.}

\item{gwf_democracy_strict}{A measure of democracy from \link{gwf_all}, obtained
by coding all democracies as 1 and all non-democracies as 0.  It is
\code{NA} for all non-democratic non-autocracies (e.g., warlord regimes,
foreign occupation, etc.).}

\item{gwf_democracy_extended_strict}{A measure of democracy from
\link{gwf_all_extended}, obtained by coding all democracies as 1 and all
non-democracies as 0. It is \code{NA} for all non-democratic
non-autocracies (e.g., warlord regimes, foreign occupation, etc.). See the
documentation for \link{gwf_all_extended} for details on how the dataset was
extended to the period before 1945 for some countries.}

\item{kailitz_binary}{A measure of democracy from \link{kailitz}, obtained by
coding all liberal democracies as 1 and all other regimes as 0.}

\item{kailitz_tri}{A trichotomous measure of democracy from \link{kailitz},
obtained by coding all liberal democracies as 2, all electoral autocracies
as 1, and all other regimes as 0.}

\item{lexical_index}{The \link{LIED} measure of electoral democracy, ranging
from 0 to 6.}

\item{lexical_index_plus}{The \link{LIED} measure of polyarchy, ranging from 0
to 7 (including political liberties).}

\item{magaloni_democracy}{A measure of democracy from \link{magaloni}, obtained
by coding all democracies as 1 and all non-democracies as 0.}

\item{magaloni_democracy_extended}{A measure of democracy from
\link{magaloni_extended}, obtained by coding all democracies as 1 and all
non-democracies as 0. See the documentation for \link{magaloni_extended} for
details on how the dataset was extended to the period before 1950 for some
countries.}

\item{mainwaring}{The measure of democracy from \link{mainwaring}, where 0 is
non-democracy, 0.5 represents hybrid regimes, and 1 is democracy.}

\item{pacl}{The measure of democracy from \link{pacl}, where 0 is non-democracy,
and 1 is democracy.}

\item{pacl_update}{The measure of democracy from \link{pacl_update}, where 0 is
non-democracy, and 1 is democracy.}

\item{PEPS1i, PEPS1q, PEPS1v, PEPS2i, PEPS2q, PEPS2v}{The measures of
democracy from \link{peps}. Higher values indicate more democracy.}

\item{PIPE_democracy, PIPE_regime}{The measures of democracy from \link{PIPE}.
Higher values indicate more democracy. See the documentation of \link{PIPE} for
large caveats - these measures are calculated on the basis of unclear
instructions and may contain errors.}

\item{pitf, pitf_binary}{The measures of democracy from \link{pitf}, converted
to numberic values. Higher values indicate more democracy.}

\item{pmm_arat}{The measure of democracy in \link{arat_pmm}.}

\item{pmm_blm}{The measure of democracy in \link{blm_pmm}.}

\item{pmm_bollen}{The measure of democracy in \link{bollen_pmm}.}

\item{pmm_fh}{The measure of democracy in \link{fh_pmm}. Check the documentation
of \link{fh_pmm} for caveats.}

\item{pmm_hadenius}{The measure of democracy in \link{hadenius_pmm}.}

\item{pmm_mainwaring}{The measure of democracy in \link{mainwaring_pmm}.}

\item{pmm_munck}{The measure of democracy in \link{munck_pmm}.}

\item{pmm_pacl}{The measure of democracy in \link{pacl_pmm}.}

\item{pmm_polity}{The measure of democracy in \link{polity_pmm}. Check the
documentation of \link{polity_pmm} for caveats.}

\item{pmm_polyarchy}{The measure of democracy in \link{polyarchy_pmm}.}

\item{pmm_prc}{The measure of democracy in \link{prc_pmm}. Check the
documentation of \link{prc_pmm} for caveats.}

\item{pmm_vanhanen}{The measure of democracy in \link{vanhanen_pmm}. Check the
documentation of \link{vanhanen_pmm} for caveats.}

\item{polity}{The polity measure of democracy in \link{polity} (version 5, with
NAs for -88, -77, -66).}

\item{polity2}{The polity2 measure of democracy in \link{polity} (version 5,
interpolated for periods of interregnum, occupation, and the like - see
documentation in the polity manual).}

\item{polityIV}{The polity measure of democracy in \link{polity} (version IV,
with NAs for -88, -77, -66).}

\item{polity2IV}{The polity2 measure of democracy in \link{polity} (version IV,
interpolated for periods of interregnum, occupation, and the like - see
documentation in the polity manual).}

\item{polyarchy_contestation_dimension}{The contestation dimension
(\code{CONTEST}) in \link{polyarchy_dimensions}.}

\item{polyarchy_inclusion_dimension}{The inclusion dimension
(\code{INCLUS}) in \link{polyarchy_dimensions}.}

\item{polyarchy_original_contestation}{The contestation dimension
(\code{cont}) in \link{polyarchy}.}

\item{polyarchy_original_polyarchy}{The original polyarchy scale
(\code{poly}) in \link{polyarchy}, reversed so that higher values imply more
democracy. The codebook suggests this was superceded by
\code{polyarchy_original_contestation}.}

\item{prc}{The measure of democracy in \link{prc}, where 1 is non-democracy, 3
are hybrid regimes, and 4 are democracies. Transitional regimes (2 in the
original scale) are coded NA.}

\item{reign_democracy}{A measure of democracy from \link{reign}, obtained by
coding all presidential and parliamentary democracies as 1, all other
regimes as 0.}

\item{csvmdi}{The continuous Support Vector Machine democracy index from
\link{svmdi}, 2020 version.}

\item{svmdi_2016}{The continuous Support Vector Machine democracy index
from \link{svmdi}, 2016 version.}

\item{dsvmdi}{The dichotomous Support Vector Machine democracy index from
\link{svmdi}, 2020 version.}

\item{svolik_democracy}{A measure of democracy from \link{svolik_regime}. 0 =
non-democracy, 1 = democracy. }

\item{uds_2010_mean, uds_2010_median, uds_2011_mean, uds_2011_median,
uds_2014_mean, uds_2014_median}{The mean and median posterior scores of the
\link{uds} index (2010, 2011, and 2014 releases). Higher values are more
democratic.}

\item{ulfelder_democracy}{The dichotmous measure of democracy in
\link{ulfelder}.}

\item{ulfelder_democracy_extended}{The dichotomous measure of democracy in
\link{ulfelder_extended}.}

\item{utip_dichotomous}{A dichotomous measure of democracy from \link{utip}.
Equals 1 if the regime is a social democracy, conservative democracy, or
one party democracy, 0 otherwise.}

\item{utip_dichotomous_strict}{A dichotomous measure of democracy from
\link{utip}. Equals 1 if the regime is a social democracy or conservative
democracy, 0 otherwise.}

\item{utip_trichotomous}{A trichotomous measure of democracy from \link{utip}.
Equals 2 if the regime is a social democracy or conservative democracy, 1
if it is a one-party democracy, 0 otherwise.}

\item{vanhanen_competition}{The competition index from \link{vanhanen}.}

\item{vanhanen_democratization}{The democratization index from \link{vanhanen}.}

\item{vanhanen_participation}{The participation index from \link{vanhanen}.}

\item{vaporeg_binary_strict}{A dichotomous measure of democracy from the
\link{vaporeg} dataset. Coded as 1 if the regime is classified as a full
democracy (VaPoReg_s = 10), and 0 otherwise. This strict definition excludes
semi-democracies and other intermediate forms. See \link{vaporeg} for details.}

\item{vaporeg_binary_non_strict}{A less restrictive dichotomous measure of
democracy from \link{vaporeg}, coded as 1 if the regime is a full democracy
(VaPoReg_s = 10) or a semi-democracy (VaPoReg_s = 20), and 0 otherwise. See
\link{vaporeg} for details.}

\item{vaporeg_trichotomous}{A trichotomous democracy measure from
\link{vaporeg}, based on the VaPoReg_s classification. Coded as 2 for full
democracies (10), 1 for semi-democracies (20), and 0 for all other regime
types. This measure distinguishes between fully and partially democratic
regimes. See \link{vaporeg} for details.}

\item{v2x_api}{The additive polyarchy index from \link{vdem}.}

\item{v2x_delibdem}{The deliberative democracy index from \link{vdem}.}

\item{v2x_egaldem}{The egalitarian democracy index from \link{vdem}.}

\item{v2x_libdem}{The liberal democracy index from \link{vdem}.}

\item{v2x_mpi}{The multiplicative polyarchy index from \link{vdem}.}

\item{v2x_partipdem}{The participative democracy index from \link{vdem}.}

\item{v2x_polyarchy}{The polyarchy (electoral democracy) index from
\link{vdem}.}

\item{wgi_democracy}{The voice and accountability index from \link{wgi}.}

\item{wth_democ1}{A dichotomous measure of democracy from
\link{wahman_teorell_hadenius}, obtaining by coding 1 all democracies according
to the \code{regime1ny} variable, 0 all other regimes.}

\item{wth_democrobust}{A dichotomous measure of democracy from
\link{wahman_teorell_hadenius}, obtaining by coding 1 all democracies according
to the \code{regimenyrobust} variable, 0 all other regimes.}

}
}

\section{State-system choices}{


State system choices are not always straightforward or clear. Most democracy
datasets use some version of the Correlates of War system of states, though
not consistently or entirely correctly. Gleditsch and Ward developed their own system V-Dem has developed its own system
of states, which is not always consistent with the Gleditsch and Ward system.

I have made some state-system choices to fit (mostly) with the Gleditsch
and Ward system of states. The overarching rule is: no democracy score
appears for a state that ceased to exist at the close of the calendar year.
With few exceptions, I thus trim and rewrite country–year observations so
that every surviving row corresponds to a political unit recognized by the
Gleditsch–Ward (GW) state–system on 31 December of the calendar year,
unless the country is not in the GW panel (in which case I make some ad-hoc
decisions, as for the British Mandate of Palestine and Sardinia, which end
in 1948 and 1861, respectively).

The main exception is that I do not delete country–year observations for
Vietnam before 1954 or Korea before 1947 (unless these are separated into
North and South, as in the \link{PIPE} dataset), since these states have extant
successors, and several datasets (including \link{LIED} and \link{vdem_simple}
measure them for those periods). Thus, the following countries should have
no observations in the year of their dissolution (and years after), unless
the dissolution occurred on 31 December, in which case the last observation
is for the year of dissolution (e.g., 1992 for Czechoslovakia):\tabular{lllr}{
   extended_country_name \tab panel \tab dissolution_date \tab last_observation \cr
   Austria-Hungary \tab GWn \tab 1918-11-13 \tab 1917 \cr
   Baden \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Bavaria \tab GWn \tab 1871-01-17 \tab 1870 \cr
   British Mandate of Palestine \tab none \tab 1948-05-14 \tab 1947 \cr
   Czechoslovakia \tab GWn \tab 1992-12-31 \tab 1992 \cr
   German Democratic Republic \tab GWn \tab 1990-10-02 \tab 1989 \cr
   Germany (Prussia) \tab GWn \tab 1945-05-07 \tab 1944 \cr
   Hanover \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Hesse-Kassel (Electoral) \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Hesse-Darmstadt (Ducal) \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Mecklenburg-Schwerin \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Modena \tab GWn \tab 1861-03-16 \tab 1860 \cr
   Papal States \tab GWn \tab 1870-09-22 \tab 1869 \cr
   Parma \tab GWn \tab 1861-03-16 \tab 1860 \cr
   Vietnam, Republic of \tab GWn \tab 1975-04-30 \tab 1974 \cr
   Sardinia \tab polity \tab 1861-03-16 \tab 1860 \cr
   Saxony \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Tuscany \tab GWn \tab 1861-03-16 \tab 1860 \cr
   Sicily (Two Sicilies) \tab GWn \tab 1861-03-16 \tab 1860 \cr
   Wurttemberg \tab GWn \tab 1871-01-17 \tab 1870 \cr
   Yemen, People's Republic of \tab GWn \tab 1990-05-21 \tab 1989 \cr
   Yugoslavia \tab GWn \tab 2006-06-04 \tab 2005 \cr
   Zanzibar \tab GWn \tab 1964-04-26 \tab 1963 \cr
   Great Colombia \tab GWn \tab 1830-09-22 \tab 1829 \cr
   Orange Free State \tab GWn \tab 1910-05-30 \tab 1909 \cr
   Tibet \tab GWn \tab 1950-10-01 \tab 1949 \cr
   Transvaal \tab GWn \tab 1910-05-30 \tab 1909 \cr
   United Provinces of Central America \tab GWn \tab 1839-12-31 \tab 1839 \cr
}


The total number of rows affected is small. If you want to keep the
observations for the year of dissolution (e.g., measures for the GDR in 1990),
set \code{keep_dissolution_year = TRUE}.

As noted, Korea and Vietnam do contain observations for years after
their exit from the system of states in the Gleditsch and Ward panel in
1910 and 1893, respectively, since they have extant successors and are
measured by some datasets for those periods:\tabular{lll}{
   extended_country_name \tab panel \tab dissolution_date \cr
   Korea \tab GWn \tab 1910-08-22 \cr
   Vietnam (Annam/Cochin China/Tonkin) \tab GWn \tab 1893-01-01 \cr
}


The treatment of Palestine is also complicated. Several datasets (e.g.,
\link{vdem}, \link{fh}, and \link{LIED}) contain separate measurements for the West Bank
and Gaza; \link{fh} includes the West Bank, Gaza, and "Israeli-Occupied
Territories" (I'm assuming the occupied territories in Lebanon and Syria).
I've labeled them as follows:\tabular{lllrr}{
   extended_country_name \tab original_country_name \tab dataset \tab min_year \tab max_year \cr
   British Mandate of Palestine \tab Palestine \tab vaporeg \tab 1919 \tab 1947 \cr
   British Mandate of Palestine \tab Palestine/British Mandate \tab LIED \tab 1918 \tab 1947 \cr
   British Mandate of Palestine \tab Palestine/British Mandate \tab vdem \tab 1918 \tab 1947 \cr
   Israeli-Occupied Territories \tab Israeli-Occupied Territories \tab fh \tab 1996 \tab 2009 \cr
   Palestine, State of \tab Palestine \tab eiu \tab 2006 \tab 2024 \cr
   Palestine, State of \tab Palestine \tab vaporeg \tab 1948 \tab 2006 \cr
   Palestine, State of \tab Palestinian Authority-Administered Territories \tab fh \tab 1996 \tab 2009 \cr
   Palestine, State of \tab West Bank and Gaza \tab wgi_democracy \tab 1996 \tab 2023 \cr
   Palestine, State of \tab West Bank and Gaza Strip \tab fh \tab 1977 \tab 1995 \cr
   Palestine, State of/Gaza Strip Only \tab Gaza Strip \tab fh \tab 2010 \tab 2024 \cr
   Palestine, State of/Gaza Strip Only \tab Palestine Gaza Strip \tab vaporeg \tab 2007 \tab 2024 \cr
   Palestine, State of/Gaza Strip Only \tab Palestine/Gaza \tab LIED \tab 1948 \tab 2023 \cr
   Palestine, State of/Gaza Strip Only \tab Palestine/Gaza \tab vdem \tab 1948 \tab 2024 \cr
   Palestine, State of/West Bank Only \tab Palestine West Bank \tab vaporeg \tab 2007 \tab 2024 \cr
   Palestine, State of/West Bank Only \tab Palestine/West Bank \tab LIED \tab 1948 \tab 2023 \cr
   Palestine, State of/West Bank Only \tab Palestine/West Bank \tab vdem \tab 1948 \tab 2024 \cr
   Palestine, State of/West Bank Only \tab West Bank \tab fh \tab 2010 \tab 2024 \cr
}


There are some other oddities, primarily concerning \link{fh} and the \link{polity} and
Polity-derived datasets such as \link{pitf}. \link{fh} includes measures for many
non-sovereign territories, including a measure for the "Russian-occupied
territories of Ukraine"; I've left it as is (\verb{extended_country_name = }
\code{"Russian-occupied territories of Ukraine"}, not allocated to either Ukraine
or Russia). And \link{polity} includes separate
observations for both Russia 1922 and USSR 1922; I've excluded the USSR
1922 observation since it officially began on 31 December 1922. I similarly
exclude the observation for the USSR in 1991, since it was dissolved on 26
December 1991. Polity also contains two observations for Ethiopia in 1993,
one for before (\code{polity_ccode = 530}) and one for after  the splitting off of
Eritrea (\code{polity_ccode = 529}). I've kept the one after the split. You can keep
both USSR and Ethiopia observations by setting the flag
\code{keep_polity_extras = TRUE}.

Some datasets based on the Correlates of War system include an observation for
both the German Federal Republic in 1990 (COW code 260) and re-unified Germany
(COW code 255). I've only kept the observation for re-unified Germany, but you can
keep both by setting the flag \code{keep_cow_extras = TRUE}. Note also that though
I've excluded Germany 1945 (COW code 255, per the table above, since it ends
in 1945 in the Gleditsch-Ward panel), there are observations for the (occupied)
territory of both the later German Federal Republic and the later German Democratic
Republic in 1945. Setting \code{keep_cow_extras = TRUE} will keep these observations.

Finally, the \link{PIPE} dataset makes a number of unusual choices for the state
system, including country-year observations for North and South Vietnam in
periods before their existence as separate states, and distinct
Austria-Hungary (Austria) and Austria-Hungary (Hungary) observations. I've
deleted most of these:\tabular{lll}{
   Entity in PIPE \tab Years dropped \tab Reason \cr
   Austria‑Hungary (Austria) & (Hungary) \tab <= 1918 \tab Kept existing Austria-Hungary observations. \cr
   Czech Republic \tab <= 1992 \tab Czechoslovakia till 31 Dec 1992. \cr
   Vietnam North \tab <=  1870 \tab Anachronistic rows. \cr
   Vietnam (unified) \tab 1954–1975 \tab Split state period. \cr
   Turkey \tab < 1922 \tab Ottoman Empire until Oct 1923 - kept existing Ottoman Empire observations. \cr
   Pakistan \tab 1950–1970 \tab Pre‑independence East Pakistan period mis-label. \cr
   Serbia \tab 1990 \tab Still inside Yugoslavia. \cr
   Serbia and Montenegro \tab 1991 \tab State not yet created. \cr
}


I do not recommend that you keep these observations, but you can keep them by
setting the flag \code{keep_PIPE_extras = TRUE}.
}

\examples{
# You can select only some datasets

democracy_data_gwf <- generate_democracy_scores_dataset(datasets = c("gwf", "pacl", "bmr"),
         output_format = "wide")

# all PMM datasets
democracy_data_pmm <- generate_democracy_scores_dataset(selection = "pmm")

\dontrun{
# This produces scaled scores
generate_democracy_scores_dataset(keep_only_last_in_year = FALSE, scale_scores = TRUE)

# These require downloads:

democracy_data_long <- generate_democracy_scores_dataset(datasets = "pacl", force_redownload = TRUE)
democracy_data_wide <- generate_democracy_scores_dataset(datasets = "pacl", force_redownload = TRUE,
                                                         output_format = "wide")
}
}
\references{
D. Pemstein, S. Meserve, and J. Melton. "Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type". In: \emph{Political Analysis} 18.4 (2010), pp. 426-449. DOI: \href{https://dx.doi.org/10.1093/pan/mpq020}{10.1093/pan/mpq020}.

D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: \href{https://dx.doi.org/10.7910/DVN/WWYOHU}{10.7910/DVN/WWYOHU}.
}
