% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_democracy_scores_dataset.R
\name{generate_democracy_scores_dataset}
\alias{generate_democracy_scores_dataset}
\title{Generate a data frame containing all the democracy measures accessible from
this package}
\usage{
generate_democracy_scores_dataset(
  datasets,
  selection,
  output_format = "long",
  use_extended = TRUE,
  include_extra_pmm = FALSE,
  verbose = TRUE,
  target_system = c("GWn", "cown"),
  force_redownload = FALSE,
  scale_scores = FALSE,
  keep_only_last_in_year = TRUE,
  exclude_downloadable = FALSE
)
}
\arguments{
\item{datasets}{Character vector indicating which datasets to use in
producing the combined data frame. Can be any or all of (an unambiguous
abbreviation of) "\link{anckar}", "\link{anrr}", "\link{LIED}", "\link{PIPE}", "\link{arat_pmm}",
"\link{blm}", "\link{blm_pmm}", "\link{bmr}", "\link{bnr}", "\link{bnr_extended}", "\link{bti}",
"\link{bollen_pmm}", "\link{doorenspleet}", "\link{eiu}", "\link{fh_pmm}", "\link{gwf_all}",
"\link{gwf_all_extended}", "\link{hadenius_pmm}", "\link{kailitz}", "\link{magaloni}",
"\link{magaloni_extended}", "\link{mainwaring}", "\link{mainwaring_pmm}", "\link{munck_pmm}",
"\link{pacl}", "\link{pacl_pmm}", "\link{pacl_update}", "\link{peps}", "\link{pitf}",
"\link{polity_pmm}", "\link{polyarchy}", "\link{polyarchy_dimensions}", "\link{polyarchy_pmm}",
"\link{prc_gasiorowski}", "\link{prc_pmm}", "\link{svmdi}", "\link{svolik_regime}",
"\link{uds_2010}", "\link{uds_2011}", "\link{uds_2014}", "\link{ulfelder}",
"\link{ulfelder_extended}", "\link{utip}", "\link{vanhanen}", "\link{wahman_teorell_hadenius}",
"\link{reign}", "\link{polityIV}", "\link{polity}" (or
"\href{download_polity_annual.html}{polity_annual}"), "\href{download_fh.html}{fh}",
"\link{fh_electoral}", "\link{wgi}". Default is all of them.}

\item{output_format}{Character. Whether to output a "wide" (each measure of
democracy in a separate column) or a "long" (a column with measure names, a
column with values) version of the data frame. Default is "long".}

\item{use_extended}{Whether to use "extended" (that is, including values
before 1945 for some regimes) versions of some datasets (\link{gwf}, \link{ulfelder},
\link{bnr}, and \link{magaloni}). Default is \code{TRUE}.}

\item{include_extra_pmm}{Whether to include versions of some measures found
in Pemstein, Meserve, and Melton's replication dataset for their 2010 piece
introducing the Unified Democracy Scores (Pemstein, Meserve, and Melton
2010, 2013). See \link{blm_pmm}, \link{prc_pmm}, \link{fh_pmm}, \link{pacl_pmm},
\link{vanhanen_pmm}, and \link{polity_pmm} for details. This is included mostly for
use by the \href{https://github.com/xmarquez/QuickUDS}{QuickUDS} package.}

\item{verbose}{Provides a running commentary on what the function is is
doing. Default is \code{TRUE}.}

\item{target_system}{Character vector describing which state system to use
for the combined file, and which country codes to use. Can be one or both
of "GWn" (Gleditsch and Ward, numeric codes) and "cown" (numeric codes for
the Correlates of War system). The default is \emph{both} Gleditsch-ward and
Correlates of War (\code{c("GWn", "cown")}).}

\item{force_redownload}{Whether to re-download all datasets that can be
re-downloaded, including those archived with this package. Used only for
debugging; default is \code{FALSE}.}

\item{scale_scores}{Whether to scale each measure (substracting their mean
and dividing by their standard deviation). Default is \code{FALSE}.}

\item{keep_only_last_in_year}{Whether to keep only the last regime
measurement in a given country-year. Some datasets (e.g., \link{prc}, \link{reign})
contain more than one regime measurement per country-year in some cases (if
the regime changed multiple times during the year); setting this to
\code{TRUE} discards all except the regime measurement as of 31 December of
the year, the standard practice in most datasets. Default is \code{TRUE}.
This setting is only of interest if you set \code{output_format = "long"},
since it is ignored when \code{output_format = "wide"}, which automatically
discards all regime measurements except the last in the year.}

\item{exclude_downloadable}{Whether to exclude all datasets that must be
explicitly downloaded (\link{polity}, \link{fh}, \link{reign}, \link{wgi}), using only archived
datasets. This speeds up the process considerably, but you lose some of the
better democracy measures out there. Default is \code{FALSE}.}
}
\value{
A \link{tibble} with the selected democracy measures and state system
information, in two versions: "long" and "wide". These contain the
following variables:
}
\description{
This function automates the process of joining together all the archived and
downloadable datasets in this package, producing a comprehensive data frame
with every democracy measure. It automatically converts each democracy
measure to a numeric value, ensures that larger values are associated with
more democracy, and that all country-years are appropriately matched.
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Long version}{


In the "long" version of the dataset (\code{format = "long"}), the output
data frame also contains the following variables:
}

\section{Variables}{


\describe{

\item{year}{The calendar year. Most measures of democracy reflect the
country's situation as of 31 December of the year. If
\code{keep_last_in_year = FALSE} (and \code{format = "long"}), a single
country year may nevertheless contain more than one measurement for some
measures (e.g., \link{prc}).}

\item{measure}{The name of the measure. (e.g., "blm", "fh_total_reversed").
}

\item{variable}{The numerical value of the measure, in the original scale
(if \code{scale_scores = FALSE}) or as a z-score (if \code{scale_scores =
  TRUE}).}

\item{index_type}{The index type (dichotomous, trichotomous,
ordinal/graded, continuous).}

\item{dataset}{The name of the dataset.} }
}

\section{Wide version}{


In the "wide" version of the dataset (\code{format = "wide"}), the output
data frame can also contain any of the following variables (in the scales
described below, unless \code{scale_scores = TRUE}, in which case the
measures are converted to z-scores):

\describe{

\item{year}{The calendar year. In the "wide" version, all  measures of
democracy reflect the country's situation as of 31 December of the year as
much as possible.}

\item{anckar_democracy}{The \link{anckar} measure of democracy, as a numeric
value. Up to 2010 this should be identical to \code{bmr_democracy_omitteddata}.
0 = non-democracy, 1 = democracy.}

\item{anrr}{The \link{anrr} measure of democracy, as a numeric value. 0 =
non-democracy, 1 = democracy.}

\item{blm}{The \link{blm} measure of democracy, as a numeric value. Can be 0
(authoritarian), 0.5 (hybrid), or 1 (democracy). }

\item{bmr_democracy}{The \link{bmr} measure of democracy, as a numeric value.
Can be 0 (non-democracy) or 1 (democracy). }

\item{bmr_democracy_omitteddata}{According to the \link{bmr} codebook, "this is
the same measure as bmr, except it records an NA for countries occupied
during an international war (e.g., the Netherlands 1940-44) or experiencing
state collapse during a civil war (e.g., Lebanon 1976-89). The democracy
variable instead fills in these years as continuations of the same regime
type." There are some -1 values that I've converted to \code{NA} (the measure is
supposed to be between 0 and 1).}

\item{bmr_democracy_femalesuffrage}{According to the \link{bmr} codebook, this
is the same measure as \code{bmr}, except that it also requires that at least
half of adult women have the right to vote. 30 countries change values.}

\item{bnr}{The \link{bnr} event measure of democracy, reversed, so that 0
indicates non-democracy and 1 indicates democracy. Since the \code{event}
variable of the \link{bnr} dataset only codes terminations of democracy
(ignoring years when the country is non-democratic), this variable is
mostly equal to 1.}

\item{bnr_extended}{The \link{bnr_extended} variable, with 0 indicating
non-democracy and 1 indicating democracy. See the documentation for
\link{bnr_extended} for details of how this variable extends \link{bnr}, filling in
years of non-democracy back to 1913.}

\item{bti}{The \link{bti} - Berteslmann Transformation Democracy Status index.
Ranges from 1 to 10. }

\item{doorenspleet}{The \link{doorenspleet} measure of democracy, with 1
indicating non-democracy and 2 indicating democracy.}

\item{eiu}{The \link{eiu} measure of democracy, ranging from 0 (least
democratic) to 10 (most democratic). The report says that the index "is
based on five categories: electoral process and pluralism; civil liberties;
the functioning of government; political participation; and political
culture" which form "one interrelated whole"}

\item{fh_electoral}{The \link{fh_electoral} measure of electoral democracy, with
0 indicating a lack of electoral democracy, 1 indicating electoral
democracy status.}

\item{fh_total_reversed}{The \href{fh}{Freedom House} combined measure of
political and civil rights, reversed, so that 0 is least free and 12 is
most free.}

\item{gwf_democracy}{A measure of democracy from \link{gwf_all}, obtained by
coding all democracies as 1 and all non-democracies, including all
non-democratic nonautocracies, as 0.}

\item{gwf_democracy_extended}{A measure of democracy from
\link{gwf_all_extended}, obtained by coding all democracies as 1 and all
non-democracies, including all non-democratic nonautocracies, as 0. See the
documentation for \link{gwf_all_extended} for details on how the dataset was
extended to the period before 1945 for some countries.}

\item{gwf_democracy_strict}{A measure of democracy from \link{gwf_all}, obtained
by coding all democracies as 1 and all non-democracies as 0.  It is
\code{NA} for all non-democratic non-autocracies (e.g., warlord regimes,
foreign occupation, etc.).}

\item{gwf_democracy_extended_strict}{A measure of democracy from
\link{gwf_all_extended}, obtained by coding all democracies as 1 and all
non-democracies as 0. It is \code{NA} for all non-democratic
non-autocracies (e.g., warlord regimes, foreign occupation, etc.). See the
documentation for \link{gwf_all_extended} for details on how the dataset was
extended to the period before 1945 for some countries.}

\item{kailitz_binary}{A measure of democracy from \link{kailitz}, obtained by
coding all liberal democracies as 1 and all other regimes as 0.}

\item{kailitz_tri}{A trichotomous measure of democracy from \link{kailitz},
obtained by coding all liberal democracies as 2, all electoral autocracies
as 1, and all other regimes as 0.}

\item{lexical_index}{The \link{LIED} measure of electoral democracy, ranging
from 0 to 6.}

\item{magaloni_democracy}{A measure of democracy from \link{magaloni}, obtained
by coding all democracies as 1 and all non-democracies as 0.}

\item{magaloni_democracy_extended}{A measure of democracy from
\link{magaloni_extended}, obtained by coding all democracies as 1 and all
non-democracies as 0. See the documentation for \link{magaloni_extended} for
details on how the dataset was extended to the period before 1950 for some
countries.}

\item{mainwaring}{The measure of democracy from \link{mainwaring}, where 0 is
non-democracy, 0.5 represents hybrid regimes, and 1 is democracy.}

\item{pacl}{The measure of democracy from \link{pacl}, where 0 is non-democracy,
and 1 is democracy.}

\item{pacl_update}{The measure of democracy from \link{pacl_update}, where 0 is
non-democracy, and 1 is democracy.}

\item{PEPS1i, PEPS1q, PEPS1v, PEPS2i, PEPS2q, PEPS2v}{The measures of
democracy from \link{peps}. Higher values indicate more democracy.}

\item{PIPE_democracy, PIPE_regime}{The measures of democracy from \link{PIPE}.
Higher values indicate more democracy. See the documentation of \link{PIPE} for
large caveats - these measures are calculated on the basis of unclear
instructions and may contain errors.}

\item{pitf, pitf_binary}{The measures of democracy from \link{pitf}, converted
to numberic values. Higher values indicate more democracy.}

\item{pmm_arat}{The measure of democracy in \link{arat_pmm}.}

\item{pmm_blm}{The measure of democracy in \link{blm_pmm}.}

\item{pmm_bollen}{The measure of democracy in \link{bollen_pmm}.}

\item{pmm_fh}{The measure of democracy in \link{fh_pmm}. Check the documentation
of \link{fh_pmm} for caveats.}

\item{pmm_hadenius}{The measure of democracy in \link{hadenius_pmm}.}

\item{pmm_mainwaring}{The measure of democracy in \link{mainwaring_pmm}.}

\item{pmm_munck}{The measure of democracy in \link{munck_pmm}.}

\item{pmm_pacl}{The measure of democracy in \link{pacl_pmm}.}

\item{pmm_polity}{The measure of democracy in \link{polity_pmm}. Check the
documentation of \link{polity_pmm} for caveats.}

\item{pmm_polyarchy}{The measure of democracy in \link{polyarchy_pmm}.}

\item{pmm_prc}{The measure of democracy in \link{prc_pmm}. Check the
documentation of \link{prc_pmm} for caveats.}

\item{pmm_vanhanen}{The measure of democracy in \link{vanhanen_pmm}. Check the
documentation of \link{vanhanen_pmm} for caveats.}

\item{polity}{The polity measure of democracy in \link{polity} (version 5, with
NAs for -88, -77, -66).}

\item{polity2}{The polity2 measure of democracy in \link{polity} (version 5,
interpolated for periods of interregnum, occupation, and the like - see
documentation in the polity manual).}

\item{polityIV}{The polity measure of democracy in \link{polity} (version IV,
with NAs for -88, -77, -66).}

\item{polity2IV}{The polity2 measure of democracy in \link{polity} (version IV,
interpolated for periods of interregnum, occupation, and the like - see
documentation in the polity manual).}

\item{polyarchy_contestation_dimension}{The contestation dimension
(\code{CONTEST}) in \link{polyarchy_dimensions}.}

\item{polyarchy_inclusion_dimension}{The inclusion dimension
(\code{INCLUS}) in \link{polyarchy_dimensions}.}

\item{polyarchy_original_contestation}{The contestation dimension
(\code{cont}) in \link{polyarchy}.}

\item{polyarchy_original_polyarchy}{The original polyarchy scale
(\code{poly}) in \link{polyarchy}, reversed so that higher values imply more
democracy. The codebook suggests this was superceded by
\code{polyarchy_original_contestation}.}

\item{prc}{The measure of democracy in \link{prc}, where 1 is non-democracy, 3
are hybrid regimes, and 4 are democracies. Transitional regimes (2 in the
original scale) are coded NA.}

\item{reign_democracy}{A measure of democracy from \link{reign}, obtained by
coding all presidential and parliamentary democracies as 1, all other
regimes as 0.}

\item{csvmdi}{The continuous Support Vector Machine democracy index from
\link{svmdi}, 2020 version.}

\item{svmdi_2016}{The continuous Support Vector Machine democracy index
from \link{svmdi}, 2016 version.}

\item{dsvmdi}{The dichotomous Support Vector Machine democracy index from
\link{svmdi}, 2020 version.}

\item{svolik_democracy}{A measure of democracy from \link{svolik_regime}. 0 =
non-democracy, 1 = democracy. }

\item{uds_2010_mean, uds_2010_median, uds_2011_mean, uds_2011_median,
uds_2014_mean, uds_2014_median}{The mean and median posterior scores of the
\link{uds} index (2010, 2011, and 2014 releases). Higher values are more
democratic.}

\item{ulfelder_democracy}{The dichotmous measure of democracy in
\link{ulfelder}.}

\item{ulfelder_democracy_extended}{The dichotomous measure of democracy in
\link{ulfelder_extended}.}

\item{utip_dichotomous}{A dichotomous measure of democracy from \link{utip}.
Equals 1 if the regime is a social democracy, conservative democracy, or
one party democracy, 0 otherwise.}

\item{utip_dichotomous_strict}{A dichotomous measure of democracy from
\link{utip}. Equals 1 if the regime is a social democracy or conservative
democracy, 0 otherwise.}

\item{utip_trichotomous}{A trichotomous measure of democracy from \link{utip}.
Equals 2 if the regime is a social democracy or conservative democracy, 1
if it is a one-party democracy, 0 otherwise.}

\item{vanhanen_competition}{The competition index from \link{vanhanen}.}

\item{vanhanen_democratization}{The democratization index from \link{vanhanen}.}

\item{vanhanen_participation}{The participation index from \link{vanhanen}.}

\item{wgi_democracy}{The voice and accountability index from \link{wgi}.}

\item{wth_democ1}{A dichotomous measure of democracy from
\link{wahman_teorell_hadenius}, obtaining by coding 1 all democracies according
to the \code{regime1ny} variable, 0 all other regimes.}

\item{wth_democrobust}{A dichotomous measure of democracy from
\link{wahman_teorell_hadenius}, obtaining by coding 1 all democracies according
to the \code{regimenyrobust} variable, 0 all other regimes.}

}
}

\examples{
democracy_data_long_no_download <- generate_democracy_scores_dataset(exclude_downloadable = TRUE,
         keep_only_last_in_year = FALSE)

# You can select only some datasets

democracy_data_gwf <- generate_democracy_scores_dataset(datasets = c("gwf_all"),
         output_format = "wide")

# all PMM datasets
democracy_data_pmm <- generate_democracy_scores_dataset(selection = "pmm")

\dontrun{
# This produces scaled scores
generate_democracy_scores_dataset(exclude_downloadable = TRUE,
         keep_only_last_in_year = FALSE,
         scale_scores = TRUE)

# These require downloads:

democracy_data_long <- generate_democracy_scores_dataset()
democracy_data_wide <- generate_democracy_scores_dataset(output_format = "wide")
}
}
\references{
Pemstein, Daniel, Stephen Meserve, and James Melton. 2010.
Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime
Type. Political Analysis 18 (4): 426-449.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication
data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures
of Regime Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}
}
