% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kailitz}
\alias{kailitz}
\title{The Steffen Kailitz Dataset of Authoritarian Regime Types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9606 rows and 18 columns.
}
\source{
Kailitz, Steffen. 2013. Classifying political regimes revisited:
legitimation and durability. Democratization 20 (1): 39-60. Original data
available at \url{http://dx.doi.org/10.1080/13510347.2013.738861}.
}
\usage{
kailitz
}
\description{
Contains the authoritarian regime types dataset from Kailitz, Steffen. 2013.
Classifying political regimes revisited: legitimation and durability.
Democratization 20 (1): 39-60. Original data available at
\url{http://dx.doi.org/10.1080/13510347.2013.738861}.
}
\section{Variables}{


\describe{

\item{year}{The calendar year. Note that more than one regime type may
obtain in a single year fo a given country.}

\item{kailitz_country}{The country name.}

\item{combined_regime}{The "combined regime" for the country year. This is
just the concatenation of all regimes for that year if the country has more
than one regime type. For example, Spain has a "combined regime" of
"Military Autocracy-One party Autocracy-Personalist Autocracy" for the each
of the years 1946-1974.}

\item{kailitz_binary}{ Dichotomous
democracy indicator. 1 if \code{combined_regime} is "Liberal democracy", 0
otherwise.}

\item{kailitz_tri}{ Trichotomous
democracy indicator. 2 if \code{combined_regime} is "Liberal Democracy", 1
if it is "Electoral Autocracy" or "Electoral Autocracy-Liberal Democracy",
0 otherwise.}

\item{personal}{ Binary indicator of
personalism. \code{TRUE} if \code{combined_regime} contains "Personal
Autocracy", \code{FALSE} otherwise.}

\item{communist}{ Binary indicator
of communism. \code{TRUE} if \code{combined_regime} contains "Communist
Ideocracy", \code{FALSE} otherwise.}

\item{military}{ Binary indicator of
military autocracy. \code{TRUE} if \code{combined_regime} contains
"Military Autocracy", \code{FALSE} otherwise.}

\item{party}{ Binary indicator of
one party autocracy \code{TRUE} if \code{combined_regime} contains "One
party Autocracy", \code{FALSE} otherwise.}

\item{monarchy}{ Binary indicator of
monarchy. \code{TRUE} if \code{combined_regime} contains "Monarchy",
\code{FALSE} otherwise.}

\item{electoral}{ Binary indicator
of electoral autocracy. \code{TRUE} if \code{combined_regime} contains
"Electoral autocracy", \code{FALSE} otherwise.}

\item{failure}{ Binary indicator of
state failure or occupation. \code{TRUE} if \code{combined_regime} contains
"State Failure or Occupation", \code{FALSE} otherwise.}

\item{transition}{ Binary indicator
of transition. \code{TRUE} if \code{combined_regime} contains "Transition",
\code{FALSE} otherwise.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{anckar}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_reign}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{wahman_teorell_hadenius}}

Other authoritarianism: 
\code{\link{anckar}},
\code{\link{download_reign}()},
\code{\link{gwf_all}},
\code{\link{magaloni}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}

Other dichotomous democracy indexes: 
\code{\link{anckar}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{doorenspleet}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}

Other trichotomous democracy indexes: 
\code{\link{blm}},
\code{\link{mainwaring}},
\code{\link{prc_gasiorowski}},
\code{\link{utip}}
}
\concept{authoritarianism}
\concept{democracy}
\concept{dichotomous democracy indexes}
\concept{trichotomous democracy indexes}
\keyword{datasets}
