% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{magaloni}
\alias{magaloni}
\alias{magaloni_extended}
\title{Autocracies of the World, 1950-2012 (Version 1.0).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8629 rows and 28 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10291 rows and 10 columns.
}
\usage{
magaloni

magaloni_extended
}
\description{
The "Autocracies of the World" dataset by Beatriz Magaloni, Jonathan Chu, and
Eric Min.  Please cite as Magaloni, Beatriz, Jonathan Chu, and Eric Min.
2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford
University. Original data and codebook available at
\url{http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/}.
The documentation here is directly derived from this codebook.
}
\section{Overview}{


This dataset classifies the world's political regimes, primarily focusing
on distinctions among different types of autocracies, between 1950 and
2012. Over the last decade, academic work on political institutions has
placed more focus on differences within autocracies-a group of polities
that were long treated as a homogenous pool defined by "lack of democracy."
Advancing the literature, contributions such as Geddes (2003) have observed
and delineated differences in the governing institutions, power holders,
and nature of governance among various autocratic governments. Such studies
have given rise to a wealth of insightful research.

Attempts to study autocracies in a larger-scale, quantitative manner have
resulted in the creation of datasets that classify countries over time
using varying taxonomies. The field of extant datasets is relatively new
and still developing. (For examples, see Wahman et al. 2013; Geddes et
al. 2011; and Cheibub et al. 2010.) We seek to contribute to this
research agenda with this first edition of a new dataset.

Our dataset includes several features that we believe are quite useful.
These include:
\itemize{
\item Avoiding "missing," "transitional," and hybrid classifications. Many
countries have tumultuous years of instability, transition, and flux.
However, this does not eliminate the fact that a government of some sort
exists during these years. Substantial efforts were made to figure out the
institutions underlying transitional regimes. For example, a military
government can oversee a transition to democracy. Thus, rather than leave
certain observations blank, we create a separate variable for transitional
years to allow researcher the flexibility of deciding whether or not those
observations are relevant to their analysis. (The only exception is Somalia
from 1991 to 2006, where the polity seems truly "stateless.") Similarly, we
sought to eliminate hybrid regime classifications. There are high quality
datasets available for scholars who are interested analyzing hybrid
authoritarian regimes (such as Geddes et al. 2011). We have found,
however, that hybrid classifications often (1) hinder useful quantitative
analysis or (2) are actually essentially one type of regime, but only
appear hybrid due to certain window-dressing institutional features.
Finding the "essential" regime type seems to be a valuable contribution.
\item Providing two novel and objective measures of personalism. Currently
available datasets recognize that certain autocratic regimes are highly
"personalistic" and have created a personalist regime type. We agree that
personalism is an important feature of many autocracies, but we depart from
previous work in our belief that personalism is a quality distinct from the
regime's institutions and is a concept that should be operationalized as
its own variable. In other words, all autocracies have some degree of
personalism: a military regime, for example, could have a very
non-personalistic, corporate leadership or it can have a highly
personalistic general as the head of state. In addition to treating
personalism as a distinct concept, we have observed that measures of
personalism lack objective and clear criteria for measurement. Most
existing datasets rely on general consensus of the qualitative literature
to identify a personalistic regime. Given these two concerns, we provide
two measures of personalism (detailed below) in this dataset.
\item Classifying through 2012. In covering up to 2012, the dataset will
hopefully retain some longevity as other complementary datasets, many of
which end in the mid-2000s, are extended into more recent years. We also
hope to update this dataset periodically with more years and useful
covariates.
}
}

\section{Summary of Variables}{


Several sets of variables come in pairs: \code{variable_r} and
\code{variable_nr}. These correspond with rounded and non-rounded values.
The rationale and methodology behind these pairs of data is provided in the
section entitled "Transition Years and Rounding." (See online codebook).

Two datasets are provided. The original dataset (\code{magaloni}) and an
"extended" but simplified dataset (\code{magaloni_extended}) that extends
the duration of the first regime (\code{regime_nr} and \code{demo_nr}) for
each country backwards in time before 1950 (using the variable
\code{duration_nr}). The \code{magaloni_extended} dataset omits all the
indicators of personalism or head of government.

\describe{

\item{cyear}{Numeric expression of country-year, made by concatenating
ccode and year.}

\item{cntyr}{Alphanumeric expression of country-year, made by concatenating
scode and year.}

\item{magaloni_ccode}{Numeric country codes, based on the Correlates of War
dataset, as in the original dataset. Use \code{cown} instead.}

\item{scode}{Three-letter country codes, based on the Correlates of War
dataset.}

\item{magaloni_country}{String version of country, as in original dataset.
Use \code{country_name} instead.}

\item{year}{Four-digit calendar year. In the \code{magaloni_extended} the
year is extended to back before 1950 using \code{duration_nr}, which adds
1662 observations to \code{regime_nr} and \code{demo_nr}.}

\item{un_region}{Region, based on classifications by the United Nations, as
in the original dataset. Use \code{region} instead, which is based on the
same classifications but is more comprehensive - no missing values.}

\item{un_continent}{Continent, based on classifications by the United
Nations, as in the original dataset. Use \code{continent} instead, which is
based on the same classifications but is more comprehensive - no missing
values.}

\item{reg_id}{A unique number identifying a specific regime in a specific
country. The number is created by combining the country code (ccode) with a
basic regime count in the country in the following manner: 100*(ccode) +
running total. For example, the first country-year in the data for Haiti
(country code is 41) is 4101, and changes to 4102 the year that regime type
changes in the country. Note that this ID is created using the non-rounded
regime classifications.}

\item{demo_r}{A dummy variable for whether a given country-year has a
democratic regime, using the rounding rule. (Note that these
classifications do not necessarily align with a strict threshold based on
Polity scores.)}

\item{demo_nr}{A dummy variable for whether a given country-year has a
democratic regime, without rounding. (Note that these classifications do
not necessarily align with a strict threshold based on Polity scores.)}

\item{regime_r}{The regime type of a given country-year, using the rounding
rule so that the regime type that constituted the majority (or plurality)
of the year is used. The list of potential regimes includes:
\itemize{
\item Democracy
\item Multiparty
\item Single party
\item Military
\item Monarchy
}

Definitions of each are provided in the "Classifying Autocracies" section.
Note that these classifications are mutually exclusive for any given
country-year.}

\item{regime_nr}{The regime type of a given country-year, in which the
classification is based on the regime type that exists at the end of the
year. This is the more "conventional" approach of other regime datasets.}

\item{duration_r}{The age of a regime up to the given country-year, using
the rounding rule. The count begins at 1.}

\item{duration_nr}{The age of a regime up to the given country-year,
without rounding. The first year in which a new regime takes hold is
recorded as 1, regardless of when in the year this new regime appears.}

\item{personal1}{A three-point measure of the country-year's regime's
degree of personalism. This is based on the seven-point \code{xconst}
(executive constraints) scale of the Polity IV dataset. The following
conversion was used:

xconst 1 Highly personal 2

xconst 2-4 Moderately personal 1

xconst 5-7 Weakly/not personal 0

In contrast to \code{personal2}, whenever Polity IV does not provide an
\code{xconst} measure (periods of transition, interregnum, or occupation),
this measure is left blank.}

\item{personal2}{Identical to \code{personal1}, except missing values are
filled in. Whenever possible, missing values are imputed using the xconst
measures that do exist for a given regime. When a regime has the same
recorded \code{xconst} throughout, this value is used for missing values.
When xconst is not the same throughout a given regime, values are prorated
to fill middle years and based on the closest xconst measure to fill years
that are on either end of the regime's lifespan.}

\item{lindex}{A newly constructed measure of personalism within each
regime. (See the "Personalism" subsection for a substantive explanation of
this metric.) The variable is essentially a Herfendahl index (sum of
squared shares) using the column exname. For a given country-year in a
unique regime (see \code{reg_id}), the following calculation is made:
$sum_{i=1}^{m} (exec_i/n)^2$ where n is the age of the regime up to that
year, and exec is the number of years that a unique executive i (out of a
total m executives up to that year) has led the regime. As such, a regime
led by only one person up through that year yields a personalism index of
\enumerate{
\item A theoretical scenario where leadership changes every single year would
yield 1/n. These calculations are made using the non-rounded values. We
note that this is a relatively sensitive measure in the early/formative
years of an individual regime, but we propose this is a useful way of
considering personalism as an evolving attribute of a regime over time.
More discussion on the relative merits of this measure can be found below.}
}

\item{exname}{The name of the executive head of state in the country-year.
Country years from 1950 to 2008 are based on Cheibub, Gandhi, and Vreeland
(2010).}

\item{change}{A dummy for whether a regime change occurred in this
country-year. (Note that in cases such as coups and counter-coups, such
changes can occur without the overarching regime type changing.)}

\item{tdate}{The date a country underwent transition to a new regime type.
The format is MM.DD.YYYY, and this date is used in the rounding rule (see
the section entitled "Transitional Years and Rounding"). When a specific
month or date cannot be identified, "00" is used.}

\item{trans}{A dummy variable indicating whether the country was in the
midst of transition, as noted by the Polity IV dataset.}

\item{occup}{A dummy variable indicating whether the country was occupied
by a foreign power in the given year, as noted by the Polity IV dataset.}

\item{interreg}{A dummy variable indicating whether the country was in the
midst of interregnum in the given year, as noted by the Polity IV dataset.}
}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other authoritarianism: 
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{gwf_all}},
\code{\link{kailitz}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}
}
\concept{authoritarianism}
\concept{democracy}
\concept{dichotomous indexes of democracy}
\keyword{datasets}
