% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{peps}
\alias{peps}
\title{Participation-Enhanced Polity Score}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15089 rows and 28 columns.}
\source{
Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M.
Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and
Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of
Democracy" Studies in Comparative International Development 42, 2 (Summer,
2006).
}
\usage{
peps
}
\description{
A dataset containing the Participation-Enhanced Polity score described in
Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett,
Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong
(2006) "Voting Counts: Participation in the Measurement of Democracy" Studies
in Comparative International Development 42, 2 (Summer, 2006). The complete
dataset is available here:
\url{http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm}. The
documentation below is based on the iformation in that website.
}
\details{
\describe{

\item{peps_ccode}{COW code, as used by the PEPS dataset.}

\item{year}{Year.}

\item{polity1raw}{Raw Polity1 score with missing data
included, directly from Polity.}

\item{Polity1}{Clean Polity1, irregular
codes defined as missing data.}

\item{Polity2}{Clean Polity2, irregular codes
defined as missing data.}

\item{Polity3}{Clean Preferred Polity score.}

\item{Pdemocracy}{Clean Polity democracy, irregular codes defined as missing
data.}

\item{Pautocracy}{Clean Polity autocracy, irregular codes defined as
missing data.}

\item{FHfree}{Sum of Freedom House polrights & civliberties,
perfect score is 12.}

\item{toI}{IDEA Votes/Voting age population. n = 4966,
missing = 11437.}

\item{toV}{Vanhanen votes/two-thirds of Vanhanen
population. n = 14441, missing = 1962.}

\item{toQ}{(mostly)IDEA votes/Voting
age population, with participation coded zero for noncompetitive elections. n
=  8010, missing = 8393.}

\item{PEPS1i}{PEPS1, using toI. n = 9163, missing =
7240.}

\item{PEPS2i}{PEPS2, using toI. n = 4125, missing = 12278.}

\item{PEPS1q}{PEPS1, using toQ. n = 9439, missing = 6964.}

\item{PEPS2q}{PEPS2, using toQ. n = 6890, missing = 9513.}

\item{PEPS1v}{PEPS1, using toV. n = 14090, missing = 2313.}

\item{PEPS2v}{PEPS1, using toV. n = 13241, missing = 3162.}

\item{pop}{Population in thousands, Penn World Tables.}

\item{rgdpch}{Real
per capita GDP, chain method, Penn World Tables.}

\item{suffrage}{Legal
suffrage as percent of Voting age population, from Paxton and Bollen.}

\item{cv}{Compulsory voting. Coded as 4=strongly enforced,3=weakly
enforced,2=NA,1=not enforced,0=none. Source: IDEA.}

\item{peps_country}{The
country names found in the original data. these had to be fixed - lots of
them were blank.} }
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{anckar}}, \code{\link{arat_pmm}},
  \code{\link{blm}}, \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{bollen_pmm}}, \code{\link{doorenspleet}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{eiu}}, \code{\link{fh_pmm}},
  \code{\link{gwf_all}}, \code{\link{hadenius_pmm}},
  \code{\link{kailitz}}, \code{\link{magaloni}},
  \code{\link{mainwaring}}, \code{\link{munck_pmm}},
  \code{\link{pacl}}, \code{\link{pitf}},
  \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{prc_gasiorowski}},
  \code{\link{svmdi}}, \code{\link{svolik_regime}},
  \code{\link{uds_2014}}, \code{\link{ulfelder}},
  \code{\link{utip}}, \code{\link{vanhanen}},
  \code{\link{wahman_teorell_hadenius}}

Other continuous indexes of democracy: \code{\link{polyarchy_dimensions}}

Other polity: \code{\link{pitf}}
}
\concept{continuous indexes of democracy}
\concept{democracy}
\concept{polity}
\keyword{datasets}
