% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pitf}
\alias{pitf}
\alias{pitf_p4}
\title{PITF democracy indicator}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16574 rows and 13 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16581 rows and 13 columns.
}
\usage{
pitf

pitf_p4
}
\description{
The democracy indicator used by the political instability task force and
described in Goldstone, Jack, Robert Bates, David Epstein, Ted Gurr, Michael
Lustik, Monty Marshall, Jay Ulfelder, and Mark Woodward. 2010. A Global Model
for Forecasting Political Instability. American Journal of Political Science
54 (1): 190-208. DOI:10.1111/j.1540-5907.2009.00426.x and in Taylor, Sean J.
and Ulfelder, Jay, A Measurement Error Model of Dichotomous Democracy Status
(May 20, 2015). Available at SSRN: \url{http://ssrn.com/abstract=2726962} or
\url{http://dx.doi.org/10.2139/ssrn.2726962}. This is derived from the Polity
dataset (pitf uses version 5; pitf_p4 uses PolityIV updated to 2018).
}
\section{Variables}{


\describe{

\item{pitf_country}{The original country name in the Polity dataset.}

\item{pitf_ccode}{The \code{ccode} variable in the polity dataset. Should
be identical to \code{cown}.}

\item{year}{The calendar year.}

\item{exrec}{The polity "executive recruitment" variable. Can be: 1 =
Ascription

2 = Dual Executive (Designation)

3 = Designation

4 = Self-Selection

5 = Gradual Transition from Self-Selection Dual Executive (Election)

6 = Transitional or Restricted Election

7 = Competitive Election.

See the polity codebook for more details.}

\item{parcomp}{The polity "competitiveness of participation" variable. See the
polity codebook for more details. Here's what the codebook says:

The Competitiveness of Participation: The competitiveness of participation
refers to the extent to which alternative preferences for policy and
leadership can be pursued in the political arena. Political competition
implies a significant degree of civil interaction, so polities which are
coded Unregulated (1) on Regulation of Participation (PARREG, variable 2.5)
are not coded for competitiveness. Polities in transition between
Unregulated and any of the regulated forms on variable 2.5 also are not
coded on variable 2.6. Competitiveness is coded on a five category scale:

(0) Not Applicable: This is used for polities that are coded as
Unregulated, or moving to/from that position, in Regulation of Political
Participation (variable 2.6).

(1) Repressed: No significant oppositional activity is permitted outside
the ranks of the regime and ruling party. Totalitarian party systems,
authoritarian military dictatorships, and despotic monarchies are typically
coded here. However, the mere existence of these structures is not
sufficient for a Repressed coding. The regime's institutional structure
must also be matched by its demonstrated ability to repress oppositional
competition.

(2) Suppressed: Some organized, political competition occurs outside
government, without serious factionalism; but the regime systematically and
sharply limits its form, extent, or both in ways that exclude substantial
groups (20 percent or more of the adult population) from participation. Suppressed
competition is distinguished from Factional competition (below) by the
systematic, persisting nature of the restrictions: large classes of people,
groups, or types of peaceful political competition are continuously
excluded from the political process. As an operational rule, the banning of
a political party which received more than 10 percent of the vote in a recent
national election is sufficient evidence that competition is "suppressed."
However, other information is required to determine whether the appropriate
coding is (2) Suppressed or (3) Factional competition. This category is
also used to characterize transitions between Factional and Repressed
competition. Examples of "suppression" are:

i. Prohibiting some kinds of political organizations, either by type or
group of people involved (e.g., no national political parties or no ethnic
political organizations).

ii. Prohibiting some kinds of political action (e.g., Communist parties may
organize but are prohibited from competing in elections).

iii. Systematic harassment of political opposition (leaders killed, jailed,
or sent into exile; candidates regularly ruled off ballots; opposition
media banned, etc.). This is evidence for either Factional, Suppressed, or
Repressed, depending on the nature of the regime, the opposition, and the
persistence of political groups. Note 3.6: A newly enacted right to engage
in political activities is most likely a change from category 1 to 2.

(3) Factional: Polities with parochial or ethnic-based political factions
that regularly compete for political influence in order to promote
particularist agendas and favor group members to the detriment of common,
secular, or cross-cutting agendas.

(4) Transitional: Any transitional arrangement from Restricted, Suppressed,
or Factional patterns to fully Competitive patterns, or vice versa.
Transitional arrangements are accommodative of competing, parochial
interests but have not fully linked parochial with broader, general
interests. Sectarian and secular interest groups coexist.

(5) Competitive: There are relatively stable and enduring, secular
political groups which regularly compete for political influence at the
national level; ruling groups and coalitions regularly, voluntarily
transfer central power to competing groups. Competition among groups seldom
involves coercion or disruption. Small parties or political groups may be
restricted in the Competitive pattern.

By combining scores on Regulation of Political Participation (variable 3.5)
and the Competitiveness of Participation (variable 3.6), a relatively
detailed picture of the extent of political competition and opposition
emerges.}

\item{pitf}{A five category indicator of democracy described in Goldstone
et al 2010. Can be:

0-Full autocracy (exrec < 7, parcomp !=0 and parcomp < 3)

1-Partial autocracy (exrec < 7, parcomp = 0 or parcomp > 2)

2-Partial democracy with factionalism (exrec > 6, parcomp = 3)

3-Partial democracy (exrec > 6, parcomp = 0 or parcomp = 4 or parcomp = 5
but exrec != 8)

4-Full democracy (exrec = 8, parcomp = 5). See Goldstone et al. 2010 for
full details.}

\item{pitf_binary}{A simplification of the pitf indicator of democracy
described in Taylor, Sean J. and Ulfelder, Jay, A Measurement Error Model
of Dichotomous Democracy Status (May 20, 2015). Available at SSRN:
\url{http://ssrn.com/abstract=2726962} or
\url{http://dx.doi.org/10.2139/ssrn.2726962}. A country is a democracy (1)
"if its chief executive is chosen in competitive elections (EXREC equal to
7 or 8) and political competition is not suppressed (PARCOMP equal to 0 or
PARCOMP greater than 2)" Otherwise it is a non-democracy (0).} }
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\references{
Goldstone, Jack, Robert Bates, David Epstein, Ted Gurr, Michael Lustik, Monty
Marshall, Jay Ulfelder, and Mark Woodward. 2010. A Global Model for
Forecasting Political Instability. American Journal of Political Science 54
(1): 190-208. DOI:10.1111/j.1540-5907.2009.00426.x

Taylor, Sean J. and Ulfelder, Jay, A Measurement Error Model of Dichotomous
Democracy Status (May 20, 2015). Available at SSRN:
\url{http://ssrn.com/abstract=2726962} or
\url{http://dx.doi.org/10.2139/ssrn.2726962}
}
\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{anckar}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_reign}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{wahman_teorell_hadenius}}

Other polity: 
\code{\link{peps}}
}
\concept{democracy}
\concept{polity}
\keyword{datasets}
