% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{polity_pmm}
\alias{polity_pmm}
\title{Polity 2 from PMM}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8050 rows and 7 columns.
}
\source{
Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013.
"Replication data for: Democratic Compromise: A Latent Variable Analysis of
Ten Measures of Regime Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}.

Marshall, Monty G., Ted Robert Gurr, and Keith Jaggers. 2012. "Polity
IV: Political Regime Characteristics and Transitions, 1800-2012."
}
\usage{
polity_pmm
}
\description{
Polity 2 Index from Polity IV. 2013. Taken from Pemstein, Daniel, Stephen A.
Meserve, and James Melton. 2013. "Replication data for: Democratic
Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In:
Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}. Higher
values are more democratic. Use \link{download_polity_annual} to download the
latest version of the Polity IV data.
}
\section{Variables}{


\describe{

\item{pmm_country}{The original country name in the PMM replication data.}

\item{year}{The calendar year.}

\item{polity_pmm}{Annual polity2 index, interpolating values for
interruption, interregnum, and transition periods. Higher values are more
democratic. Taken from Pemstein, Daniel, Stephen A. Meserve, and James
Melton. 2013. "Replication data for: Democratic Compromise: A Latent
Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}. In
this version the indicator only goes to 1945, and it differs from polity2
in a few cases:\preformatted{## Joining, by = c("year", "extended_country_name", "GWn", "cown", "in_GW_system")
}\preformatted{## # A tibble: 17 x 6
## # Groups:   pmm_country, pmm_polity [17]
##    pmm_country         pmm_polity polity2 min_year max_year n_years
##    <chr>                    <int>   <dbl>    <dbl>    <dbl>   <int>
##  1 Bhutan                      -6      -5     2005     2007       3
##  2 Czech Republic               8       9     2006     2008       3
##  3 Czechoslovakia              -1       2     1947     1947       1
##  4 Equatorial Guinea           -5      -6     1996     2008      13
##  5 Ethiopia                     0       1     1993     1993       1
##  6 Ethiopia                     1      -3     2005     2008       4
##  7 Israel                       9       6     1981     1998      18
##  8 Israel                      10       6     1999     2008      10
##  9 Korea North                 -9     -10     1994     2008      15
## 10 Lebanon                      7       6     2005     2008       4
## 11 Moldova                      8       9     2005     2008       4
## 12 Montenegro                   8       9     2008     2008       1
## 13 Mozambique                   6       5     1994     2008      15
## 14 Myanmar (Burma)             -8      -6     2008     2008       1
## 15 Senegal                      8       7     2007     2008       2
## 16 United States               10       8     1967     1973       7
## 17 Yugoslavia (Serbia)          6       8     2006     2006       1
}

}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
\link{download_polity_annual}

\link{redownload_polityIV}

\link{polityIV}

Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other PMM replication data: 
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bollen_pmm}},
\code{\link{fh_pmm}},
\code{\link{hadenius_pmm}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{prc_gasiorowski}}

Other ordinal democracy indexes: 
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{fh_pmm}},
\code{\link{polityIV}}
}
\concept{PMM replication data}
\concept{democracy}
\concept{ordinal democracy indexes}
\keyword{datasets}
