% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{prc_gasiorowski}
\alias{prc_gasiorowski}
\alias{prc}
\alias{gasiorowski}
\alias{prc_pmm}
\title{The Political Regime Change (PRC) dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11384 rows and 12 columns.}
\source{
Gasiorowski, Mark J. 1996. "An Overview of the Political Regime
Change Dataset." Comparative Political Studies 29(4):469-483.

Reich, G. 2002. Categorizing Political Regimes: New Data for Old
Problems. Democratization 9 (4): 1-24.
\url{http://www.tandfonline.com/doi/pdf/10.1080/714000289}.
}
\usage{
prc_gasiorowski

prc_pmm
}
\description{
The Political Regime Change dataset. First described in Gasiorowski, Mark J.
1996. "An Overview of the Political Regime Change Dataset." Comparative
Political Studies 29(4):469-483. Available in updated form in Reich, G. 2002.
Categorizing Political Regimes: New Data for Old Problems. Democratization 9
(4): 1-24. \url{http://www.tandfonline.com/doi/pdf/10.1080/714000289}.
}
\details{
\describe{

\item{prc_gasiorowski_country}{The original country name in the dataset.}

\item{year}{Calendar year. In the \code{prc_gasiorowski} dataset, more than
one regime may be coded for a country-year. For example, Bulgaria in 1990 is
coded as an authoritarian regime (to mid 1990), then a transitional regime,
and then as a democracy, so there are three different values for Bulgaria
1990. In the \code{prc_pmm}, only the last regime (democracy, in the Bulgaria
case for 1990) is coded for the year, so there is only one value for each
country-year. }.

\item{regime}{The regime type for the year. Can be A (authoritarian), T
(transition), S (semidemocratic), and D (democratic).}

\item{start}{The start year of the regime.}

\item{end}{The end year of the regime.}

\item{prc}{Numeric version of \code{regime}. 1 = authoritarian, 2 =
transition, 3 = semidemocratic, 4 = democratic.}

\item{prc_at_end_year}{Value of \code{prc} at the end of the year. (May
differ from value at the beginning of the year when a transition occurs).
Only in the \code{pmm_gasiorowski} dataset.}

\item{prc_at_beginning_year}{Value of \code{prc} at the beginning of the
year. (May differ from value at the end of the year when a transition
occurs). Only in the \code{pmm_gasiorowski} dataset.}

\item{prc_notrans}{Same as \code{prc} but with transition values set to NA.
Only in \code{prc} dataset.}

}
}
\section{Standard descriptive variables (generated by this package)}{

 \describe{

 \item{extended_country_name}{The name of the country in the Gleditsch-Ward
 system of states, or the official name of the
 entity (for non-sovereign entities and states not in the Gleditsch and Ward
 system of states) or else a common name for disputed cases that do not have
 an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
 scheme sometimes indicates the common name of the country and (in
 parentheses) the name of an earlier incarnation of the state: thus, they
 have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
 For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
 System Membership: A Revised List of the Independent States since 1816."
 International Interactions 25: 393-413. The list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
 Ward list of independent states. For details, see Gleditsch, Kristian S. &
 Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the
 Independent States since 1816." International Interactions 25: 393-413. The
 list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{cown}{The Correlates of War numeric country code, 2016 version. This
 differs from Gleditsch and Ward's numeric country code in a few cases. See
 \url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
 the full list.}

 \item{in_GW_system}{Whether the state is "in system" (that is, is
 independent and sovereign), according to Gleditsch and Ward, for this
 particular date. Matches at the end of the year; so, for example South
 Vietnam 1975 is `FALSE` because, according to Gleditsch and Ward, the
 country ended on April 1975 (being absorbed by North Vietnam). It is also
 `TRUE` for dates beyond 2012 for countries that did not end by then, depsite
 the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note}{


The \code{prc_pmm} version of this data is taken from from Pemstein, Daniel,
Stephen A. Meserve, and James Melton. 2013. "Replication data for:
Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime
Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}. The
original dataset is found in \code{prc_gasiorowski}. The \code{prc_pmm} data only goes
back to 1945. It also differs from the original data in a few cases,
because it only includes one year per country - whereas \code{prc_gasiorowski}
can include more than one regime classificaiton for a given year. For more
detail on the differences, see the vignette. It is included here for completeness.
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{arat_pmm}}, \code{\link{blm}},
  \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{bollen_pmm}}, \code{\link{doorenspleet}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{fh_pmm}}, \code{\link{gwf_all}},
  \code{\link{hadenius_pmm}}, \code{\link{kailitz}},
  \code{\link{magaloni}}, \code{\link{mainwaring}},
  \code{\link{munck_pmm}}, \code{\link{pacl}},
  \code{\link{peps}}, \code{\link{pitf}},
  \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{svolik_regime}},
  \code{\link{uds_2014}}, \code{\link{ulfelder}},
  \code{\link{utip}}, \code{\link{vanhanen}},
  \code{\link{wahman_teorell_hadenius}}

Other trichotomous democracy indexes: \code{\link{blm}},
  \code{\link{kailitz}}, \code{\link{mainwaring}},
  \code{\link{utip}}

Other PMM replication data: \code{\link{arat_pmm}},
  \code{\link{blm}}, \code{\link{bollen_pmm}},
  \code{\link{fh_pmm}}, \code{\link{hadenius_pmm}},
  \code{\link{mainwaring}}, \code{\link{munck_pmm}},
  \code{\link{pacl}}, \code{\link{polity_pmm}}

Other events data: \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{ulfelder}}
}
\keyword{datasets}
