% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\encoding{UTF-8}
\name{prepare_democracy_data}
\alias{prepare_democracy_data}
\title{Prepare democracy data before replicating the UDS model}
\usage{
prepare_democracy_data(data, .funs)
}
\arguments{
\item{data}{A dataset of democracy scores. For the function to do anything,
the column names must contain at least one of the following strings:
\link{anckar}, \link{anrr}, \link{arat}, \link{blm}, \link{bmr}, \link{bti}, \link{bollen},
\link{doorenspleet}, \link{wgi}, \link{gwf}, \link{hadenius}, \link{kailitz}, \link{lied},
\link{munck}, \link{pacl}, \link{peps}, \link{polyarchy}, \link{polity}, \link{prc}, \link{PIPE},
\link{svmdi}, \link{svolik}, \link{ulfelder}, \link{utip}, \code{v2x},
\code{vanhanen_democratization} (from \link{vanhanen}), \link{vanhanen_pmm}, or
\link{wth}. For details of these variables, see
the documentation for \link{generate_democracy_scores_dataset} or the
documentation for the individual datasets.}

\item{.funs}{A named list of functions to modify the columns. It defaults to
the following:

\code{funs(arat = cut(., breaks = c(0, 50, 60, 70, 80, 90, 100, 109), labels
 = 1:7, include.lowest = TRUE, right = FALSE),

 hadenius = cut(., breaks =  c(0, 1, 2, 3, 4, 7, 8, 9, 10), labels = 1:8,
 include.lowest = TRUE, right = FALSE),

 bollen = cut(., breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100),
 labels = 1:10, include.lowest = TRUE, right = FALSE),

 bti = ~cut(.x, breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 vanhanen = cut(., breaks = c(0, 5, 10, 15, 20, 25, 30, 35, 50), labels =
 1:8, include.lowest = TRUE, right = FALSE),

 munck = cut(., breaks = c(0, 0.5, 0.75, 0.99, 1), labels = 1:4,
 include.lowest = TRUE, right = FALSE),

 polyarchy_dimensions = cut(., breaks = 20, include.lowest = TRUE, right =
 FALSE, ordered_result = TRUE),

 polity = ifelse(. < -10, NA, .), v2x = cut(., breaks = 20, include.lowest =
 TRUE, right = FALSE, ordered_result = TRUE),

 v2x_* = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 svmdi = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 eiu = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 wgi = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 peps = round(.),

 other = as.numeric(unclass(factor(.))))}}
}
\value{
A data frame with the transformed scores, if any.
}
\description{
This function is designed to take the democracy data included in this package
and put it in a form suitable for use with the \link[mirt:mirt]{mirt::mirt} package to replicate the
UDS model. It takes a data frame and tries to determine, from the column
names, which variables contain democracy scores.
}
\details{
If the column names contain the strings \code{arat}, \code{blm}, \code{bollen},\code{wgi},
\code{hadenius}, \code{munck}, \code{pacl}, \code{peps}, \code{polyarchy_inclusion_dimension},
\code{polyarchy_contestation_dimension}, \code{polity}, \code{prc}, \code{v2x}, \code{vanhanen_pmm}, or
\code{vanhanen_democratization}, the function performs the following
transformations by default:

\code{arat}: Following Pemstein, Meserve, and Melton's replication code (Pemstein,
Meserve, and Melton 2013), the function cuts Arat (1991)'s 0-109 democracy
score (\link{arat_pmm}) into 7 intervals with the following cutoffs: 50, 60, 70,
80, 90, and 100. The resulting score is ordinal from 1 to 8.

\code{bollen}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Bollen's (2001)'s
0-100 democracy score (\link{bollen_pmm}) into 10 intervals with the following
cutoffs: 10,20,30,40,50,60,70,80, and 90. The resulting score is ordinal from
1 to 10.

\code{bti}: the function assumes this is the Bertelsmann Transformation Index
(\link{bti}), and it will cut it into 20 categories. The resulting score is ordinal
from 1 to 20.

\code{wgi}: If the World Governance Indicator's index of voice and accountability
(\link{wgi}) is included in the file, the function cuts it into 20 categories. The
resulting score is ordinal from 1 to 20.

\code{hadenius_pmm}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Hadenius (1992)'s 0-10
democracy score (\link{hadenius_pmm}) into 8 intervals with the following cutoffs:
1, 2, 3, 4, 7, 8, and 9. The resulting score is ordinal from 1 to 8.

\code{munck}: Following Pemstein, Meserve, and Melton's replication code (Pemstein,
Meserve, and Melton 2013), the function cuts Munck's (2009)'s 0-1 democracy
score (\link{munck_pmm}) into 4 intervals with the following cutoffs: 0.5,0.5,0.75,
and 0.99. The resulting score is ordinal from 1 to 4.

\code{peps}: If any of the variants of the Participation-Enhanced Polity Score
(Moon et al 2006, \link{peps}) is included in the file, the function rounds its
value (eliminates the decimal) and then transforms it into an ordinal measure
from 1 to 21.

\code{polity}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function takes the polity scores
(\link{polity} or \link{polityIV})and puts NA for any values below -10, and then
transforms it into an ordinal measure from 1 to 21.

\code{polyarchy_inclusion_dimension}, \code{polyarchy_contestation_dimension}: If any of
the polyarchy inclusion or contestation dimensions from Coppedge, Alvarez and
Maldonado (2008, \link{polyarchy}) are included, it cuts them into into 20
categories. The resulting score is ordinal from 1 to 20.

\code{csvdmi} or \code{svdmi_2016}: the function assumes this is one of the continuous
indexes of democracy from the SVMDI dataset (Grundler and Krieger 2018,
\link{svmdi}), and it will cut it into 20 categories. The resulting score is
ordinal from 1 to 20.

\code{v2x}: If any of the \code{v2x_} continuous indexes of democracy from the V-Dem
dataset (Coppedge et al 2021) are included in the file, the function cuts them
into 20 categories. The resulting score is ordinal from 1 to 20.

\code{vanhanen_democratization} or \code{vanhanen_pmm}: Following Pemstein, Meserve, and
Melton's replication code (Pemstein, Meserve, and Melton 2013), the function
cuts Vanhanen's (2012)'s index of democratization (\link{vanhanen}) into 8
intervals with the following cutoffs: 5,10,15,20,25,30, and 35. The resulting
score is ordinal from 1 to 8.

The function also recognizes the following column names (or partial column
names - it also recognizes, e.g., \code{pmm_blm}) as measures of democracy: \code{anrr}
(from Acemoglu, Naidu, Restrepo, and Robinson 2019, \link{anrr}), \code{anckar} (from
Anckar and Fredriksson 2018 \link{anckar}), \code{blm} (from Bowman, Lehoucq, and
Mahoney 2005, \link{blm}), \code{bmr} (from Boix, Miller, and Rosato 2012, \link{bmr}),
\code{doorenspleet} (from Doorenspleet 2000, \link{doorenspleet}), \code{e_v2x} (the
"ordinal" indexes from the V-dem project, Coppedge et al 2021), \code{freedomhouse}
or \code{fh} (from Freedom House, \link{fh} - freedom scale must be reversed so that
"more freedom" is higher), \code{gwf} (from Geddes, Wright, and Frantz 2014, \link{gwf}
, the dichotomous democracy indicator only), \code{kailitz} (from Kailitz 2013 -
democracy/non-democracy indicator, \link{kailitz}), \code{lied} or \code{lexical_index} (from
Skaaning, Gerring, and Bartusevicius 2015, \link{LIED}), \code{mainwaring} (from
Mainwaring and Perez Linan 2008, \link{mainwaring}), \code{magaloni} (from Magaloni,
Min, Chu 2013 - democracy/non-democracy indicator, \link{magaloni}), \code{pacl} (from
Cheibub, Gandhi, and Vreeland 2010, \link{pacl} or \link{pacl_update}), \code{pitf} (from
Goldstone et al 2010 or Taylor and Ulfelder 2015, \link{pitf}), \code{polyarchy} (from
Coppedge and Reinicke 1991, \link{polyarchy}), \code{prc} (from Gasiorowski 1996 or
Reich 2002, \link{prc}), \code{PIPE} (from Przeworski 2010, \link{PIPE}), \code{reign} (from Bell
2016, \link{reign}), \code{svmdi} (from Grundler and Krieger 2018, 2016, \link{svmdi}),
\code{svolik} (from Svolik 2012, democracy/dictatorship indicator only, \link{svolik}),
\code{ulfelder} (from Ulfelder 2012, \link{ulfelder}), \code{utip} (from Hsu 2008, \link{utip}),
and \code{wth} or \code{wahman_teorell_hadenius} (from Wahman, Teorell, and Hadenius
2013, \link{wahman_teorell_hadenius}). In each of these cases the function
transforms the values of these scores by running
\code{as.numeric(unclass(factor(x)))}, which transforms them into ordinal variables
from 1 to the number of categories.

For details of these scores, see the documentation for
\link{generate_democracy_scores_dataset} or the documentation for the individual
datasets.

It is also possible to change these defaults.
}
\section{Note}{
 Warning! The function does not perform any sanity checks. It
will try to transform anything that has the right name. You should always
check that the results make sense.
}

\examples{
democracy <- generate_democracy_scores_dataset(
  selection = "pmm", output_format = "wide", exclude_pmm_duplicates = FALSE
  )

summary(democracy)
summary(prepare_democracy_data(democracy))

}
\references{
D. Acemoglu, S. Naidu, P. Restrepo, et al. "Democracy Does CauseGrowth". In: \emph{Journal of Political Economy} 127.1 (2019), pp. 47-100.DOI: 10.1086/700936.

C. Anckar and C. Fredriksson. "Classifying Political Regimes1800–2016: A Typology and a New Dataset". In: \emph{European PoliticalScience} 18.1 (Mar. 2018), pp. 84-96. DOI: 10.1057/s41304-018-0149-8.

Z. F. Arat. \emph{Democracy and Human Rights in Developing Countries}.Boulder: Lynne Rienner Publishers, 1991.

C. Bell. \emph{The Rulers, Elections, and Irregular Governance Dataset(REIGN)}. 2016.

M. Bernhard, T. Nordstrom, and C. Reenock. "Economic Performance,Institutional Intermediation, and Democratic Survival". In: \emph{Journal ofPolitics} 63.3 (2001), pp. 775-803. DOI: 10.1111/0022-3816.00087.

Bertelsmann Stiftung. \emph{Transformation Index of the BertelsmannStiftung 2024}. Tech. rep. Bertelsmann Stiftung, 2024.

C. Bjørnskov and M. Rode. "Regime Types and Regime Change: A NewDataset on Democracy, Coups, and Political Institutions". In: \emph{TheReview of International Organizations} 15.2 (2020), pp. 531-551. DOI:10.1007/s11558-019-09345-1.

C. Boix, M. Miller, and S. Rosato. "A Complete Data Set ofPolitical Regimes, 1800–2007". In: _Comparative Political Studies_46.12 (2012), pp. 1523-1554. DOI: 10.1177/0010414012463905.

K. A. Bollen. \emph{Cross-National Indicators of Liberal Democracy,1950-1990}. 2001. DOI: 10.3886/ICPSR02532.v2.

K. Bollen and P. Paxton. "Subjective Measures of LiberalDemocracy". In: \emph{Comparative Political Studies} 33.1 (2000), pp. 58-86.DOI: 10.1177/0010414000033001003.

K. Bowman, F. Lehoucq, and J. Mahoney. "Measuring PoliticalDemocracy: Case Expertise, Data Adequacy, and Central America". In:\emph{Comparative Political Studies} 38.8 (2005), pp. 939-970. DOI:10.1177/0010414005277083.

J. A. Cheibub, J. Gandhi, and J. R. Vreeland. "Democracy andDictatorship Revisited". In: \emph{Public Choice} 143.1-2 (2009), pp.67-101. DOI: 10.1007/s11127-009-9491-2.

M. Coppedge, A. Alvarez, and C. Maldonado. "Two PersistentDimensions of Democracy: Contestation and Inclusiveness". In: \emph{Thejournal of politics} 70.03 (2008), pp. 632-647. DOI:10.1017/S0022381608080663.

M. Coppedge, J. Gerring, C. H. Knutsen, et al. \emph{V-Dem CodebookV14}. Tech. rep. Varieties of Democracy (V-Dem) Project, 2024.

M. Coppedge, J. Gerring, C. H. Knutsen, et al. \emph{V-Dem CodebookV15}. Report. Varieties of Democracy (V-Dem) Project, 2025.

M. Coppedge, J. Gerring, C. H. Knutsen, et al. \emph{V-Dem CodebookV13}.

M. Coppedge and W. H. Reinicke. "Measuring Polyarchy". In:\emph{Studies in Comparative International Development} 25.1 (1990), pp.51-72.

R. Doorenspleet. "Reassessing the Three Waves of Democratization".In: \emph{World Politics} 52.03 (2000), pp. 384-406. DOI:10.1017/S0043887100016580.

Freedom House. \emph{Freedom in the World 2024: The Mounting Damage ofFlawed Elections and Armed Conflict}. Tech. rep. Freedom House, 2024.

M. Gasiorowski. "An Overview of the Political Regime ChangeDataset". In: \emph{Comparative Political Studies} 29.4 (1996), pp. 469-483.DOI: 10.1177/0010414096029004004.

B. Geddes, J. Wright, and E. Frantz. "Autocratic Breakdown andRegime Transitions: A New Data Set". In: _Perspectives on Politics_12.1 (2014), pp. 313-331. DOI: 10.1017/S1537592714000851.

K. S. Gleditsch and M. D. Ward. "Interstate System Membership: ARevised List of Independent States since the Congress of Vienna". In:\emph{International Interactions} 25.4 (1999), pp. 393-413. DOI:10.1080/03050629908434958.

J. Goldstone, R. Bates, D. Epstein, et al. "A Global Model forForecasting Political Instability". In: \emph{American Journal of PoliticalScience} 54.1 (2010), pp. 190-208. DOI:10.1111/j.1540-5907.2009.00426.x.

K. Gründler and T. Krieger. "Democracy and Growth: Evidence from aMachine Learning Indicator". In: \emph{European Journal of PoliticalEconomy} 45 (2016), pp. 85-107. DOI: 10.1016/j.ejpoleco.2016.05.005.

K. Gründler and T. Krieger. \emph{Machine Learning Indices, PoliticalInstitutions, and Economic Development}. Report. CESifo Group Munich,2018.

K. Gründler and T. Krieger. "Using Machine Learning for MeasuringDemocracy: A Practitioners Guide and a New Updated Dataset for 186Countries from 1919 to 2019". In: \emph{European Journal of PoliticalEconomy} (), p. 102047. DOI: 10.1016/j.ejpoleco.2021.102047.

A. Hadenius. \emph{Democracy and Development}. New York: CambridgeUniversity Press, 1992.

A. Hadenius and J. Teorell. "Pathways from Authoritarianism". In:\emph{Journal of Democracy} 18.1 (2007), pp. 143-157. DOI:10.1353/jod.2007.0009.

F. House. \emph{Freedom in the World 2025: The Uphill Battle toSafeguard Rights}. Tech. rep. Freedom House, Feb. 2025.

S. Hsu. "The Effect of Political Regimes on Inequality,1963-2002". In: \emph{UTIP Working Paper} (2008).

S. Kailitz. "Classifying Political Regimes Revisited: Legitimationand Durability". In: \emph{Democratization} 20.1 (2013), pp. 39-60.

S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Codebook.Version 1.7}. Tech. rep. Dresden: Hannah Arendt Institute forTotalitarianism Studies, 2024.

S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Dataset}.Dresden, 2024.

D. Kaufmann and A. Kraay. \emph{Worldwide Governance Indicators}. 2020.

B. Magaloni, J. Chu, and E. Min. \emph{Autocracies of the World,1950-2012 (Version 1.0)}. Dataset. 2013.

S. Mainwaring, D. Brinks, and A. Pérez-Liñán. "ClassifyingPolitical Regimes in Latin America". In: \emph{Studies in ComparativeInternational Development} 36.1 (2001), pp. 37-65. DOI:10.1007/bf02687584.

S. Mainwaring, A. Pérez-Liñán, and D. Brinks. "Political Regimesin Latin America, 1900-2007 (with Daniel Brinks)". In: \emph{Democracies andDictatorships in Latin America: Emergence, Survival, and Fall}. NewYork: Cambridge University Press, 2014.

M. G. Marshall and T. R. Gurr. _Polity 5: Political RegimeCharacteristics and Transitions, 1800-2018. Dataset Users' Manual._Manual. 2020.

M. G. Marshall, T. R. Gurr, and K. Jaggers. \emph{Polity IV Project:Political Regime Characteristics and Transitions, 1800-2018. DatasetUsers' Manual.} Manual. 2019.

B. E. Moon, J. H. Birdsall, S. Ciesluk, et al. "Voting Counts:Participation in the Measurement of Democracy". In: \emph{Studies inComparative International Development} 41.2 (2006), pp. 3-32. DOI:10.1007/BF02686309.

G. Munck. \emph{Measuring Democracy: A Bridge between Scholarship andPolitics}. Baltimore: The Johns Hopkins University Press, 2009.

X. Márquez. "A Quick Method for Extending the Unified DemocracyScores". In: \emph{Available at SSRN 2753830} (2016). DOI:10.2139/ssrn.2753830.

X. Márquez. \emph{democracyData: A Package for Accessing andManipulating Existing Measures of Democracy.} 2020.

D. Pemstein, K. L. Marquardt, E. Tzelgov, et al. \emph{The V-demMeasurement Model: Latent Variable Analysis for Cross-national andCross-temporal Expert-coded Data}. Tech. rep. 21. Varieties ofDemocracy Institute, University of Gothenburg, 2022.

D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for:Democratic Compromise: A Latent Variable Analysis of Ten Measures ofRegime Type}. 2013. DOI: 10.7910/DVN/WWYOHU.

D. Pemstein, S. Meserve, and J. Melton. "Democratic Compromise: ALatent Variable Analysis of Ten Measures of Regime Type". In:\emph{Political Analysis} 18.4 (2010), pp. 426-449. DOI: 10.1093/pan/mpq020.

A. Przeworski. \emph{Political Institutions and Political Events (PIPE)Data Set}. Data Set. 2013.

G. Reich. "Categorizing Political Regimes: New Data for OldProblems". In: \emph{Democratization} 9.4 (2002), pp. 1-24. DOI:10.1080/714000289.

S. Skaaning, J. Gerring, and H. Bartusevičius. "A Lexical Index ofElectoral Democracy". In: \emph{Comparative Political Studies} 48.12 (2015),pp. 1491-1525. DOI: 10.1177/0010414015581050.

M. Svolik. \emph{The Politics of Authoritarian Rule}. Cambridge and NewYork: Cambridge University Press, 2012.

S. J. Taylor and J. Ulfelder. "A Measurement Error Model ofDichotomous Democracy Status". In: \emph{Available at SSRN} (2015). DOI:10.2139/ssrn.2726962.

The Economist Intelligence Unit. \emph{Democracy Index 2022: FrontlineDemocracy and the Battle for Ukraine}. Tech. rep. The EconomistIntelligence Unit, 2023.

The Economist Intelligence Unit. \emph{Democracy Index 2024: What'sWrong with Representative Democracy?} Tech. rep. London, UK: TheEconomist Intelligence Unit Limited, 2025.

J. Ulfelder. \emph{Democracy/Autocracy Data Set}. 2012. DOI:10.7910/DVN/M11WFC.

J. Ulfelder and M. Lustik. "Modelling Transitions to and fromDemocracy". In: \emph{Democratization} 14.3 (2007), pp. 351-387. DOI:10.1080/13510340701303196.

T. E. I. Unit. \emph{Democracy Index 2023: Age of Conflict}. Tech. rep.London, United Kingdom: Economist Intelligence Unit, 2024.

T. Vanhanen. \emph{Measures of Democracy 1810-2018 (Dataset). Version8.0 (2019-06-17).} Tampere, 2019.

M. Wahman, J. Teorell, and A. Hadenius. "Authoritarian RegimeTypes Revisited: Updated Data in Comparative Perspective". In:\emph{Contemporary Politics} 19.1 (2013), pp. 19-34. DOI:10.1080/13569775.2013.773200.
}
