% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\encoding{UTF-8}
\name{prepare_democracy_data}
\alias{prepare_democracy_data}
\title{Prepare democracy data before replicating the UDS model}
\usage{
prepare_democracy_data(data, .funs)
}
\arguments{
\item{data}{A dataset of democracy scores. For the function to do anything,
the column names must contain at least one of the following strings:
\link{anckar}, \link{anrr}, \link{arat}, \link{blm}, \link{bmr}, \link{bti}, \link{bollen},
\link{doorenspleet}, \link{wgi}, \link{gwf}, \link{hadenius}, \link{kailitz}, \link{lied},
\link{munck}, \link{pacl}, \link{peps}, \link{polyarchy}, \link{polity}, \link{prc}, \link{PIPE},
\link{svmdi}, \link{svolik}, \link{ulfelder}, \link{utip}, \code{v2x},
\href{vanhanen.html}{vanhanen_democratization}, \link{vanhanen_pmm}, or
\href{wahman_teorell_hadenius.html}{wth}. For details of these variables, see
the documentation for \link{generate_democracy_scores_dataset} or the
documentation for the individual datasets.}

\item{.funs}{A names list of functions to modify the columns. It defaults to
the following:

\code{funs(arat = cut(., breaks = c(0, 50, 60, 70, 80, 90, 100, 109), labels
 = 1:7, include.lowest = TRUE, right = FALSE),

 hadenius = cut(., breaks =  c(0, 1, 2, 3, 4, 7, 8, 9, 10), labels = 1:8,
 include.lowest = TRUE, right = FALSE),

 bollen = cut(., breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100),
 labels = 1:10, include.lowest = TRUE, right = FALSE),

 bti = ~cut(.x, breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 vanhanen = cut(., breaks = c(0, 5, 10, 15, 20, 25, 30, 35, 50), labels =
 1:8, include.lowest = TRUE, right = FALSE),

 munck = cut(., breaks = c(0, 0.5, 0.75, 0.99, 1), labels = 1:4,
 include.lowest = TRUE, right = FALSE),

 polyarchy_dimensions = cut(., breaks = 20, include.lowest = TRUE, right =
 FALSE, ordered_result = TRUE),

 polity = ifelse(. < -10, NA, .), v2x = cut(., breaks = 20, include.lowest =
 TRUE, right = FALSE, ordered_result = TRUE),

 v2x_* = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 svmdi = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 eiu = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 wgi = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
 ordered_result = TRUE),

 peps = round(.),

 other = as.numeric(unclass(factor(.))))}}
}
\value{
A data frame with the transformed scores, if any.
}
\description{
This function is designed to take the democracy data included in this package
and put it in a form suitable for use with the \link{mirt} package to replicate the
UDS model. It takes a data frame and tries to determine, from the column
names, which variables contain democracy scores.
}
\details{
If the column names contain the strings \code{arat}, \code{blm}, \code{bollen},\code{wgi},
\code{hadenius}, \code{munck}, \code{pacl}, \code{peps}, \code{polyarchy_inclusion_dimension},
\code{polyarchy_contestation_dimension}, \code{polity}, \code{prc}, \code{v2x}, \code{vanhanen_pmm}, or
\code{vanhanen_democratization}, the function performs the following
transformations by default:

\code{arat}: Following Pemstein, Meserve, and Melton's replication code (Pemstein,
Meserve, and Melton 2013), the function cuts Arat (1991)'s 0-109 democracy
score (\link{arat_pmm}) into 7 intervals with the following cutoffs: 50, 60, 70,
80, 90, and 100. The resulting score is ordinal from 1 to 8.

\code{bollen}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Bollen's (2001)'s
0-100 democracy score (\link{bollen_pmm}) into 10 intervals with the following
cutoffs: 10,20,30,40,50,60,70,80, and 90. The resulting score is ordinal from
1 to 10.

\code{bti}: the function assumes this is the Bertelsmann Transformation Index
(\link{bti}), and it will cut it into 20 categories. The resulting score is ordinal
from 1 to 20.

\code{wgi}: If the World Governance Indicator's index of voice and accountability
(\link{wgi}) is included in the file, the function cuts it into 20 categories. The
resulting score is ordinal from 1 to 20.

\code{hadenius_pmm}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Hadenius (1992)'s 0-10
democracy score (\link{hadenius_pmm}) into 8 intervals with the following cutoffs:
1, 2, 3, 4, 7, 8, and 9. The resulting score is ordinal from 1 to 8.

\code{munck}: Following Pemstein, Meserve, and Melton's replication code (Pemstein,
Meserve, and Melton 2013), the function cuts Munck's (2009)'s 0-1 democracy
score (\link{munck_pmm}) into 4 intervals with the following cutoffs: 0.5,0.5,0.75,
and 0.99. The resulting score is ordinal from 1 to 4.

\code{peps}: If any of the variants of the Participation-Enhanced Polity Score
(Moon et al 2006, \link{peps}) is included in the file, the function rounds its
value (eliminates the decimal) and then transforms it into an ordinal measure
from 1 to 21.

\code{polity}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function takes the polity scores
(\link{polity} or \link{polityIV})and puts NA for any values below -10, and then
transforms it into an ordinal measure from 1 to 21.

\code{polyarchy_inclusion_dimension}, \code{polyarchy_contestation_dimension}: If any of
the polyarchy inclusion or contestation dimensions from Coppedge, Alvarez and
Maldonado (2008, \link{polyarchy}) are included, it cuts them into into 20
categories. The resulting score is ordinal from 1 to 20.

\code{csvdmi} or \code{svdmi_2016}: the function assumes this is one of the continuous
indexes of democracy from the SVMDI dataset (Grundler and Krieger 2018,
\link{svmdi}), and it will cut it into 20 categories. The resulting score is
ordinal from 1 to 20.

\code{v2x}: If any of the \code{v2x_} continuous indexes of democracy from the V-Dem
dataset (Coppedge et al 2021) are included in the file, the function cuts them
into 20 categories. The resulting score is ordinal from 1 to 20.

\code{vanhanen_democratization} or \code{vanhanen_pmm}: Following Pemstein, Meserve, and
Melton's replication code (Pemstein, Meserve, and Melton 2013), the function
cuts Vanhanen's (2012)'s index of democratization (\link{vanhanen}) into 8
intervals with the following cutoffs: 5,10,15,20,25,30, and 35. The resulting
score is ordinal from 1 to 8.

The function also recognizes the following column names (or partial column
names - it also recognizes, e.g., \code{pmm_blm}) as measures of democracy: \code{anrr}
(from Acemoglu, Naidu, Restrepo, and Robinson 2019, \link{anrr}), \code{anckar} (from
Anckar and Fredriksson 2018 \link{anckar}), \code{blm} (from Bowman, Lehoucq, and
Mahoney 2005, \link{blm}), \code{bmr} (from Boix, Miller, and Rosato 2012, \link{bmr}),
\code{doorenspleet} (from Doorenspleet 2000, \link{doorenspleet}), \code{e_v2x} (the
"ordinal" indexes from the V-dem project, Coppedge et al 2021), \code{freedomhouse}
or \code{fh} (from Freedom House, \link{fh} - freedom scale must be reversed so that
"more freedom" is higher), \code{gwf} (from Geddes, Wright, and Frantz 2014, \link{gwf}
, the dichotomous democracy indicator only), \code{kailitz} (from Kailitz 2013 -
democracy/non-democracy indicator, \link{kailitz}), \code{lied} or \code{lexical_index} (from
Skaaning, Gerring, and Bartusevicius 2015, \link{LIED}), \code{mainwaring} (from
Mainwaring and Perez Linan 2008, \link{mainwaring}), \code{magaloni} (from Magaloni,
Min, Chu 2013 - democracy/non-democracy indicator, \link{magaloni}), \code{pacl} (from
Cheibub, Gandhi, and Vreeland 2010, \link{pacl} or \link{pacl_update}), \code{pitf} (from
Goldstone et al 2010 or Taylor and Ulfelder 2015, \link{pitf}), \code{polyarchy} (from
Coppedge and Reinicke 1991, \link{polyarchy}), \code{prc} (from Gasiorowski 1996 or
Reich 2002, \link{prc}), \code{PIPE} (from Przeworski 2010, \link{PIPE}), \code{reign} (from Bell
2016, \link{reign}), \code{svmdi} (from Grundler and Krieger 2018, 2016, \link{svmdi}),
\code{svolik} (from Svolik 2012, democracy/dictatorship indicator only, \link{svolik}),
\code{ulfelder} (from Ulfelder 2012, \link{ulfelder}), \code{utip} (from Hsu 2008, \link{utip}),
and \code{wth} or \code{wahman_teorell_hadenius} (from Wahman, Teorell, and Hadenius
2013, \link{wahman_teorell_hadenius}). In each of these cases the function
transforms the values of these scores by running
\code{as.numeric(unclass(factor(x)))}, which transforms them into ordinal variables
from 1 to the number of categories.

For details of these scores, see the documentation for
\link{generate_democracy_scores_dataset} or the documentation for the individual
datasets.

It is also possible to change these defaults.
}
\section{Note}{
 Warning! The function does not perform any sanity checks. It
will try to transform anything that has the right name. You should always
check that the results make sense.
}

\examples{
democracy <- generate_democracy_scores_dataset(selection = "pmm", output_format = "wide")
summary(democracy)
summary(prepare_data(democracy))

}
\references{
\link{1} D. Acemoglu, S. Naidu, P. Restrepo, et al. “Democracy Does Cause Growth”. In: \emph{Journal of Political
Economy} 127.1 (2019), pp. 47-100. DOI: 10.1086/700936.

\link{2} C. Anckar and C. Fredriksson. “Classifying political regimes 1800-2016: a typology and a new dataset”.
In: \emph{European Political Science} (2018). DOI: 10.1057/s41304-018-0149-8. <URL:
https://doi.org/10.1057/s41304-018-0149-8>.

\link{3} Z. F. Arat. \emph{Democracy and human rights in developing countries}. Boulder: Lynne Rienner Publishers,
1991.

\link{4} C. Bell. \emph{The Rulers, Elections, and Irregular Governance Dataset (REIGN)}. 2016. <URL:
http://oefresearch.org/datasets/reign>.

\link{5} M. Bernhard, T. Nordstrom, and C. Reenock. “Economic Performance, Institutional Intermediation, and
Democratic Survival”. In: \emph{Journal of Politics} 63.3 (2001), pp. 775-803. DOI: 10.1111/0022-3816.00087.

\link{6} Bertelsmann Stiftung. \emph{Transformation Index of the Bertelsmann Stiftung 2020}. Tech. rep. Bertelsmann
Stiftung, 2020.

\link{7} C. Bjørnskov and M. Rode. “Regime types and regime change: A new dataset on democracy, coups, and
political institutions”. In: \emph{The Review of International Organizations} 15.2 (2020), pp. 531-551. DOI:
10.1007/s11558-019-09345-1.

\link{8} C. Boix, M. Miller, and S. Rosato. “A Complete Dataset of Political Regimes, 1800-2007”. In:
\emph{Comparative Political Studies} 46.12 (2012), pp. 1523-1554. DOI: 10.1177/0010414012463905.

\link{9} K. A. Bollen. \emph{Cross-National Indicators of Liberal Democracy, 1950-1990}. Data set. 2001.

\link{10} K. Bollen and P. Paxton. “Subjective Measures of Liberal Democracy”. In: \emph{Comparative Political
Studies} 33.1 (2000), pp. 58-86. DOI: 10.1177/0010414000033001003.

\link{11} K. Bowman, F. Lehoucq, and J. Mahoney. “Measuring Political Democracy: Case Expertise, Data Adequacy,
and Central America”. In: \emph{Comparative Political Studies} 38.8 (2005), pp. 939-970. DOI:
10.1177/0010414005277083.

\link{12} J. Cheibub, J. Gandhi, and J. Vreeland. “Democracy and dictatorship revisited”. In: \emph{Public Choice}
143.1 (2010), pp. 67-101. DOI: 10.1007/s11127-009-9491-2.

\link{13} M. Coppedge, A. Alvarez, and C. Maldonado. “Two Persistent Dimensions of Democracy: Contestation and
Inclusiveness”. In: \emph{The journal of politics} 70.03 (2008), pp. 632-647. DOI: 10.1017/S0022381608080663.

\link{14} M. Coppedge, J. Gerring, C. H. Knutsen, et al. \emph{V-Dem Country-Year/Country-Date Dataset v11}. 2021.
<URL: https://www.v-dem.net/en/data/data-version-11/>.

\link{15} M. Coppedge, J. Gerring, C. H. Knutsen, et al. \emph{V-Dem Codebook v11}. 2021. <URL:
https://www.v-dem.net/en/data/data-version-11/>.

\link{16} M. Coppedge and W. H. Reinicke. “Measuring Polyarchy”. In: \emph{Studies in Comparative International
Development} 25.1 (1990), pp. 51-72. DOI: 10.1007/Bf02716905.

\link{17} R. Doorenspleet. “Reassessing the Three Waves of Democratization”. In: \emph{World Politics} 52.03 (2000),
pp. 384-406. DOI: 10.1017/S0043887100016580.

\link{18} Freedom House. \emph{Freedom in the World 2021: Democracy under Siege}. Tech. rep. Freedom House, 2021.
<URL: https://freedomhouse.org/report/freedom-world/2021/democracy-under-siege>.

\link{19} M. Gasiorowski. “An Overview of the Political Regime Change Dataset”. In: \emph{Comparative Political
Studies} 29.4 (1996), pp. 469-483. DOI: 10.1177/0010414096029004004.

\link{20} B. Geddes, J. Wright, and E. Frantz. “Autocratic Breakdown and Regime Transitions: A New Data Set”.
In: \emph{Perspectives on Politics} 12.1 (2014), pp. 313-331. DOI: 10.1017/S1537592714000851.

\link{21} K. Gleditsch and M. D. Ward. “Interstate system membership: A revised list of independent states since
the congress of Vienna”. In: \emph{International Interactions} 25.4 (1999), pp. 393-413. DOI:
10.1080/03050629908434958.

\link{22} J. Goldstone, R. Bates, D. Epstein, et al. “A Global Model for Forecasting Political Instability”. In:
\emph{American Journal of Political Science} 54.1 (2010), pp. 190-208. DOI: 10.1111/j.1540-5907.2009.00426.x.

\link{23} K. Gründler and T. Krieger. “Democracy and growth: Evidence from a machine learning indicator”. In:
\emph{European Journal of Political Economy} 45 (2016), pp. 85-107. DOI:
https://doi.org/10.1016/j.ejpoleco.2016.05.005. <URL:
http://www.sciencedirect.com/science/article/pii/S0176268016300222>.

\link{24} K. Gründler and T. Krieger. \emph{Machine Learning Indices, Political Institutions, and Economic
Development}. Report. CESifo Group Munich, 2018. <URL:
https://www.cesifo-group.de/DocDL/cesifo1_wp6930.pdf>.

\link{25} K. Gründler and T. Krieger. “Using Machine Learning for measuring democracy: A practitioners guide and
a new updated dataset for 186 countries from 1919 to 2019”. In: \emph{European Journal of Political Economy}
(2021), pp. 102-47. DOI: https://doi.org/10.1016/j.ejpoleco.2021.102047.

\link{26} A. Hadenius. \emph{Democracy and development}. New York: Cambridge University Press, 1992.

\link{27} A. Hadenius and J. Teorell. “Pathways from Authoritarianism”. In: \emph{Journal of Democracy} 18.1 (2007),
pp. 143-157.

\link{28} S. Hsu. “The Effect of Political Regimes on Inequality, 1963-2002”. In: \emph{UTIP Working Paper} (2008).

\link{29} S. Kailitz. “Classifying political regimes revisited: legitimation and durability”. In:
\emph{Democratization} 20.1 (2013), pp. 39-60. DOI: 10.1080/13510347.2013.738861.

\link{30} D. Kaufmann and A. Kraay. \emph{Worldwide Governance Indicators}. 2020. <URL:
http://www.govindicators.org>.

\link{31} B. Magaloni, J. Chu, and E. Min. \emph{Autocracies of the World, 1950-2012 (Version 1.0)}. Dataset. 2013.
<URL: http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset>.

\link{32} S. Mainwaring, D. Brinks, and A. Pérez-Liñán. “Classifying Political Regimes in Latin America”. In:
\emph{Studies in Comparative International Development} 36.1 (2001), pp. 37-65. DOI: 10.1007/bf02687584.

\link{33} S. Mainwaring, A. Pérez-Liñán, and D. Brinks. “Political Regimes in Latin America, 1900-2007 (with
Daniel Brinks)”. In: \emph{Democracies and Dictatorships in Latin America: Emergence, Survival, and Fall}. New
York: Cambridge University Press, 2014. Chap. Political Regimes in Latin America, 1900-2007 (with Daniel
Brinks).

\link{34} X. Marquez. “A Quick Method for Extending the Unified Democracy Scores”. In: \emph{Available at SSRN
2753830} (2016). <URL: https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2753830>.

\link{35} X. Márquez. \emph{democracyData: A package for accessing and manipulating existing measures of democracy.}
2020. <URL: http://github.com/xmarquez/democracyData>.

\link{36} M. G. Marshall and T. R. Gurr. \emph{Polity 5: Political Regime Characteristics and Transitions, 1800-2018.
Dataset Users' Manual.} Center for Systemic Peace, 2020.

\link{37} M. G. Marshall, T. R. Gurr, and K. Jaggers. \emph{Polity IV Project: Political Regime Characteristics and
Transitions, 1800-2018. Dataset Users' Manual.} Center for Systemic Peace, 2019.

\link{38} B. E. Moon, J. H. Birdsall, S. Ciesluk, et al. “Voting Counts: Participation in the Measurement of
Democracy”. In: \emph{Studies in Comparative International Development} 41.2 (2006), pp. 3-32. DOI:
10.1007/BF02686309.

\link{39} G. Munck. \emph{Measuring Democracy: A Bridge between Scholarship and Politics}. Baltimore: The Johns
Hopkins University Press, 2009.

\link{40} D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication data for: Democratic Compromise: A Latent
Variable Analysis of Ten Measures of Regime Type}. 2013. <URL: http://hdl.handle.net/1902.1/PMM>.

\link{41} D. Pemstein, S. Meserve, and J. Melton. “Democratic Compromise: A Latent Variable Analysis of Ten
Measures of Regime Type”. In: \emph{Political Analysis} 18.4 (2010), pp. 426-449. DOI: 10.1093/pan/mpq020.

\link{42} A. Przeworski. \emph{Political Institutions and Political Events (PIPE) Data Set}. Data set. 2013. <URL:
https://sites.google.com/a/nyu.edu/adam-przeworski/home/data>.

\link{43} G. Reich. “Categorizing Political Regimes: New Data for Old Problems”. In: \emph{Democratization} 9.4
(2002), pp. 1-24. DOI: 10.1080/714000289.

\link{44} S. Skaaning, J. Gerring, and H. Bartusevicius. “A Lexical Index of Electoral Democracy”. In:
\emph{Comparative Political Studies} 48.12 (2015), pp. 1491-1525. DOI: 10.1177/0010414015581050.

\link{45} M. Svolik. \emph{The Politics of Authoritarian Rule}. Cambridge: Cambridge University Press, 2012.

\link{46} S. J. Taylor and J. Ulfelder. “A Measurement Error Model of Dichotomous Democracy Status”. In:
\emph{Available at SSRN} (2015). DOI: 10.2139/ssrn.2726962.

\link{47} The Economist Intelligence Unit. \emph{Democracy Index 2020: In sickness and in health?} Tech. rep. The
Economist Intelligence Unit, 2021.

\link{48} J. Ulfelder. \emph{Democracy/Autocracy Data Set}. 2012. <URL: http://hdl.handle.net/1902.1/18836>.

\link{49} J. Ulfelder and M. Lustik. “Modelling Transitions To and From Democracy”. In: \emph{Democratization} 14.3
(2007), pp. 351-387. DOI: 10.1080/13510340701303196.

\link{50} T. Vanhanen. \emph{Measures of Democracy 1810-2012}. Tampere, 2014. <URL:
http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html>.

\link{51} M. Wahman, J. Teorell, and A. Hadenius. “Authoritarian Regime Types Revisited: Updated Data in
Comparative Perspective”. In: \emph{Contemporary Politics} 19.1 (2013), pp. 19-34. <URL:
https://sites.google.com/site/authoritarianregimedataset/data>.
}
