% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_UDS.R
\name{raterinfo}
\alias{raterinfo}
\title{Extract rater info from a UD model in a tidy format.}
\usage{
raterinfo(model)
}
\arguments{
\item{model}{A \link[mirt:SingleGroupClass-class]{mirt::SingleGroupClass} model
of the democracy scores.}
}
\value{
A data frame with rater information for each democracy index over the
range of the latent variable \code{theta}.
}
\description{
Extract rater info from a UD model in a tidy format.
}
\examples{
\donttest{
# Replicate the official UDS scores (2011 release)
library(dplyr)
library(mirt)
democracy_data <- generate_democracy_scores_dataset(selection = "_pmm", output_format = "wide")

democracy_data <- prepare_democracy_data(democracy_data)

replication_2011_model <- mirt(democracy_data \%>\% select(matches("pmm")), model = 1, itemtype = "graded", SE = TRUE, verbose = FALSE)

raterinfo(replication_2011_model)}
}
