% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redownload_functions.R
\name{redownload_blm}
\alias{redownload_blm}
\alias{redownload_bmr}
\alias{redownload_bnr}
\alias{redownload_gwf}
\alias{redownload_lied}
\alias{redownload_pacl}
\alias{redownload_peps}
\alias{redownload_utip}
\alias{redownload_wahman_teorell_hadenius}
\alias{redownload_polyarchy_original}
\alias{redownload_polyarchy_dimensions}
\alias{redownload_magaloni}
\alias{redownload_uds}
\alias{redownload_ulfelder}
\alias{redownload_pipe}
\title{Download and process various democracy datasets.}
\source{
Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring
Political Democracy: Case Expertise, Data Adequacy, and Central America.
Comparative Political Studies 38 (8): 939-970.
\url{http://cps.sagepub.com/content/38/8/939}. Data available at
\url{http://www.blmdemocracy.gatech.edu/}.

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete
Data Set of Political Regimes, 1800-2007. Comparative Political Studies 46
(12): 1523-1554. Available at \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/FJLMKT}

Michael Bernhard, Timothy Nordstrom, and Christopher Reenock,
"Economic Performance, Institutional Intermediation and Democratic
Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding
description available at
\url{http://users.clas.ufl.edu/bernhard/content/data/data.htm}

Barbara Geddes, Joseph Wright, and Erica Frantz.
2014. "Autocratic Breakdown and Regime Transitions: A New Data Set."
Perspectives on Politics 12(2): 313-331. The full data and codebook can be
downloaded here \url{http://sites.psu.edu/dictators/}.

Skaaning, Svend-Erik; John
Gerring; and Henrikas Bartusevicius (2015). "A Lexical Index of Electoral
Democracy." Comparative Political Studies, Vol. 48, No. 12, pp. 1491-1525.
Original data and variable descriptions available at
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/29106}

Cheibub, Jose Antonio, Jennifer Gandhi, and James
Raymond Vreeland. 2010. "Democracy and Dictatorship Revisited." Public
Choice, vol. 143, no. 2-1, pp. 67-101. DOI: 10.1007/s11127-009-9491-2. The
full data and codebook can be downloaded here
\url{https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited}

Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett,
Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong
(2006) "Voting Counts: Participation in the Measurement of Democracy" Studies
in Comparative International Development 42, 2 (Summer, 2006). The complete
dataset is available here:
\url{http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm}.

The University of Texas Inequality Project Categorical Dataset of Political
Regimes. Described in Sara Hsu, "The Effect of Political Regimes on
Inequality, 1963-2002," UTIP Working Paper No. 53 (2008),
\url{http://utip.gov.utexas.edu/papers/utip_53.pdf}. Data available for
download at
\url{http://utip.gov.utexas.edu/data/}

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian
regime types revisited: updated data in comparative perspective.
Contemporary Politics 19 (1): 19-34. The dataset and codebook can be
downloaded from
\url{https://sites.google.com/site/authoritarianregimedataset/data}

Michael Coppedge and Wolfgang Reinicke, "Measuring Polyarchy," Studies in
Comparative International Development 25:1 (Spring 1990): 51-72. Data
available at \url{http://www3.nd.edu/~mcoppedg/crd/datacrd.htm}

Michael Coppedge, Angel Alvarez, and Claudia
Maldonado, "Two Persistent Dimensions of Democracy: Contestation and
Inclusiveness," Journal of Politics 70:3 (July 2008): 632-647.

Magaloni, Beatriz, Jonathan Chu, and Eric Min.
2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford
University. Original data and codebook available at
\url{http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/}.

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic
Compromise: A Latent Variable Analysis of Ten Measures of Regime Type.
Political Analysis 18 (4): 426-449. Original data and codebook available at
\url{http://www.unified-democracy-scores.org/}.

Jay Ulfelder. 2012. Democracy/Autocracy Data Set. \url{http://hdl.handle.net/1902.1/18836}.

Adam Przeworski. 2013 \emph{Political Institutions and Political Events
(PIPE) Data Set}. Data set.
\url{https://sites.google.com/a/nyu.edu/adam-przeworski/home/data}.
}
\usage{
redownload_blm(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bmr(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bnr(url, verbose = TRUE, extend = FALSE, return_raw = FALSE,
  ...)

redownload_gwf(url, verbose = TRUE, extend = FALSE, dataset = c("all",
  "autocratic only"), return_raw = FALSE, ...)

redownload_lied(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_pacl(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_peps(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_utip(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_wahman_teorell_hadenius(url, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_polyarchy_original(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_polyarchy_dimensions(url, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_magaloni(url, verbose = TRUE, extend = FALSE,
  return_raw = FALSE, ...)

redownload_uds(release_year = 2014, verbose = TRUE, return_raw = FALSE,
  ...)

redownload_ulfelder(url, verbose = TRUE, return_raw = FALSE,
  extend = FALSE, ...)

redownload_pipe(url, verbose = TRUE, return_raw = FALSE, ...)
}
\arguments{
\item{url}{The URL of the dataset. This defaults to:

\itemize{

\item For \link{blm}: \code{http://www.blmdemocracy.gatech.edu/blm final data.xls}

\item For \link{bmr}:
\url{https://dataverse.harvard.edu/api/access/datafile/3130643}

\item For \link{bnr}:
\url{http://users.clas.ufl.edu/bernhard/content/data/meister1305.dta}

\item For \link{gwf_all} and \link{gwf_autocratic}:
\url{http://sites.psu.edu/dictators/wp-content/uploads/sites/12570/2016/05/GWF-Autocratic-Regimes-1.2.zip}

\item For \link{LIED}:
\url{https://dataverse.harvard.edu/api/access/datafile/2863101}

\item For \link{pacl}:
\url{https://uofi.box.com/shared/static/bba3968d7c3397c024ec.dta}

\item For \link{peps}:
\url{http://www.lehigh.edu/~bm05/democracy/PEPS1pub.dta}

\item For \link{utip}: \code{http://utip.lbj.utexas.edu/data/political regime data set RV.xls}

\item For \link{wahman_teorell_hadenius}:
\url{https://sites.google.com/site/authoritarianregimedataset/data/ARD_V6.dta?attredirects=0&d=1}

\item For \link{polyarchy}:
\url{https://www3.nd.edu/~mcoppedg/crd/poly8500.sav}

\item For \link{polyarchy_dimensions}:
\url{http://www3.nd.edu/~mcoppedg/crd/DahlDims.sav}

\item For \link{magaloni}:
\url{http://cddrl.fsi.stanford.edu/sites/default/files/res/Data_Set.xls}

\item For UDS: for the 2014 release, it defaults to
\url{http://www.unified-democracy-scores.org/files/20140312/z/uds_summary.csv.gz};
for the the 2011 release, to
\url{http://www.unified-democracy-scores.org/files/20110104/uds_summary.csv.gz};
and for the 2010 release, to
\url{http://www.unified-democracy-scores.org/files/20100726/uds_summary.csv.gz}

\item For \link{ulfelder}:
\url{https://dataverse.harvard.edu/api/access/datafile/2420018}

\item For \link{PIPE}:
\url{https://sites.google.com/a/nyu.edu/adam-przeworski/home/data} }}

\item{verbose}{Whether to print a running commentary of what the function is
doing while processing the data.}

\item{return_raw}{Whether to return the raw data, without any processing.
Default is \code{FALSE}.}

\item{...}{Other parameters passed to \link{country_year_coder}.}

\item{extend}{(Only for \link{redownload_bnr}, \link{redownload_gwf},
\link{redownload_magaloni}, and \link{redownload_ulfelder}). Whether to
extend the dataset back in time using a full panel of independent countries
(for the \link{redownload_bnr} case) or the appropriate duration variable
(\code{gwf_duration}, \code{duration_nr}, or \code{rgjdura} and
\code{rgjdurd}, respectively, for \link{redownload_gwf},
\link{redownload_magaloni}, and \link{redownload_ulfelder}). For example,
the United States enters the GWF dataset in 1946, where \code{gwf_duration}
is already 75; one can extend the dataset to indicate that the country was
classified as a democracy from 1872. Default is \code{FALSE}.}

\item{dataset}{(Only for \link{redownload_gwf}). The dataset to output. Geddes, Wright, and Frantz provide
two country-year files, one with autocratic regimes only ("autocratic only"), and one with both
democratic and non-democratic regimes ("all"). Default is "all".}

\item{release_year}{(Only in \link{redownload_uds}). The year of the UDS release
to be downloaded. Can be 2014, 2011, or 2010. Default is 2014.}
}
\value{
A \link{tibble} with the processed dataset, unless \code{return_raw}
is \code{TRUE}, in which case the function returns the raw data without
processing.
}
\description{
Download and process various democracy datasets. Note that the datasets
returned by the \code{redownload_*} family of functions (\link{blm},
\link{bmr}, \link{bnr}, \link{gwf_autocratic}, \link{gwf_autocratic_extended},
\link{gwf_all}, \link{gwf_all_extended}, \link{LIED}, \link{magaloni},
\link{pacl}, \link{PIPE}, \link{peps}, \link{polyarchy},
\link{polyarchy_dimensions}, \link{uds_2014}, \link{uds_2010},
\link{uds_2011}, \link{ulfelder}, \link{utip}, \link{wahman_teorell_hadenius})
are all available directly from this package and are unlikely to have changed
since the package was installed. Access the respective dataset by typing its
name, and refer to their documentation for details. You will not normally need
to redownload them, unless you want to process the raw data yourself (set
\code{return_raw = TRUE}) or suspect they have changed since the package was
installed.
}
\examples{
\dontrun{
redownload_blm()
}
\dontrun{
bmr <- redownload_bmr()
bmr
}
\dontrun{
redownload_bnr()
redownload_bnr(full_panel = FALSE)}
\dontrun{
gwf <- redownload_gwf()
gwf
}
\dontrun{
redownload_lied()}
\dontrun{
redownload_pacl()}
\dontrun{
redownload_peps()}
\dontrun{
redownload_utip()}
\dontrun{
redownload_wahman_teorell_hadenius()}
\dontrun{
redownload_polyarchy_original()}
\dontrun{
redownload_polyarchy_dimensions()}
\dontrun{
redownload_magaloni()}
\dontrun{
redownload_uds()
redownload_uds(2011)}
\dontrun{
redownload_ulfelder()}
\dontrun{
redownload_pipe()}
}
