% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redownload_functions.R
\name{redownload_blm}
\alias{redownload_blm}
\alias{redownload_anckar}
\alias{redownload_bmr}
\alias{redownload_bnr}
\alias{redownload_gwf}
\alias{redownload_pacl}
\alias{redownload_peps}
\alias{redownload_utip}
\alias{redownload_wahman_teorell_hadenius}
\alias{redownload_polyarchy_original}
\alias{redownload_polyarchy_dimensions}
\alias{redownload_magaloni}
\alias{redownload_svmdi}
\alias{redownload_ulfelder}
\alias{redownload_pipe}
\alias{redownload_bti}
\alias{redownload_pacl_update}
\title{Download and process various democracy datasets.}
\source{
Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring
Political Democracy: Case Expertise, Data Adequacy, and Central America.
Comparative Political Studies 38 (8): 939-970.
\url{http://cps.sagepub.com/content/38/8/939}. Data available at
\url{http://www.blmdemocracy.gatech.edu/}.

Anckar, Carsten and Cecilia Fredriksson (2018). "Classifying
political regimes 1800-2016: a typology and a new dataset." European
Political Science, doi: 10.1057/s41304-018-0149-8. Data, article, and
codebook available at: \url{https://doi.org/10.1057/s41304-018-0149-8}

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete
Data Set of Political Regimes, 1800-2007. Comparative Political Studies 46
(12): 1523-1554. Available at \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/FJLMKT}

Michael Bernhard, Timothy Nordstrom, and Christopher Reenock,
"Economic Performance, Institutional Intermediation and Democratic
Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding
description available at
\url{http://users.clas.ufl.edu/bernhard/content/data/data.htm}

Barbara Geddes, Joseph Wright, and Erica Frantz.
2014. "Autocratic Breakdown and Regime Transitions: A New Data Set."
Perspectives on Politics 12(2): 313-331. The full data and codebook can be
downloaded here \url{http://sites.psu.edu/dictators/}.

Cheibub, Jose Antonio, Jennifer Gandhi, and James
Raymond Vreeland. 2010. "Democracy and Dictatorship Revisited." Public
Choice, vol. 143, no. 2-1, pp. 67-101. DOI: 10.1007/s11127-009-9491-2. The
full data and codebook can be downloaded here
\url{https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited}

Bruce E. Moon, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett,
Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong
(2006) "Voting Counts: Participation in the Measurement of Democracy" Studies
in Comparative International Development 42, 2 (Summer, 2006). The complete
dataset is available here:
\url{http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm}.

The University of Texas Inequality Project Categorical Dataset of Political
Regimes. Described in Sara Hsu, "The Effect of Political Regimes on
Inequality, 1963-2002," UTIP Working Paper No. 53 (2008),
\url{http://utip.gov.utexas.edu/papers/utip_53.pdf}. Data available for
download at
\url{http://utip.gov.utexas.edu/data/}

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian
regime types revisited: updated data in comparative perspective.
Contemporary Politics 19 (1): 19-34. The dataset and codebook can be
downloaded from
\url{https://sites.google.com/site/authoritarianregimedataset/data}

Michael Coppedge and Wolfgang Reinicke, "Measuring Polyarchy," Studies in
Comparative International Development 25:1 (Spring 1990): 51-72. Data
available at \url{http://www3.nd.edu/~mcoppedg/crd/datacrd.htm}

Michael Coppedge, Angel Alvarez, and Claudia
Maldonado, "Two Persistent Dimensions of Democracy: Contestation and
Inclusiveness," Journal of Politics 70:3 (July 2008): 632-647.

Magaloni, Beatriz, Jonathan Chu, and Eric Min.
2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford
University. Original data and codebook available at
\url{http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/}.

Grundler, Klaus, and Tommy Krieger. 2018. "Machine Learning
Indicators, Political Institutions, and Economic Development." CESifo
Working Paper. Original data available at
\url{https://www.dropbox.com/s/a7yqs5txt3qpwn0/Index\%20Upload.xlsx?dl=0}. Working paper available at
\url{https://www.cesifo-group.de/DocDL/cesifo1_wp6930.pdf}

Jay Ulfelder. 2012. Democracy/Autocracy Data Set. \url{http://hdl.handle.net/1902.1/18836}.

Adam Przeworski. 2013 \emph{Political Institutions and Political Events
(PIPE) Data Set}. Data set.
\url{https://sites.google.com/a/nyu.edu/adam-przeworski/home/data}.

Transformation Index of the Bertelsmann Stiftung 2020. Bertelsmann
Stiftung. Available at
\url{https://www.bti-project.org/en/index/political-transformation.html}

Bjørnskov, C. and M. Rode (2020). "Regime types and regime change: A
new dataset on democracy, coups, and political institutions." The Review of
International Organizations 15(2): 531-551. Available at
\url{http://www.christianbjoernskov.com/bjoernskovrodedata/}
}
\usage{
redownload_blm(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_anckar(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bmr(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bnr(url, verbose = TRUE, extend = FALSE, return_raw = FALSE, ...)

redownload_gwf(
  url,
  verbose = TRUE,
  extend = FALSE,
  dataset = c("all", "autocratic only"),
  return_raw = FALSE,
  ...
)

redownload_pacl(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_peps(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_utip(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_wahman_teorell_hadenius(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

redownload_polyarchy_original(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_polyarchy_dimensions(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_magaloni(
  url,
  verbose = TRUE,
  extend = FALSE,
  return_raw = FALSE,
  ...
)

redownload_svmdi(
  url,
  release_year = 2020,
  verbose = TRUE,
  return_raw = FALSE,
  ...
)

redownload_ulfelder(
  url,
  verbose = TRUE,
  return_raw = FALSE,
  extend = FALSE,
  ...
)

redownload_pipe(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_bti(url, verbose = TRUE, return_raw = FALSE, ...)

redownload_pacl_update(url, verbose = TRUE, return_raw = FALSE, ...)
}
\arguments{
\item{url}{The URL of the dataset. This defaults to:

\itemize{

\item For \link{anckar}:
\url{https://static-content.springer.com/esm/art\%3A10.1057\%2Fs41304-018-0149-8/MediaObjects/41304_2018_149_MOESM2_ESM.xlsx}

\item For \link{blm}:
\url{http://www.blmdemocracy.gatech.edu/blm\%20final\%20data.xls}

\item For \link{bmr}:
\url{https://dataverse.harvard.edu/api/access/datafile/3130643}

\item For \link{bnr}:
\url{http://users.clas.ufl.edu/bernhard/content/data/meister1305.dta}

\item For \link{bti}:
\url{https://www.bti-project.org/content/en/downloads/data/BTI\%202006-2020\%20Scores.xlsx}

\item For \link{gwf_all} and \link{gwf_autocratic}:
\url{http://sites.psu.edu/dictators/wp-content/uploads/sites/12570/2016/05/GWF-Autocratic-Regimes-1.2.zip}

\item For \link{LIED}:
\url{https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/29106/SXRLK1}

\item For \link{pacl}:
\url{https://uofi.box.com/shared/static/bba3968d7c3397c024ec.dta}

\item For \link{pacl_update}
\url{http://www.christianbjoernskov.com/wp-content/uploads/2020/09/Bj\%C3\%B8rnskov-Rode-integrated-dataset-v3.2.xlsx}

\item For \link{peps}:
\url{http://www.lehigh.edu/~bm05/democracy/PEPS1pub.dta}

\item For \link{svmdi}:
\url{https://www.dropbox.com/s/a7yqs5txt3qpwn0/Index\%20Upload.xlsx?dl=1}.

For the 2016 release, it defaults to
\url{http://www.wiwi.uni-wuerzburg.de/fileadmin/12010400/Data.dta}

\item For \link{utip}:
\url{http://utip.lbj.utexas.edu/data/political\%20regime\%20data\%20set\%20RV.xls}

\item For \link{wahman_teorell_hadenius}:
\url{https://sites.google.com/site/authoritarianregimedataset/data/ARD_V6.dta?attredirects=0&d=1}

\item For \link{polyarchy}:
\url{https://www3.nd.edu/~mcoppedg/crd/poly8500.sav}

\item For \link{polyarchy_dimensions}:
\url{http://www3.nd.edu/~mcoppedg/crd/DahlDims.sav}

\item For \link{magaloni}:
\url{https://fsi-live.s3.us-west-1.amazonaws.com/s3fs-public/res/Data_Set.xls}

\item For \link{ulfelder}:
\url{https://dataverse.harvard.edu/api/access/datafile/2420018}

\item For \link{PIPE}:
\url{https://sites.google.com/a/nyu.edu/adam-przeworski/home/data}

}}

\item{verbose}{Whether to print a running commentary of what the function is
doing while processing the data.}

\item{return_raw}{Whether to return the raw data, without any processing.
Default is \code{FALSE}.}

\item{...}{Other parameters passed to \link{country_year_coder}.}

\item{extend}{(Only for \link{redownload_bnr}, \link{redownload_gwf},
\link{redownload_magaloni}, and \link{redownload_ulfelder}). Whether to
extend the dataset back in time using a full panel of independent countries
(for the \link{redownload_bnr} case) or the appropriate duration variable
(\code{gwf_duration}, \code{duration_nr}, or \code{rgjdura} and
\code{rgjdurd}, respectively, for \link{redownload_gwf},
\link{redownload_magaloni}, and \link{redownload_ulfelder}). For example,
the United States enters the GWF dataset in 1946, where \code{gwf_duration}
is already 75; one can extend the dataset to indicate that the country was
classified as a democracy from 1872. Default is \code{FALSE}.}

\item{dataset}{(Only for \link{redownload_gwf}). The dataset to output. Geddes, Wright, and Frantz provide
two country-year files, one with autocratic regimes only ("autocratic only"), and one with both
democratic and non-democratic regimes ("all"). Default is "all".}

\item{release_year}{(Only in \link{redownload_svmdi}). The year of the release
to be downloaded. For \link{svmdi}, it can be 2016 or 2020.}
}
\value{
A \link{tibble} with the processed dataset, unless \code{return_raw}
is \code{TRUE}, in which case the function returns the raw data without
processing.
}
\description{
Download and process various democracy datasets. Note that the datasets
returned by the \code{redownload_*} family of functions (\link{blm},
\link{bmr}, \link{bnr}, \link{bti}, \link{gwf_autocratic},
\link{gwf_autocratic_extended}, \link{gwf_all}, \link{gwf_all_extended},
\link{LIED}, \link{magaloni}, \link{pacl}, \link{pacl_update}, \link{PIPE}, \link{peps},
\link{polityIV}, \link{polyarchy}, \link{polyarchy_dimensions}, \link{uds_2014},
\link{uds_2010}, \link{uds_2011}, \link{ulfelder}, \link{utip},
\link{wahman_teorell_hadenius}, \link{anckar}, \link{svmdi}) are all available
directly from this package and are unlikely to have changed since the package
was installed. Access the respective dataset by typing its name, and refer to
their documentation for details. You will not normally need to redownload
them, unless you want to process the raw data yourself (set \code{return_raw =
TRUE}) or suspect they have changed since the package was installed.
}
\examples{
\dontrun{
blm <- redownload_blm()
blm
}
\dontrun{
redownload_anckar()}
\dontrun{
bmr <- redownload_bmr()
bmr
}
\dontrun{
redownload_bnr()
redownload_bnr(full_panel = FALSE)}
\dontrun{
gwf <- redownload_gwf()
gwf
}
\dontrun{
redownload_pacl()}
\dontrun{
redownload_peps()}
\dontrun{
redownload_utip()}
\dontrun{
redownload_wahman_teorell_hadenius()}
\dontrun{
redownload_polyarchy_original()}
\dontrun{
redownload_polyarchy_dimensions()}
\dontrun{
redownload_magaloni()}
\dontrun{
redownload_svmdi(release_year = 2016)
redownload_svmdi() # For release year 2020}
\dontrun{
redownload_ulfelder()}
\dontrun{
redownload_pipe()}
\dontrun{
bti <- redownload_bti()
bti
}
\dontrun{
redownload_pacl_update()}
}
