% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{REIGN}
\alias{REIGN}
\alias{reign}
\title{The Rulers, Elections, and Irregular Governance Dataset (Regime data)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13937 rows and 14 columns.
}
\source{
Bell, Curtis. 2016. The Rulers, Elections, and Irregular Governance
Dataset (REIGN). Broomfield, CO: OEF Research. Available at
\url{http://oefresearch.org/datasets/reign}
}
\usage{
REIGN
}
\description{
Bell, Curtis. 2016. The Rulers, Elections, and Irregular Governance Dataset
(REIGN). Broomfield, CO: OEF Research. The dataset, variable descriptions,
and technical notes, are available at
\url{http://oefresearch.org/datasets/reign}. The variable descriptions below
are taken from
\url{http://oefresearch.org/sites/default/files/REIGN_descriptions.pdf}.
Collection stopped in 2021; this dataset is now archived in this package.
}
\section{Overview}{


The REIGN regime characteristics dataset is based on the regime
classifications of Geddes, Wright, and Frantz (\link{gwf_autocratic}), but it is
updated monthly, and it has some differences with \link{gwf_autocratic}. These
are described as follows in the online technical notes:

REIGN data is updated to the present month (up to August 2021). We also
added the following countries to the  dataset: The Bahamas, Jamaica,
Trinidad and Tobago, Barbados, Dominica, Grenada, St. Lucia, St. Vincent
and the Grenadines, Antigua and Barbuda, St. Kitts and Nevis, Belize,
Guyana, Surinam, Andorra, San Marino, Malta, Cyprus, Cape Verde, Sao Tome
and Principe, Equatorial Guinea, Djibouti, Comoros, Mauritius, Seychelles,
South Sudan, Bhutan, Maldives, South Vietnam, Brunei, East Timor, Papua New
Guinea, Vanuatu, Solomon Islands, Kiribati, Tuvalu, Fiji, Tonga, Nauru,
Marshall Islands, Palau, Micronesia, and Samoa.

We also added political systems that lasted for less than a year and
specified start and end months. This allows for a more granular look at
transitional periods and interim governments. These new short-lived regimes
appear in countries including Haiti, the Dominican Republic, Guatemala,
Peru, Benin, Niger, Ghana, and Nigeria.

Finally, GWF generally does not allow for yearly changes in institutions
over the course of a regime. If a government begins as personalist, it
stays personalist until that leader leaves power. In several cases we code
a change in regime type following reforms during the tenure of a single
leader, including Lanzana in Burkina Faso, Rawlings in Ghana, etc.

Further discussion of the differences between REIGN and GWF
(\link{gwf_autocratic}) is available here:
\url{http://oefresearch.org/sites/default/files/REIGN\%20Data\%202016.10.05.pdf}
}

\section{Variables}{


\describe{
\item{reign_cowcode}{The COW code of the country in REIGN.}

\item{reign_country}{The country name in REIGN.}

\item{gwf_casename}{The case name in the \link{gwf} dataset.}

\item{gwf_startdate}{The startdate of the case in the \link{gwf} dataset.}

\item{gwf_enddate}{The enddate of the case in the \link{gwf} dataset.}

\item{Start}{The start date of the regime.}

\item{End}{The end date of the regime, or 31 December of the current year.}

\item{gwf_regimetype}{The regime type. This is what the codebook says about the
regime types:

DEMOCRACIES

We use a procedural definition of democracy, meaning we are interested in
the institutional rules that dictate how leaders gain power and not in
other correlates of democracy, including strong traditions of freedom of
speech and assembly. Non-competitive single-party systems are classified as
such, even if some are somewhat more liberal than others (see Botswana and
Namibia). Democracies have reasonably free-and-fair competitions for
political power.
\enumerate{
\item Presidential Democracy: Democracy in which the executive is distinct
from the legislative branch and considerable decision-making authority is
granted to the executive. Presidential systems have presidents who serve as
chief executives rather than figureheads.
\item Parliamentary Democracy: Democracy in which legislatures are more
powerful and executives are less autonomous. Generally speaking, countries
with powerful prime ministers and general elections are parliamentary
democracies. Hybrid semi-presidential systems are classified case-by-case,
but are usually grouped with parliamentary democracies.
}

NON-DEMOCRACIES

The Autocratic Regimes dataset by Professors Barbara Geddes, Joseph Wright,
and Erica Frantz classifies all non-democratic governments into one of ten
categories. Though our data differs from this dataset in important ways, it
uses the same typology. We document major differences in the extended
technical notes.
\enumerate{
\item Personalist Systems: Power is highly concentrated in the hands of a
non-monarch dictator who is relatively unconstrained by a military or
political party. Contemporary examples include Russia, Sudan, and Chad.
\item Monarchies: Power is highly concentrated in the hands of a monarch who
is much more than just a figurehead. Contemporary examples include
Swaziland, Kuwait, and Morocco.
\item Single-Party Systems: Power is held by the head of a party. Executive
power is effectively checked by the party or ruling committee. Contemporary
examples include China, Angola, and Ethiopia.
\item Oligarchies: Power is held by the head of party, but unlike other
single-party systems this party explicitly represents the interests of one
elite segment of society. Past examples include apartheid-era South Africa
and Rhodesia under Ian Smith.
\item Party-Personalist Hybrids: An intermediate hybrid where a party
apparatus supports a dictator, yet the party's identity is concentrated
around the person in power and it has few meaningful checks on executive
power. Examples include Eritrea, and North Korea.
\item Military Juntas: A military committee runs the country. One officer
typically serves as head, but this head serves the interests of the
committee and his power is checked by other members of the military. Recent
examples include Thailand and Algeria.
\item Indirect Military Juntas: The military has de facto power, but rules
behind a civilian puppet. See pre-Mobutu Zaire and Suriname under Bouterse
(1980-1988).
\item Personalist-Military Hybrids: A hybrid of military and personalist
institutions in which a dictator holds most power and is relatively
unchecked, yet the dictator's authority is rooted in military support.
These systems often evolve from juntas when power is consolidated around a
single individual. Examples include Chile under Pinochet, Pakistan under
Zia and Musharraf, and Fiji under Bainimarama.
\item Party-Military Hybrids: Militarized single-party states in which most or
all members of the ruling party are military elites. Examples include
Algeria from 1962-1992, El Salvador before 1982, and Congo-Brazzaville
between the 1968 coup and 1991.
\item Party-Personalist-Military Hybrids: A dictator rules with the support
of a militarized single-party state, but is relatively unchecked by these
institutions. Examples include Egypt after 1952, Indonesia under Suharto,
and Syria under the Assads.
}

INTERIM PERIODS

Finally, we include four forms of government that are explicitly
provisional or transitional:
\enumerate{
\item Warlordism: occurs only in countries that are torn apart by conflict to
the extent that they do not have a functional government. As of 2016, only
Libya and Yemen meet this definition. War-torn countries like Syria are not
included because a strong government continues to hold power in the capital
and a significant part of the country.
\item Foreign-Occupied governments occur where foreign politicians or
militaries hold de facto power over a government.
\item Civilian Provisional and Military Provisional governments are explicitly
temporary arrangementsthat usually proceed completed transitions to
democracy or follow coups and constitutional crises. Interim regimes are
only called "military provisional" if the military is holding power until
an election or some other formalized legitimizing event can occur.}
}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other authoritarianism: 
\code{\link{anckar}},
\code{\link{gwf_all}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{utip}},
\code{\link{wahman_teorell_hadenius}}
}
\concept{authoritarianism}
\concept{democracy}
\keyword{datasets}
