% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{svmdi}
\alias{svmdi}
\alias{svmdi_2016}
\title{Suport Vector Machine Democracy Index by Grundler and Krieger}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12588 rows and 9 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5376 rows and 13 columns.
}
\source{
K. Gr端ndler and T. Krieger. "Democracy and Growth: Evidence from a Machine Learning Indicator". In: \emph{European Journal of Political Economy} 45 (2016), pp. 85-107. DOI: \href{https://dx.doi.org/10.1016/j.ejpoleco.2016.05.005}{10.1016/j.ejpoleco.2016.05.005}.

K. Gr端ndler and T. Krieger. "Democracy and Growth: Evidence from a Machine Learning Indicator". In: \emph{European Journal of Political Economy} 45 (2016), pp. 85-107. DOI: \href{https://dx.doi.org/10.1016/j.ejpoleco.2016.05.005}{10.1016/j.ejpoleco.2016.05.005}.
}
\usage{
svmdi

svmdi_2016
}
\description{
This is the index described in K. Gr端ndler and T. Krieger. \emph{Machine Learning Indices, Political Institutions, and Economic Development}. Report. CESifo Group Munich, 2018. An earlier
version of this index (used in K. Gr端ndler and T. Krieger. "Democracy and Growth: Evidence from a Machine Learning Indicator". In: \emph{European Journal of Political Economy} 45 (2016), pp. 85-107. DOI: \href{https://dx.doi.org/10.1016/j.ejpoleco.2016.05.005}{10.1016/j.ejpoleco.2016.05.005}.) is available as
\link{svmdi_2016}; there are some important differences between the values of the
earlier and later versions of the index.
}
\section{Variables}{


\describe{

\item{svmdi_country}{The country name, as in the original dataset, with
minimal modification. Use \code{extended_country_name} instead if you want
a consistent name.}

\item{svmdi_iso}{Three letter ISO country code. In the original dataset.}

\item{year}{The calendar year.}

\item{region1}{Geographical region. (In the 2016 release only).}

\item{region2}{Geographical region. (In the 2016 release only).}

\item{csvmdi}{Continuous version of the support vector machine democracy
index (In the 2020 release only). Ranges from 0 = least demcratic to 1 =
most democratic. }

\item{dsvmdi}{Discrete version of the support vector machine democracy
index (In the 2020 release only). 0 = Not demcratic, 1 =
democratic. }

\item{svmdi}{Continuous version of the support vector machine democracy
index (In the 2016 release only). Ranges from 0 = least demcratic to 1 =
most democratic. }

\item{svmdi_sd}{Standard deviation of the continuous version of the support
vector machine democracy index (In the 2016 release only). }

\item{svmdi_q5}{Lower confidence bound of the continuous version of the
support vector machine democracy index (In the 2016 release only). }

\item{svmdi_q95}{Upper confidence bound of the continuous version of the
support vector machine democracy index (In the 2016 release only). }

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}}
}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
