% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{uds_2014}
\alias{uds_2014}
\alias{uds}
\alias{uds_2011}
\alias{uds_2010}
\title{The Unified Democracy Scores}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9850 rows and 12 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9137 rows and 12 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7558 rows and 12 columns.
}
\source{
D. Pemstein, S. Meserve, and J. Melton. "Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type". In: \emph{Political Analysis} 18.4 (2010), pp. 426-449. DOI: \href{https://dx.doi.org/10.1093/pan/mpq020}{10.1093/pan/mpq020}.

D. Pemstein, S. A. Meserve, and J. Melton. \emph{Replication Data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type}. 2013. DOI: \href{https://dx.doi.org/10.7910/DVN/WWYOHU}{10.7910/DVN/WWYOHU}.
}
\usage{
uds_2014

uds_2011

uds_2010
}
\description{
The 2014, 2011, and 2010 releases of the Unified Democracy Scores by
Pemstein, Meserve, and Melton (2010).
}
\details{
\describe{

\item{uds_country}{The name of the country in the UDS release}

\item{year}{Calendar year.}

\item{uds_ccode}{The correlates of war numeric code, as in the original UDS
release. This is different for Yugoslavia/Serbia from the actual COW code
after 2006. Use the \code{cown} variable instead for merging with COW-indexed
data.}

\item{mean}{The posterior mean of the latent variable in the UD release.}

\item{sd}{The posterior standard deviation of the latent variable in the UD
release.}

\item{median}{The posterior median of the latent variable in the UD release.}

\item{pct025}{Unified democracy score posterior 2.5 percentile.}

\item{pct975}{Unified democracy score posterior 97.5 percentile.}

}
}
\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\seealso{
The \link{generate_extended_uds} function to extend these scores, as well as vignette("Replicating_and_extending_the_UD_scores").

Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vaporeg}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other latent variable indexes of democracy: 
\code{\link{extended_uds}},
\code{\link{polyarchy_dimensions}}
}
\concept{democracy}
\concept{latent variable indexes of democracy}
\keyword{datasets}
