% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vanhanen}
\alias{vanhanen}
\alias{vanhanen_pmm}
\title{Vanhanen measures of democracy, 1800-2012}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15735 rows and 9 columns.}
\source{
Vanhanen, Tatu. Measures of Democracy 1810-2012 (computer file). FSD1289,
version 6.0 (2014-01-31). Vanhanen, Tatu & Lundell, Krister (data
collection). Tampere. Finnish Social Science Data Archive (distributor),
2014. Data available at
\url{http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html}
}
\usage{
vanhanen

vanhanen_pmm
}
\description{
A dataset containing Tatu Vanhanen's measures of democracy. To cite use
Vanhanen, Tatu, Measures of Democracy 1810-2012 (computer file). FSD1289,
version 6.0 (2014-01-31). Vanhanen, Tatu & Lundell, Krister (data
collection). Tampere. Finnish Social Science Data Archive (distributor),
2014. Data available at
\url{http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html}.
}
\section{Variables}{


\describe{

\item{vanhanen_country}{The country name in the original dataset.}

\item{year}{Year.}

\item{vanhanen_competition}{From Vanhanen's introduction to the dataset:
The smaller parties' share of the votes cast in parliamentary or
presidential elections, or both, is used to indicate the degree of
competition. It is calculated by subtracting the percentage of votes won by
the largest party from 100. If the largest party gets, for example, 40
percent of the votes, the share of the smaller parties is 60 percent. If
data on the distribution of votes are not available, the value of this
variable is calculated on the basis of the distribution of seats in
parliament. The distribution of seats is used also in cases in which it
seems to indicate power relations more realistically than the distribution
of votes.}

\item{vanhanen_participation}{From Vanhanen's introduction to the dataset:
The percentage of the population which actually voted in the same
elections is used to measure the degree of participation (= Participation).
This percentage is calculated from the total population, not from the adult
or enfranchized population. I use the total population as the basis of
calculation because more statistical data are available on total
populations than on age structures of electorates. In principle, these two
empirical variables are very simple and easy to use. In practice, however,
there are several points where more detailed rules of interpretation are
needed.

Definition of a party

First, it is necessary to define what is meant by 'a party' and 'the
largest party' in these calculations. My basic assumption is that the
relative strength of the largest political party provides the most
realistic indicator of the distribution of political power in modern
states. Competing groups have formed more or less permanent political
parties since the nineteenth century, but it is not always obvious which
groups should be regarded as 'parties.' Historically, factions, political
cliques and groups of notables preceded parties. Parties as we understand
them have emerged since the first half of the nineteenth century (see
Duverger 1954; LaPalombara and Weiner 1966; Sartori 1976; von Beyme 1984;
Blondel 1995). Many definitions of political parties emphasize that a party
is an organized group and that its principal aim is to win political power
(Michels 1962; LaPalombara and Weiner 1966). According to Giovanni Sartori
(1976: 63-64): 'A party is any political group identified by an official
label that presents at elections, and is capable of placing through
elections (free or non-free), candidates for public office.' I think that
this definition provides sufficient criteria to distinguish 'parties' and
'the largest party' from other political groups. It is plausible to regard
as 'parties' all political groups which take part in elections and are
identified by an official label. Usually, but not always, it is easy to
distinguish between parties taking part in elections. Party alliances are
problematic. It is not always clear whether the alliance or its individual
member parties should be regarded as 'parties.' In such cases, a party's
behaviour in elections is used as the decisive criterion. If a party
belongs to a larger alliance permanently, we are not justified in regarding
it as a separate party. The alliance should then be treated as a separate
'party,' because the purpose is to measure the relative strength of
competing and independent groups.

In parliamentary elections 'the largest party' refers to the party which
received the largest single share of the votes or of the seats in
parliament. Sometimes, depending on the type of electoral system, the
proportion of seats may be considerably higher than the proportion of
votes, whereas the reverse situation is hardly possible.

In presidential elections, the term 'largest party' refers to the votes
received by the presidential candidate who won the election. A problem is,
however, whether we should take into account the first or the second round
votes, if there are two rounds of voting. The percentage of votes obtained
by the winning candidate may be significantly higher in the second
round than in the first. The round of voting, if it is not the first round,
is indicated in country tables. The purpose has been to take into account
the round that reflects the strength of parties and political groups most
reliably.

Indirect elections and elections without parties

Interpretation is needed in indirect elections, too. How should we
calculate the degree of participation in such elections? My basic rule has
been that only votes cast in final election are counted. When president is
elected by indirect elections, usually by parliament, only the number of
actual electors is taken into account, which means that the degree of
participation drops to zero. The same interpretation is applied to indirect
parliamentary elections (in China, for example). However, if the real
election takes place at the election of electors, as in the presidential
elections of the United States, the number of votes and the distribution of
votes in that election are taken into account.

Another problem of interpretation concerns countries where members of
parliament are elected but political parties are not allowed to take part
in elections, or to form party groups in parliament after elections. Such
election results are usually interpreted to mean that one party has taken
all the votes or the seats. This interpretation is based on the assumption
that the ruling group does not allow political competition for power in
elections. Parties are absent from elections because they are banned. In
such cases the "largest party's" share is assumed to be 100 percent.

The situation is different in countries in which only independent
candidates participate in elections, although parties are not banned and
although it would be legally possible to establish parties (Micronesia, for
example). In such cases it is plausible to assume that elections are
competitive and that elected members of parliament are not controlled by
any particular political group or by the government. Independent members of
parliament may form at least temporary political groups in the parliament
freely. Therefore, it is assumed in such cases that the "largest party's"
share is not higher than 30 percent.

Non-elected governmental institutions

A different question of interpretation arises in cases where the
composition of a governmental institution using the highest executive or
legislative power is not based on popular election. How should the degree
of competition and the degree of participation be measured in such cases?
According to my interpretation, the share of the smaller parties and the
degree of electoral participation are zero in such cases. Power is
concentrated in the hands of the ruling group. This interpretation applies
to military and revolutionary regimes, to other non-elected autocratic
governments, provisional governments, and to monarchies in which the ruler
and the government responsible to the ruler dominate and exercise executive
and usually also legislative power. There are many such historical as well
as contemporary cases. In all these cases the "largest party's" share is
assumed to be 100 percent and the degree of participation zero.

Dominant governmental institutions

The calculation of the values of competition and participation can be based
on parliamentary or presidential/executive elections, or both. In each case
it is necessary to decide which governmental institution and election
should be taken into account. This depends on the assumed importance of the
two governmental institutions. The relative importance of parliaments and
presidents (or other heads of state) varies greatly, but usually these two
governmental institutions are, at least formally, the most important
institutions wielding political power. Depending on how power is divided
between them, we can speak of parliamentary and presidential forms of
government. In the former, the legislature is dominant. The executive
branch is dependent on and responsible to the legislative branch. In the
latter, the executive branch is dominant and is not responsible to the
legislature. But it is also possible for their powers to be so well
balanced that neither has clear dominance. Thus we can distinguish three
institutional power arrangements at the national level: (a) parliamentary
dominance, (b) executive dominance, and (c) concurrent powers. In the first
case the values of competition and participation are calculated on the
basis of parliamentary elections, in the second they are calculated on the
basis of presidential or other executive elections (or the lack of
elections), and in the third case both possible elections are taken into
account.

If the support of competing parties is about the same in both elections (as
in the United States, for example), it does not make much difference how
the governmental system is classified in order to calculate the values of
competition and participation, but if the electoral systems are
significantly different in parliamentary and presidential elections, an
incorrect classification of the country's governmental system might distort
the results of the measurement. The same is true if the powers of the two
institutions differ crucially. I have attempted to classify each country's
governmental institutions as realistically as possible. All classifications
of governmental systems are indicated in country tables.

Interpretation in the classification of governmental systems has been
needed especially in the cases in which the results of presidential
elections are based on the second round of voting. Because in such cases
the share of the elected president tends to be 50.0 percent or higher, it
is useful to check the relative strength of parties by taking parliamentary
elections into account, too.

When both elections are taken into account (concurrent powers), it is
necessary to weight the relative importance of parliamentary and
presidential elections. Usually it is reasonable to give equal weight (50
percent) to both elections, but in some cases it may be more realistic to
give a weight of 75 or 25 percent to parliamentary elections and 25 or 75
percent to presidential (executive) elections. In most cases it is
relatively easy to decide which of the two branches of government is
dominant and which elections should be taken into account, but some cases
are open to different interpretations. The same applies to the weighting of
the two branches in the cases of concurrent powers. The classifications of
the governmental systems and changes of the governmental systems are
indicated in each country table. In the cases of "concurrent powers," the
estimated relative importance of the two branches of government is also
indicated (50-50 percent, 25-75 percent, or 75-25 percent).}

\item{vanhanen_democratization}{From Vanhanen's introduction to the
dataset: The two basic indicators of democratization can be used
separately to measure the level of democracy, but, because they are assumed
to indicate two different dimensions of democratization, it is reasonable
to argue that a combination of them would be a more realistic indicator of
democracy than either of them alone. They can be combined in many ways,
depending on how we weight the importance of Competition and Participation.
Some researchers (see, for example, Bollen 1979, 1980; Coppedge and
Reinicke 1988) have excluded the degree of electoral participation from
their measures of democracy because they think that it does not represent a
significant differentiating aspect of democracy. My argument is that
participation is probably as important dimension of democracy as
competition. If only a small minority of the adult population takes part in
elections, the electoral struggle for power is restricted to the upper
stratum of the population, and the bulk of the population remains outside
national politics. Power sharing is then certainly more superficial than in
societies where the majority of the adult population takes part in
elections (of course, presupposing that elections are competitive). Because
I am not sure which of these two dimensions of democratization is more
important and how much more important, I have weighted them equally in the
Index of Democratization (ID). This is an arbitrary choice, but it is based
on the assumption that both dimensions are equally important and necessary
for democratization.

However, the decision to weight them equally does not solve the problem of
how to combine them. One way would be to calculate their arithmetic mean.
Another way is to multiply them. We could also use a mixture of adding and
multiplying, for example, by first multiplying them and then adding 25
percent (or some other percentage) of the values of both indicators to the
index. The first combination would be based on the assumption that both
dimensions indicate the degree of democracy independently and that a high
level of competition can partly compensate for the lack of participation,
or vice versa. The second combination is based on the assumption that both
dimensions are necessary for democracy and that a high level of competition
cannot compensate the lack of participation, or vice versa. I have come to
the conclusion that the latter assumption is theoretically better than the
former one because it is plausible to assume that both dimensions are
important for democracy. So the two indicators - Competition and
Participation - are combined into an Index of Democratization (ID) by
multiplying them and dividing the outcome by 100.

The decision to weight indicators equally and to multiply them means that a
low value for either of the two variables is enough to keep the index value
low. A high level of participation cannot compensate for the lack of
competition, or vice versa. The Index of Democratization gets high values
only if the values of both basic variables are high. Multiplication of the
two percentages corrects one fault in Participation variable mentioned
above, namely, that this indicator thus not differentiate between important
and formal elections. There have been and still are countries where the
level of electoral participation is high but the level of democracy low,
because elections are not free and competitive. Multiplication of the two
percentages cancels the misleading information provided by Participation in
such cases and produces a low ID value. The same correction takes place in
opposite cases, when the level of competition is high but the degree of
electoral participation low.

Referendums do not affect the value of Competition variable. They affect
only the Participation variable and through it the Index of
Democratization, but the effect of referendums (Participation) on the Index
of Democratization depends crucially on the degree of Competition. If
Competition is in zero, the Index of Democratization cannot rise from zero,
although the value of Participation variable were high. In other words, the
higher the value of Competition, the more the same number of referendums
increases the value of ID.

This index of democracy is simpler than any of the alternative measures of
democracy (cf. Munck and Verkuilen 2002; Munck 2009). It differs from other
measures in two important ways: (1) it uses only two indicators, and (2)
both of them are based, in principle, on quantitative data. Most other
measures of democracy include more indicators, and most are based on more
or less qualitative data. I think that it is scientifically more justified
to use simple quantitative indicators than more complicated indicators
loaded with weights and estimates based on subjective judgements if those
simple quantitative indicators are as valid measures of the phenomeon as
the more complicated and less quantitative indicators. However, some
subjective judgments are needed also in the use of my measures of
democracy, but it is possible for other researchers to see from the dataset
what those subjective interpretations have been. One advantage of this
Index of Democratization is that empirical data on the two basic variables
are available from different sources, that statistical data on elections
are in most cases exact and reliable, and that the role of subjective
judgements in the use of electoral data is relatively limited. Further, I
would like to emphasize that because the two basic variables do not take
into account all important aspects of democracy, they are better adapted to
indicate significant differences between political systems from the
perspective of democracy than more detailed differences among democracies
or nondemocracies (cf. Vanhanen 2000a, 2000b; 2003).}
}
}

\section{Standard descriptive variables (generated by this package)}{

 \describe{

 \item{extended_country_name}{The name of the country in the Gleditsch-Ward
 system of states, or the official name of the
 entity (for non-sovereign entities and states not in the Gleditsch and Ward
 system of states) or else a common name for disputed cases that do not have
 an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
 scheme sometimes indicates the common name of the country and (in
 parentheses) the name of an earlier incarnation of the state: thus, they
 have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
 For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
 System Membership: A Revised List of the Independent States since 1816."
 International Interactions 25: 393-413. The list can be found at
 \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

 \item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
 Ward list of independent states.}

 \item{cown}{The Correlates of War numeric country code, 2016 version. This
 differs from Gleditsch and Ward's numeric country code in a few cases. See
 \url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
 the full list.}

 \item{in_GW_system}{Whether the state is "in system" (that is, is
 independent and sovereign), according to Gleditsch and Ward, for this
 particular date. Matches at the end of the year; so, for example South
 Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
 country ended on April 1975 (being absorbed by North Vietnam). It is also
 \code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
 the fact that the Gleditsch and Ward list has not been updated since.} }
}

\section{Note}{


The \code{vanhanen_pmm} version of this data is taken from from Pemstein,
Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for:
Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime
Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}. The
original dataset is found in \code{vanhanen}. The \code{vanhanen_pmm} data only goes
back to 1945. There are some missing values in PMM's data compared to the
original Vanhanen dataset. For more detail on the differences, see the
vignette. The vanhanen_pmm data is included here for completeness.
}

\seealso{
Other democracy: \code{\link{LIED}}, \code{\link{PIPE}},
  \code{\link{arat_pmm}}, \code{\link{blm}},
  \code{\link{bmr}}, \code{\link{bnr}},
  \code{\link{bollen_pmm}}, \code{\link{doorenspleet}},
  \code{\link{download_fh_electoral}},
  \code{\link{download_fh}}, \code{\link{download_reign}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{fh_pmm}}, \code{\link{gwf_all}},
  \code{\link{hadenius_pmm}}, \code{\link{kailitz}},
  \code{\link{magaloni}}, \code{\link{mainwaring}},
  \code{\link{munck_pmm}}, \code{\link{pacl}},
  \code{\link{peps}}, \code{\link{pitf}},
  \code{\link{polity_pmm}},
  \code{\link{polyarchy_dimensions}},
  \code{\link{polyarchy}}, \code{\link{prc_gasiorowski}},
  \code{\link{svolik_regime}}, \code{\link{uds_2014}},
  \code{\link{ulfelder}}, \code{\link{utip}},
  \code{\link{wahman_teorell_hadenius}}

Other continuous democracy indexes: \code{\link{arat_pmm}},
  \code{\link{bollen_pmm}},
  \code{\link{download_wgi_voice_and_accountability}},
  \code{\link{hadenius_pmm}}, \code{\link{munck_pmm}}
}
\keyword{datasets}
