% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vaporeg}
\alias{vaporeg}
\title{The Varieties of Political Regimes (VaPoReg) dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 29840 rows and 30 columns.
}
\source{
S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Dataset}. Dresden, 2024.

S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Codebook. Version 1.7}. Tech. rep. Dresden: Hannah Arendt Institute for Totalitarianism Studies, 2024.
}
\usage{
vaporeg
}
\description{
S. Kailitz. \emph{Varieties of Political Regimes (va-PoReg). Codebook. Version 1.7}. Tech. rep. Dresden: Hannah Arendt Institute for Totalitarianism Studies, 2024. The original data
and codebook can be downloaded from <`r find_url("vaporeg",
"documentation")>. The documentation below is derived from the codebook, with
modifications.
}
\section{Credits}{
 The data were originally compiled by Steffen Kailitz and
the Varieties of Political Regimes project. The dataset has been revised
and expanded based on historical sources, country reports, and
international political datasets.
}

\section{Introduction}{


The VaPoReg dataset provides a classification of political regimes globally
from 1900 to 2024. It covers independent states, colonial regimes,
occupation regimes, and semi-sovereign entities such as protectorates. The
dataset codes each country-year according to different regime typologies,
integrating various sources of political regime data.

The dataset includes unique features such as:
\itemize{
\item Distinguishing between \strong{electoral} and \strong{non-electoral} autocracies.
\item Differentiating between \strong{constitutional monarchies, absolute monarchies, and personalist regimes}.
\item Identifying \strong{colonial and occupation regimes} separately from independent states.
\item Incorporating non-sovereign entities that have historically been excluded from other datasets.
}
}

\section{Conceptual Framework}{


The dataset classifies regimes based on their \strong{institutional structure,
political legitimation, and origin of rule}. The classification emphasizes
observable characteristics, such as whether a regime emerged from a coup,
follows ideological principles, or relies on military governance.
}

\section{Variables}{


\describe{

\item{country_name}{The name of the country for a given year.}

\item{year}{The calendar year.}

\item{vaporeg_code}{The country code based on the Correlates of War (COW) dataset.}

\item{GeoNames_ID}{Geographic identifier for linking with spatial datasets.}

\item{vaporeg_s}{The \strong{standard classification} of political regimes. Coded as:
\itemize{
\item 10: Democracy
\item 20: Semidemocracy
\item 30: Electoral Oligarchy
\item 40: Non-electoral Transitional Regime
\item 50: Electoral Autocracy
\item 60: Constitutional Monarchy
\item 70: Absolute Monarchy
\item 80: One-party Autocracy
\item 90: Right-wing Autocracy
\item 100: Communist Ideocracy
\item 110: Islamist Ideocracy
\item 120: Military Autocracy
\item 130: Personalist Autocracy
\item 140: Colonial Regime
\item 150: Occupation Regime
\item 160: No Central Authority
}
}

\item{vaporeg_a}{An \strong{alternative classification}, which merges democracy and semidemocracy and groups
monarchies together as ruling monarchies.}

\item{vaporeg_cr}{A classification similar to \strong{VaPoReg_s}, but includes an additional category for \strong{"Part of Other Country"}.}

\item{vaporeg_governing_country}{For colonial and occupation regimes, this variable identifies the ruling country.}

\item{vaporeg_s_of_gc}{The \strong{standard regime classification} for the governing country.}

\item{vaporeg_a_of_gc}{The \strong{alternative regime classification} for the governing country.}

\item{vaporeg_s_start}{The date on which the political regime type,
as defined by the standard version of our political regime categorization,
began.}

\item{vaporeg_s_end}{The date on which the political regime type,
as defined by the standard version of our political regime categorization,
ended.}

\item{vaporeg_s_duration}{The number of years the political regime has lasted,
as defined by the standard version of our political regime categorization,
counting only if the regime was still in power on July 1st of each year.}

\item{vaporeg_s_change}{The variable identifies regime changes based on the
standard version of our political regimecategorization. It takes a value of
one if a different regime is in place on July 1st compared to July 1st of
the previous year.}

\item{vaporeg_a_start}{The date on which the political regime type,
according to the alternative version of our political regime categorization,
began.}

\item{vaporeg_a_end}{The date on which the political regime type,
according to the alternative version of our political regime categorization,
ended.}

\item{vaporeg_a_duration}{The number of years the political regime has lasted,
as defined by the alternative version of our political regime categorization,
counting only if the regime was still in power on July 1st ofeach year.}

\item{vaporeg_a_change}{The variable identifies regime changes based on the
standard version of our political regime categorization. It takes a value of
one if a different regime is in place on July 1st compared to July 1st of the
previous year.}

\item{vaporeg_binary_strict}{A \strong{dichotomous classification} of democracy, based
on \code{VaPoReg_s}. Coded as 1 if \code{VaPoReg_s} is a Democracy (10), 0 otherwise.}

\item{vaporeg_binary_non-strict}{A \strong{dichotomous classification} of democracy, based
on \code{VaPoReg_s}, but less strict. Coded as 1 if \code{VaPoReg_s} is a Democracy (10) or a Semidemocracy (20),
0 otherwise.}

\item{vaporeg_trichotomous}{A \strong{trichotomous classification} of democracy, based
on \code{VaPoReg_s}. Coded as 2 if \code{VaPoReg_s} is a Democracy (10), 1 if it is a Semidemocracy (20),
0 otherwise.}

\item{Gapminder_pop}{Population, per the Gapminder dataset.}

\item{gm_world_pop}{World population, per the Gapminder dataset}

\item{un_region}{UN region}

\item{un_cntinent}{UN continent}

}
}

\section{Special Cases}{

\itemize{
\item \strong{Colonial Regimes}: Defined by foreign rule, legal discrimination, and lack of sovereignty.
\item \strong{Occupation Regimes}: Military-controlled administrations imposed by foreign powers.
\item \strong{No Central Authority}: Cases where no governing entity holds effective control (e.g., Somalia post-1991).
\item \strong{Protectorates and Mandates}: Semi-sovereign territories coded separately in some instances.
}
}

\section{Sources}{


The dataset integrates multiple sources, including:
\itemize{
\item \strong{Freedom House}
\item \strong{Polity IV}
\item \strong{V-Dem (Varieties of Democracy Project)}
\item \strong{LIED (Lexical Index of Electoral Democracy)}
\item \strong{Historical country reports, election archives, and secondary literature}
}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\references{
Kailitz, Steffen. "Varieties of Political Regimes." Available at
\url{https://www.va-poreg.de/home}.
}
\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh}()},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl}},
\code{\link{pacl_update}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy}},
\code{\link{polyarchy_dimensions}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{vdem_simple}},
\code{\link{wahman_teorell_hadenius}}

Other institutions: 
\code{\link{LIED}},
\code{\link{PIPE}}
}
\concept{democracy}
\concept{institutions}
\keyword{datasets}
