% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{UTF-8}
\name{vdem_simple}
\alias{vdem_simple}
\alias{vdem}
\title{Simplified version of the V-Dem version 12 dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27555 rows and 54 columns.
}
\usage{
vdem_simple
}
\description{
This data frame contains a very reduced version of version 12 of the V-Dem
dataset, including only the 7 main indexes that measure democracy directly
(v2x_polyarchy, v2x_libdem, v2x_partipdem, v2x_delibdem, v2x_egaldem,
v2x_api, and v2x_mpi) and their associated uncertainty measures. The full
VDem dataset is available via the
\href{https://github.com/vdeminstitute/vdemdata}{vdemdata} package. To cite VDem,
use Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I.
Lindberg, Jan Teorell, Nazifa Alizada, David Altman, Michael Bernhard, Agnes
Cornell, M. Steven Fish, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Sandra
Grahn, Allen Hicken, Garry Hindle, Nina Ilchenko, Katrin Kinzelbach, Joshua
Krusell, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Juraj
Medzihorsky, Pamela Paxton, Daniel Pemstein, Josefine Pernes, Oskar Rydén,
Johannes von Römer, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning,
Jeffrey Staton, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig,
Steven Wilson and Daniel Ziblatt. 2022. "V- Dem [Country–Year/Country–Date]
Dataset v12" Varieties of Democracy (V-Dem) Project.
https://doi.org/10.23696/vdemds22.
}
\details{
Variable descriptions taken from the V-Dem codebook directly.
}
\section{V-Dem Identifiers}{


\describe{

\item{vdem_country_name}{Country name. Name of coded country. A V-Dem
country is a political unit enjoying at least some degree of functional
and/or formal sovereignty.}

\item{country_text_id}{Country name abbreviation. Abbreviated country
names.}

\item{country_id}{V-Dem country ID. Unique country ID designated for each
country.}

\item{year}{Year. Four-digit designation of the year for which an
observation is given that ranges from the start to the end of the coding
period.}

\item{historical_date}{Historical Date. Variable designating the date for
which observation is given. The date is coded in YYYY-MM-DD format.
December 31 observation always refers to the situation at the end of the
year. There can be observations on other dates signifying other events,
i.e. elections or executive appointments.}

\item{project}{V-Dem Project. Indication what project team has coded
country in respective year.}

\item{historical}{Historical V-Dem Coding. Binary indication whether the
country in question has been coded by the team of Historical V-Dem
project.}

\item{histname}{Time-specific name of coded country. Many countries go by
different names in different time-periods, for example due to name changes,
changes in territory, colonization, occupation, or independence. This
variable contains a brief description of the identity of each polity that
comprises a country's history. This variable is based on the V-Dem Country
Coding Units document.}

\item{codingstart}{Start of coding period. Year in which coding of the
country in question starts. V-Dem country coding starts in 1789, or from
when a country first enjoyed at least some degree of functional and/or
formal sovereignty.}

\item{codingend}{End of coding period. Either a maximum year of country
coding period or the year when the country ceased to exist because it lost
functional or formal sovereignty.}

\item{codingstart_contemporary}{Contemporary start of coding period. Year
in which the coding of Contemporary V-Dem project starts. Variables from
"Contemporary" project can have different question formulation, variable
type, or number of coders as opposed to the "Historical" one.}

\item{codingend_contemp}{Historical End of Coding Period. Year in which the
coding of Contemporary V-Dem project ends. Variables from the
"Contemporary" project can have different question formulation, variable
type, or number of coders as opposed to the "Historical" one.}

\item{codingstart_hist}{Historical start of coding period. Year in which
the coding of Historical V-Dem project starts. Variables from
``Historical'' project can have different question formulation, variable
type, or number of coders as opposed to the "Contemporary" one.}

\item{codingend_hist}{Historical End of Coding Period. Year in which the
coding of Historical V-Dem project ends. Variables from "Historical"
project can have different question formulation, variable type, or number
of coders as opposed to the "Contemporary" one.}

\item{gapstart1, gapstart2, gapstart3}{Time periods when a country does not
fulfill V-Dem's coding period criteria are not coded. The date that
indicates the gap start is the last date coded before the gap. For more
details about V-Dem country coding periods, please see the V-Dem Country
Coding Units document.}

\item{gapend1, gapend2, gapend3}{The periods of when a country does not
fulfill V-Dem's coding period criteria are not coded. The date that
indicates the gap end is the first date coded after the gap. For more
details about V-Demcountry coding periods, please see the V-Dem Country
Coding Units document.}

\item{gap_index}{An index for each country and continuous non-gap, i.e. it
is reasonable for interpolation to interpolate only within the same gap
index. For more details about V-Dem country coding periods, please see the
V-Dem Country Coding Units document.}

\item{vdem_cowcode}{COW Country Code. Correlates of War (COW) project country
codes.}

}
}

\section{V-Dem Main Indexes}{


Each of the indexes described below comes with associated uncertainty
measures: \verb{_codehigh} and \verb{_codelow} (the interval where the V-Dem Bayesian
measurement model places 68\% of the probability mass, or approximately one
standard deviation upper and lower bounds, possibly asymmetric) and \verb{_sd}
(frequentist standard deviation).

\describe{

\item{v2x_polyarchy}{Electoral democracy index. To what extent is the ideal
of electoral democracy in its fullest sense achieved? The electoral
principle of democracy seeks to embody the core value of making rulers
responsive to citizens, achieved through electoral competition for the
electorate's approval under circumstances when suffrage is extensive;
political and civil society organizations can operate freely; elections are
clean and not marred by fraud or systematic irregularities; and elections
affect the composition of the chief executive of the country. In between
elections, there is freedom of expression and an independent media capable
of presenting alternative views on matters of political relevance. In the
V-Dem conceptual scheme, electoral democracy is understood as an essential
element of any other conception of representative democracy --- liberal,
participatory, deliberative, egalitarian, or some other.}

\item{v2x_libdem}{Liberal democracy index. To what extent is the ideal of
liberal democracy achieved?

The liberal principle of democracy emphasizes the importance of protecting
individual and minority rights against the tyranny of the state and the
tyranny of the majority. The liberal model takes a ~negative~ view of
political power insofar as it judges the quality of democracy by the limits
placed on government. This is achieved by constitutionally protected civil
liberties, strong rule of law, an independent judiciary, and effective
checks and balances that, together, limit the exercise of executive power.
To make this a measure of liberal democracy, the index also takes the level
of electoral democracy into account.}

\item{v2x_delibdem}{Deliberative democracy index. To what extent is the
ideal of deliberative democracy achieved?

The deliberative principle of democracy focuses on the process by which
decisions are reached in a polity. A deliberative process is one in which
public reasoning focused on the common good motivates political
decisions—as contrasted with emotional appeals, solidary attachments,
parochial interests, or coercion. According to this principle, democracy
requires more than an aggregation of existing preferences. There should
also be respectful dialogue at all levels—from preference formation to
final decision—among informed and competent participants who are open to
persuasion. To make it a measure of not only the deliberative principle but
also of democracy, the index also takes the level of electoral democracy
into account.}

\item{v2x_partipdem}{Participatory democracy index. To what extent is the
ideal of participatory democracy achieved?

The participatory principle of democracy emphasizes active participation by
citizens in all political processes, electoral and non-electoral. It is
motivated by uneasiness about a bedrock practice of electoral democracy:
delegating authority to representatives. Thus, direct rule by citizens is
preferred, wherever practicable. This model of democracy thus takes
suffrage for granted, emphasizing engagement in civil society
organizations, direct democracy, and subnational elected bodies. To make it
a measure of participatory democracy, the index also takes the level of
electoral democracy into account.}

\item{v2x_mpi}{Multiplicative polyarchy index. To what extent is the
electoral principle of democracy achieved?

The electoral principle of democracy seeks to achieve responsiveness and
accountability between leaders and citizens through the mechanism of
competitive elections. This is presumed to be achieved when suffrage is
extensive; political and civil society organizations can operate freely;
elections are clean and not marred by fraud or systematic irregularities;
and the chief executive of a country is selected directly or indirectly
through elections.}

\item{v2x_egaldem}{Egalitarian democracy index. To what extent is the ideal
of egalitarian democracy achieved?

The egalitarian principle of democracy holds that material and immaterial
inequalities inhibit the exercise of formal rights and liberties, and
diminish the ability of citizens from all social groups to participate.
Egalitarian democracy is achieved when 1) rights and freedoms of
individuals are protected equally across all social groups; and 2)
resources are distributed equally across all social groups; 3) groups and
individuals enjoy equal access to power. To make it a measure of
egalitarian democracy, the index also takes the level of electoral
democracy into account.}

\item{v2x_api}{Additive polyarchy index. To what extent is the electoral
principle of democracy achieved?

electoral principle of democracy seeks to achieve responsiveness and
accountability between leaders and citizens through the mechanism of
competitive elections. This is presumed to be achieved when suffrage is
extensive; political and civil society organizations can operate freely;
elections are clean and not marred by fraud or systematic irregularities;
and the chief executive of a country is selected directly or indirectly
through elections.}

}
}

\section{Standard descriptive variables (generated by this package)}{

\describe{

\item{extended_country_name}{The name of the country in the Gleditsch-Ward
system of states, or the official name of the
entity (for non-sovereign entities and states not in the Gleditsch and Ward
system of states) or else a common name for disputed cases that do not have
an official name (e.g., Western Sahara, Hyderabad). The Gleditsch and Ward
scheme sometimes indicates the common name of the country and (in
parentheses) the name of an earlier incarnation of the state: thus, they
have Germany (Prussia), Russia (Soviet Union), Madagascar (Malagasy), etc.
For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate
System Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch and
Ward list of independent states.}

\item{cown}{The Correlates of War numeric country code, 2016 version. This
differs from Gleditsch and Ward's numeric country code in a few cases. See
\url{http://www.correlatesofwar.org/data-sets/state-system-membership} for
the full list.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Matches at the end of the year; so, for example South
Vietnam 1975 is \code{FALSE} because, according to Gleditsch and Ward, the
country ended on April 1975 (being absorbed by North Vietnam). It is also
\code{TRUE} for dates beyond 2012 for countries that did not end by then, depsite
the fact that the Gleditsch and Ward list has not been updated since.} }
}

\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I.
Lindberg, Jan Teorell, Nazifa Alizada, David Altman, Michael Bernhard,
Agnes Cornell, M. Steven Fish, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn,
Sandra Grahn, Allen Hicken, Garry Hindle, Nina Ilchenko, Katrin Kinzelbach,
Joshua Krusell, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Juraj
Medzihorsky, Pamela Paxton, Daniel Pemstein, Josefine Pernes, Oskar Rydén,
Johannes von Römer, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning,
Jeffrey Staton, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig,
Steven Wilson and Daniel Ziblatt. 2022. "V- Dem
[Country–Year/Country–Date] Dataset v12" Varieties of Democracy (V-Dem)
Project. https://doi.org/10.23696/vdemds22.
}
\seealso{
Other democracy: 
\code{\link{LIED}},
\code{\link{PIPE}},
\code{\link{REIGN}},
\code{\link{anckar}},
\code{\link{anrr}},
\code{\link{arat_pmm}},
\code{\link{blm}},
\code{\link{bmr}},
\code{\link{bnr}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{doorenspleet}},
\code{\link{download_fh_electoral}()},
\code{\link{download_fh_full}()},
\code{\link{download_fh}()},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{extended_uds}},
\code{\link{fh_pmm}},
\code{\link{gwf_all}},
\code{\link{hadenius_pmm}},
\code{\link{kailitz}},
\code{\link{magaloni}},
\code{\link{mainwaring}},
\code{\link{munck_pmm}},
\code{\link{pacl_update}},
\code{\link{pacl}},
\code{\link{peps}},
\code{\link{pitf}},
\code{\link{polityIV}},
\code{\link{polity_pmm}},
\code{\link{polyarchy_dimensions}},
\code{\link{polyarchy}},
\code{\link{prc_gasiorowski}},
\code{\link{svmdi}},
\code{\link{svolik_regime}},
\code{\link{uds_2014}},
\code{\link{ulfelder}},
\code{\link{utip}},
\code{\link{vanhanen}},
\code{\link{wahman_teorell_hadenius}}

Other continuous democracy indexes: 
\code{\link{arat_pmm}},
\code{\link{bollen_pmm}},
\code{\link{bti}},
\code{\link{download_wgi_voice_and_accountability}()},
\code{\link{eiu}},
\code{\link{hadenius_pmm}},
\code{\link{munck_pmm}},
\code{\link{svmdi}},
\code{\link{vanhanen}}
}
\concept{continuous democracy indexes}
\concept{democracy}
\keyword{datasets}
